% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Abundance}
\alias{Abundance}
\alias{abundance.2bp.exome.GRCh37}
\alias{abundance.2bp.genome.GRCh37}
\alias{abundance.3bp.exome.GRCh37}
\alias{abundance.3bp.genome.GRCh37}
\alias{abundance.4bp.exome.GRCh37}
\alias{abundance.4bp.genome.GRCh37}
\alias{abundance.5bp.exome.GRCh37}
\alias{abundance.5bp.genome.GRCh37}
\alias{abundance.2bp.exome.GRCh38}
\alias{abundance.2bp.genome.GRCh38}
\alias{abundance.3bp.exome.GRCh38}
\alias{abundance.3bp.genome.GRCh38}
\alias{abundance.4bp.exome.GRCh38}
\alias{abundance.4bp.genome.GRCh38}
\alias{abundance.5bp.exome.GRCh38}
\alias{abundance.5bp.genome.GRCh38}
\alias{abundance.2bp.exome.GRCm38}
\alias{abundance.2bp.genome.GRCm38}
\alias{abundance.3bp.exome.GRCm38}
\alias{abundance.3bp.genome.GRCm38}
\alias{abundance.4bp.exome.GRCm38}
\alias{abundance.4bp.genome.GRCm38}
\alias{abundance.5bp.exome.GRCm38}
\alias{abundance.5bp.genome.GRCm38}
\title{Nucleotide abundance}
\format{A single-column
matrix containing the counts of particular sequences in
a genome or part of a genome. This include
2-mers, 3-mers, 4-mers, 5-mers, stranded or strand-agnostic,
and genome-wide, in-transcript, or in-exome, for different
reference genome versions and for different organisms. The names should
be self explanatory.}
\usage{
abundance.2bp.exome.GRCh37

abundance.2bp.genome.GRCh37

abundance.3bp.exome.GRCh37

abundance.3bp.genome.GRCh37

abundance.4bp.exome.GRCh37

abundance.4bp.genome.GRCh37

abundance.5bp.exome.GRCh37

abundance.5bp.genome.GRCh37

abundance.2bp.exome.GRCh38

abundance.2bp.genome.GRCh38

abundance.3bp.exome.GRCh38

abundance.3bp.genome.GRCh38

abundance.4bp.exome.GRCh38

abundance.4bp.genome.GRCh38

abundance.5bp.exome.GRCh38

abundance.5bp.genome.GRCh38

abundance.2bp.exome.GRCm38

abundance.2bp.genome.GRCm38

abundance.3bp.exome.GRCm38

abundance.3bp.genome.GRCm38

abundance.4bp.exome.GRCm38

abundance.4bp.genome.GRCm38

abundance.5bp.exome.GRCm38

abundance.5bp.genome.GRCm38
}
\description{
Nucleotide abundance information for a particular organism
}
\details{
\code{abundance.2bp.genome.GRCh37}, \code{abundance.2bp.exome.GRCh37} A
matrix containing dinucleotide abundance information for \strong{Human}
GRCh37. Its row names indicate 10 different types of 2 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatDNS78ToPdf}}.

\code{abundance.2bp.genome.GRCh38}, \code{abundance.2bp.exome.GRCh38} A
matrix containing dinucleotide abundance information for \strong{Human}
GRCh38. Its row names indicate 10 different types of 2 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatDNS78ToPdf}}.

\code{abundance.2bp.genome.GRCm38}, \code{abundance.2bp.exome.GRCm38} A
matrix containing dinucleotide abundance information for \strong{Mouse}
GRCm38. Its row names indicate 10 different types of 2 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatDNS78ToPdf}}.

\code{abundance.3bp.genome.GRCh37}, \code{abundance.3bp.exome.GRCh37} A
matrix containing trinucleotide abundance information for \strong{Human}
GRCh37. Its row names indicate 32 different types of 3 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatSNS96ToPdf}}.

\code{abundance.3bp.genome.GRCh38}, \code{abundance.3bp.exome.GRCh38} A
matrix containing trinucleotide abundance information for \strong{Human}
GRCh38. Its row names indicate 32 different types of 3 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatSNS96ToPdf}}.

\code{abundance.3bp.genome.GRCm37}, \code{abundance.3bp.exome.GRCm37} A
matrix containing trinucleotide abundance information for \strong{Mouse}
GRCm37. Its row names indicate 32 different types of 3 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatSNS96ToPdf}}.

\code{abundance.4bp.genome.GRCh37}, \code{abundance.4bp.exome.GRCh37} A
matrix containing tetranucleotide abundance information for \strong{Human}
GRCh37. Its row names indicate 136 different types of 4 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatDNS136ToPdf}}.

\code{abundance.4bp.genome.GRCh38}, \code{abundance.4bp.exome.GRCh38} A
matrix containing tetranucleotide abundance information for \strong{Human}
GRCh38. Its row names indicate 136 different types of 4 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatDNS136ToPdf}}.

\code{abundance.4bp.genome.GRCm37}, \code{abundance.4bp.exome.GRCm37} A
matrix containing tetranucleotide abundance information for \strong{Mouse}
GRCm37. Its row names indicate 136 different types of 4 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatDNS136ToPdf}}.

\code{abundance.5bp.genome.GRCh37}, \code{abundance.5bp.exome.GRCh37} A
matrix containing pentanucleotide abundance information for \strong{Human}
GRCh37. Its row names indicate 512 different types of 5 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatSNS1536ToPdf}}.

\code{abundance.5bp.genome.GRCh38}, \code{abundance.5bp.exome.GRCh38} A
matrix containing pentanucleotide abundance information for \strong{Human}
GRCh38. Its row names indicate 512 different types of 5 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatSNS1536ToPdf}}.

\code{abundance.5bp.genome.GRCm37}, \code{abundance.5bp.exome.GRCm37} A
matrix containing pentanucleotide abundance information for \strong{Mouse}
GRCm37. Its row names indicate 512 different types of 5 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting function \code{\link{PlotCatSNS1536ToPdf}}.
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
\keyword{datasets}
\keyword{internal}
