% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AbundanceFile}
\alias{AbundanceFile}
\alias{abundance.2bp.exome.GRCh37}
\alias{abundance.2bp.genome.GRCh37}
\alias{abundance.3bp.exome.GRCh37}
\alias{abundance.3bp.genome.GRCh37}
\alias{abundance.4bp.exome.GRCh37}
\alias{abundance.4bp.genome.GRCh37}
\alias{abundance.5bp.exome.GRCh37}
\alias{abundance.5bp.genome.GRCh37}
\title{Nucleotide abundance file}
\format{A matrix containing nucleotide abundance information for different organism.}
\usage{
abundance.2bp.exome.GRCh37

abundance.2bp.genome.GRCh37

abundance.3bp.exome.GRCh37

abundance.3bp.genome.GRCh37

abundance.4bp.exome.GRCh37

abundance.4bp.genome.GRCh37

abundance.5bp.exome.GRCh37

abundance.5bp.genome.GRCh37
}
\description{
Nucleotide abundance information for a particular organism
}
\details{
\code{abundance.2bp.GRCh37} A matrix containing dinucleotide abundance information
for human GRCh37. Its row names indicate 10 different types of 2 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting functions \code{\link{PlotCatDNS78}} and
\code{\link{CatDNS78ToPdf}}.

\code{abundance.3bp.GRCh37} A matrix containing trinucleotide abundance information
for human GRCh37. Its row names indicate 32 different types of 3 base pairs
combinations while its column contains the occurrences of each type. It can
be used in plotting functions \code{\link{PlotCatSNS96}} and
\code{\link{CatSNS96ToPdf}}.

\code{abundance.4bp.GRCh37} A matrix containing tetranucleotide abundance
information for human GRCh37. Its row names indicate 136 different types of 4
base pairs combinations while its column contains the occurrences of each
type. It can be used in plotting functions \code{\link{PlotCatDNS136}} and
\code{\link{CatDNS136ToPdf}}.

\code{abundance.5bp.GRCh37} A matrix containing pentanucleotide abundance
information for human GRCh37. Its row names indicate 512 different types of 5
base pairs combinations while its column contains the occurrences of each
type. It can be used in plotting functions \code{\link{PlotCatSNS1536}} and
\code{\link{CatSNS1536ToPdf}}.
}
\note{
In the ID (insertion and deletion) catalog, the deletions repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
\keyword{datasets}
