% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{AddRunInformation}
\alias{AddRunInformation}
\title{Create a run information text file from generating zip archive from VCF
files.}
\usage{
AddRunInformation(
  files,
  vcf.names,
  zipfile.name,
  vcftype,
  ref.genome,
  region,
  nrow.data,
  mutation.loads
)
}
\description{
Create a run information text file from generating zip archive from VCF
files.
}
\keyword{internal}
