% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{AddSequence}
\alias{AddSequence}
\title{Add sequence context to a data frame with mutation records}
\usage{
AddSequence(df, genome)
}
\arguments{
\item{df}{An input data frame storing mutation records of a VCF file.}

\item{genome}{A reference genome as described in \code{\link{ICAMS}}.}
}
\value{
A copy of the input data.frame with a new column added
    that contains sequence context information.
}
\description{
Add sequence context to a data frame with mutation records
}
\keyword{internal}
