% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{AddSequenceID}
\alias{AddSequenceID}
\title{Add sequence context to a data frame with ID (insertion/deletion) mutation records}
\usage{
AddSequenceID(df, seq = BSgenome.Hsapiens.1000genomes.hs37d5)
}
\arguments{
\item{df}{A data frame storing mutation records of a VCF file
containing only insertions and deletions. IMPORTANT: The
representation of indels in df must have been canonicalized, so that
context bases (which are added by some indel callers) are placed in a
column "Left.context.base" and so that, for deletions, ALT is the empty
string, and, for insertions, REF is the empty string.}

\item{seq}{A particular reference genome.}
}
\value{
A data frame with 2 new columns added to the input data frame. One
  column contains sequence context information and the other column contains
  the length of the "context" string to the left of the site of the variant.
}
\description{
Add sequence context to a data frame with ID (insertion/deletion) mutation records
}
