% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{AliquotID2SampleID}
\alias{AliquotID2SampleID}
\title{AliquotID2SampleID}
\usage{
AliquotID2SampleID(aliquot.id)
}
\arguments{
\item{aliquot.id}{TODO}
}
\value{
TODO
}
\description{
AliquotID2SampleID
}
