% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CalBaseCountsFrom3MerAbundance}
\alias{CalBaseCountsFrom3MerAbundance}
\title{Calculate base counts from three mer abundance}
\usage{
CalBaseCountsFrom3MerAbundance(three.mer.abundance)
}
\description{
Calculate base counts from three mer abundance
}
\keyword{internal}
