% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{CalculateNumberOfSpace}
\alias{CalculateNumberOfSpace}
\title{Calculate the number of space needed to add strand bias statistics to
the run-information.txt file.}
\usage{
CalculateNumberOfSpace(list)
}
\arguments{
\item{list}{A list containing strand bias statistics.}
}
\value{
A matrix containing the space information.
}
\description{
Calculate the number of space needed to add strand bias statistics to
the run-information.txt file.
}
\keyword{internal}
