% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{Canonicalize1ID}
\alias{Canonicalize1ID}
\title{Given a single insertion or deletion in context categorize it.}
\usage{
Canonicalize1ID(context, ref, alt, pos, trace = 0)
}
\arguments{
\item{context}{Ample surrounding
sequence on each side of the insertion or deletion.}

\item{ref}{The reference allele (vector of length 1)}

\item{alt}{The alternative allele (vector of length 1)}

\item{pos}{The position of \code{ins.or.del.seq} in \code{context}.}

\item{trace}{If > 0 cat information how the computation is carried out.}
}
\value{
A string that is the canonical representation of the type of the given
 insertion or deletion.
}
\description{
Given a single insertion or deletion in context categorize it.
}
\keyword{internal}
