% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanonicalizeID.R
\name{CanonicalizeID}
\alias{CanonicalizeID}
\title{Determine the mutation types of insertions and deletions.}
\usage{
CanonicalizeID(context, ref, alt, pos)
}
\arguments{
\item{context}{A vector of ample surrounding
sequence on each side the variants}

\item{ref}{Vector of reference alleles; this includes
one un-altered base at the start of the reference allele
e.g. for a deletion of a single T this might be "AT", in
which case the \code{alt} allele woult be "A"}

\item{alt}{Vector of alternative alleles.}

\item{pos}{Vector of the positions of the insertions and deletions in
\code{context}. This is the position of the unaltered allele shared
by \code{ref} and \code{alt}. So in 1-based indexing, it is the
position just before the insertion or the deletion.}
}
\value{
A vector of strings that are the canonical representations
of the given insertions and deletions.
}
\description{
Determine the mutation types of insertions and deletions.
}
\keyword{internal}
