% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{CanonicalizeQUAD}
\alias{CanonicalizeQUAD}
\title{CanonicalizeQUAD}
\usage{
CanonicalizeQUAD(quad)
}
\arguments{
\item{quad}{TODO}
}
\value{
TODO
}
\description{
CanonicalizeQUAD
}
\keyword{internal}
