% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNS_catalog.R
\name{Cat1536ToPdf}
\alias{Cat1536ToPdf}
\title{Plot the 1536 mutation catalog of >= 1 samples to a PDF file}
\usage{
Cat1536ToPdf(catalog, name, id = colnames(catalog), abundance)
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the 1536 SNS mutation types
while its columns contain the counts of each mutation type from different
samples. The mutation types are in six-letters like CATTAT, first 2-letters
CA refers to (-2, -1) position, third letter T refers to the base which has
mutation, next second 2-letters TA refers to (+1, +2) position, last letter
T refers to the base after mutation.}

\item{name}{Name of the PDF file to be produced.}

\item{id}{A vector containing the identifier of each sample.}

\item{abundance}{A matrix containing pentanucleotide abundance information.}
}
\value{
invisible(TRUE)
}
\description{
Plot the 1536 mutation catalog of >= 1 samples to a PDF file
}
