% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNS_catalog.R
\name{Cat192StrandToPdf}
\alias{Cat192StrandToPdf}
\title{Plot the transcription strand bias graph of 6 SNS mutation types
("C>A", "C>G", "C>T", "T>A", "T>C", "T>G") of different samples to
a PDF file.}
\usage{
Cat192StrandToPdf(catalog, name, id = colnames(catalog),
  type = "counts", cex = 1, abundance = NULL)
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the 192 SNS mutation types
while its columns contain the counts of each mutation type from different
samples.}

\item{name}{The name of the PDF file to be produced.}

\item{id}{The ID information of the sample which has mutations.}

\item{type}{A vector of values indicating the type of graph for each sample.
If type = "counts", the graph will plot the occurrences of the 192 mutation
types in the sample. If type = "signature", the graph will plot mutation
signatures of the sample. If type = "density", the graph will plot the
rates of mutations per million trinucleotides for each mutation type. The
default value for type is "counts".}

\item{cex}{A numerical value giving the amount by which mutation class
labels, y axis labels, sample name and legend should be magnified relative
to the default.}

\item{abundance}{A matrix containing trinucleotide abundance and strand
information, to be used only when type = "density".}
}
\value{
invisible(TRUE)
}
\description{
Plot the transcription strand bias graph of 6 SNS mutation types
("C>A", "C>G", "C>T", "T>A", "T>C", "T>G") of different samples to
a PDF file.
}
