% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNS_catalog.R
\name{Cat192ToPdf}
\alias{Cat192ToPdf}
\title{Plot the SNS 192 mutation catalog of different samples to a PDF file}
\usage{
Cat192ToPdf(catalog, name, id = colnames(catalog), type = "counts",
  cex = 0.8, abundance = NULL)
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the 192 SNS mutation types
while its columns contain the counts of each mutation type from different
samples.}

\item{name}{The name of the PDF file to be produced.}

\item{id}{The ID information of the sample which has mutations.}

\item{type}{A vector of values indicating the type of graph for each sample.
If type = "counts", the graph will plot the occurrences of the 192 mutation
types in the sample. If type = "signature", the graph will plot mutation
signatures of the sample. If type = "density", the graph will plot the
rates of mutations per million trinucleotides for each mutation type. The
default value for type is "counts".}

\item{cex}{A numerical value giving the amount by which mutation class labels
on top of graph, y axis labels and sample name should be magnified relative
to the default.}

\item{abundance}{A matrix containing trinucleotide abundance and strand
information, to be used only when type = "density".}
}
\value{
invisible(TRUE)
}
\description{
Plot the SNS 192 mutation catalog of different samples to a PDF file
}
