% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catalog.row.order96}
\alias{catalog.row.order96}
\title{Canonical Order of a SNS 96 Catalog}
\format{A string of characters with length 96.}
\usage{
catalog.row.order96
}
\description{
The canonical order of row names in a SNS 96 catalog.
}
\keyword{datasets}
