% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CatalogRowHeaders}
\alias{CatalogRowHeaders}
\alias{catalog.row.headers.SNS.96}
\alias{catalog.row.headers.SNS.192}
\alias{catalog.row.headers.SNS.1536}
\alias{catalog.row.headers.DNS.78}
\alias{catalog.row.headers.DNS.144}
\alias{catalog.row.headers.DNS.136}
\alias{catalog.row.headers.ID}
\title{Row headers information for writing a catalog to disk in standardized format}
\format{A data frame which contains the row headers information for writing
  a catalog to disk in standardized format.}
\usage{
catalog.row.headers.SNS.96

catalog.row.headers.SNS.192

catalog.row.headers.SNS.1536

catalog.row.headers.DNS.78

catalog.row.headers.DNS.144

catalog.row.headers.DNS.136

catalog.row.headers.ID
}
\description{
Row headers information for writing a catalog to disk in standardized format
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
\keyword{datasets}
