% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DNS_catalog.R
\name{CatDNS144ToPdf}
\alias{CatDNS144ToPdf}
\title{Plot the transcription strand bias graph of 10 major DNS mutation types
("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN", "TA>NN",
"TC>NN", "TG>NN", "TT>NN") of different samples to a PDF file.}
\usage{
CatDNS144ToPdf(catalog, name, id = colnames(catalog), type = "counts",
  cex = 1, abundance = NULL)
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the 144 DNS mutation types
while its columns contain the counts of each mutation type from different
samples.}

\item{name}{The name of the PDF file to be produced.}

\item{id}{The ID information of the sample which has mutations.}

\item{type}{A vector of values indicating the type of graph for each sample.
If type = "counts", the graph will plot the occurrences of the 10 major DNS
mutation types in the sample. If type = "signature", the graph will plot
mutation signatures of the 10 major DNS mutation types in the sample. If
type = "density", the graph will plot the rates of mutations per million
dinucleotides for each of the 10 major DNS mutation types. The default
value for type is "counts".}

\item{cex}{A numerical value giving the amount by which mutation class
labels, y axis labels, sample name and legend should be magnified relative
to the default.}

\item{abundance}{A matrix containing dinucleotide abundance and strand
information, to be used only when type = "density".}
}
\value{
invisible(TRUE)
}
\description{
Plot the transcription strand bias graph of 10 major DNS mutation types
("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN", "TA>NN",
"TC>NN", "TG>NN", "TT>NN") of different samples to a PDF file.
}
