% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DNS_catalog.R
\name{CatDNS78ToPdf}
\alias{CatDNS78ToPdf}
\title{Plot the DNS 78 mutation catalog of different samples to a PDF file}
\usage{
CatDNS78ToPdf(catalog, name, id = colnames(catalog), type = "density",
  abundance = NULL)
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the 78 DNS mutation types
while its columns contain the counts of each mutation type from different
samples.}

\item{name}{The name of the PDF file to be produced.}

\item{id}{A vector containing the ID information of different samples.}

\item{type}{A vector of values indicating the type of plot for each sample.
If type = "density", the graph will plot the rates of mutations per million
nucleotides for each mutation type. If type = "counts", the graph will plot
the occurrences of the 78 mutation types in the sample. If type =
"signature", the graph will plot mutation signatures of the sample. The
default value for type is "density".}

\item{abundance}{A matrix containing dinucleotide abundance information, to
be used only when type = "density".}
}
\value{
invisible(TRUE)
}
\description{
Plot the DNS 78 mutation catalog of different samples to a PDF file
}
