% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_many_indels.R
\name{categorize_many_indels}
\alias{categorize_many_indels}
\title{Determine the mutation types of insertions and deletions.}
\usage{
categorize_many_indels(vcf, explain_indels = FALSE)
}
\arguments{
\item{context}{A vector of ample surrounding
sequence on each side the variants}

\item{ref}{Vector of reference alleles; this includes
one un-altered base at the start of the reference allele
e.g. for a deletion of a single T this might be "AT", in
which case the \code{alt} allele woult be "A"}

\item{alt}{Vector of alternative alleles.}

\item{pos}{Vector of the positions of the insertions and deletions in
 \code{context}. This is the position of the unaltered allele shared
by \code{ref} and \code{alt}. So in 1-based indexing, it is the
position just before the insertion or the deletion.}
}
\value{
A data frame parallel to the input vectors. The
data frame has the columns COSMIC_83, ins_or_del, previous_char, repeat_seq, repeat_count, post_char, mh_seq
}
\description{
Determine the mutation types of insertions and deletions.
}
\keyword{internal}
