% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{CaTypeAliquot2CaTypeSampleID}
\alias{CaTypeAliquot2CaTypeSampleID}
\title{CaTypeAliquot2CaTypeSampleID}
\usage{
CaTypeAliquot2CaTypeSampleID(ca.type.and.aliquot.id)
}
\arguments{
\item{ca.type.and.aliquot.id}{TODO}
}
\value{
TODO
}
\description{
CaTypeAliquot2CaTypeSampleID
}
