% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CheckAndReorderRownames}
\alias{CheckAndReorderRownames}
\title{Check whether the rownames of \code{object} are correct, if yes then put the
rows in the correct order.}
\usage{
CheckAndReorderRownames(object)
}
\description{
Check whether the rownames of \code{object} are correct, if yes then put the
rows in the correct order.
}
\keyword{internal}
