% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{CheckAndReturnIDCatalog}
\alias{CheckAndReturnIDCatalog}
\title{Check and return ID catalog}
\usage{
CheckAndReturnIDCatalog(catID, catID166, discarded.variants, annotated.vcfs)
}
\arguments{
\item{catID}{An ID catalog.}

\item{catID166}{An ID166 (genic-intergenic indel) catalog.}

\item{discarded.variants}{A list of discarded variants.}

\item{annotated.vcfs}{A list of annotated VCFs.}
}
\value{
A list of ID catalog. Also return the discarded variants and
annotated VCFs if they exit.
}
\description{
Check and return ID catalog
}
\keyword{internal}
