% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{CheckAndReturnSBSCatalogs}
\alias{CheckAndReturnSBSCatalogs}
\title{Check and return SBS catalogs}
\usage{
CheckAndReturnSBSCatalogs(
  catSBS96,
  catSBS1536,
  catSBS192 = NULL,
  discarded.variants,
  annotated.vcfs
)
}
\arguments{
\item{catSBS96}{An SBS96 catalog.}

\item{catSBS1536}{An SBS1536 catalog.}

\item{catSBS192}{An SBS192 catalog.}

\item{discarded.variants}{A list of discarded variants.}

\item{annotated.vcfs}{A list of annotated VCFs.}
}
\value{
A list of SBS catalogs. Also return the discarded variants and
annotated VCFs if they exit.
}
\description{
Check and return SBS catalogs
}
\keyword{internal}
