% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CheckCatalogAttribute}
\alias{CheckCatalogAttribute}
\title{Check attributes of catalog specified by user}
\usage{
CheckCatalogAttribute(ref.genome, region, catalog.type)
}
\arguments{
\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".

TODO(Nanhai) need to added "transcript" here also (and in other functions)}

\item{catalog.type}{One of "counts", "density", "counts.signature",
"density.signature".}
}
\value{
TRUE
}
\description{
Check attributes of catalog specified by user
}
\keyword{internal}
