% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CheckClassOfCatalog}
\alias{CheckClassOfCatalog}
\title{Check the class of catalog}
\usage{
CheckClassOfCatalog(catalog)
}
\arguments{
\item{catalog}{An S3 object with class "catalog".}
}
\value{
an object with the corresponding class type of catalog.
}
\description{
Check the class of catalog
}
\keyword{internal}
