% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CheckClassOfCatalogFromPath}
\alias{CheckClassOfCatalogFromPath}
\title{Check the class of catalog from path}
\usage{
CheckClassOfCatalogFromPath(path)
}
\arguments{
\item{path}{Path to a catalog on disk in the standardized format.}
}
\value{
An object with the corresponding class type of catalog.
}
\description{
Check the class of catalog from path
}
\keyword{internal}
