% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{CheckDBSClassInVCF}
\alias{CheckDBSClassInVCF}
\title{Check DBS mutation class in VCF with the corresponding DBS mutation matrix}
\usage{
CheckDBSClassInVCF(vcf, mat, sample.id)
}
\arguments{
\item{vcf}{An annotated DBS VCF with columns of DBS mutation
classes added by \code{AddDBSClass}.}

\item{mat}{The DBS mutation count matrix.}

\item{sample.id}{Usually the sample id, but defaults to "count".}
}
\description{
Check DBS mutation class in VCF with the corresponding DBS mutation matrix
}
\keyword{internal}
