% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF_etc.R
\name{CheckSeqContextInVCF}
\alias{CheckSeqContextInVCF}
\title{Check that the sequence context information is consistent with the value of
the column REF.}
\usage{
CheckSeqContextInVCF(vcf, column.to.use)
}
\arguments{
\item{vcf}{In-memory VCF as a data.frame; must be an SBS or DBS VCF.}

\item{column.to.use}{The column name as a string of the column in the VCF
with the context information.}
}
\value{
Throws error with location information if the value of REF is
inconsistent with the value of seq.21bases. Assumes the first base of the
reference allele is at position (size(\if{html}{\out{<context string>}})-1)/2, and generates
error if this is not an integer. Indices are 1-based.
}
\description{
Check that the sequence context information is consistent with the value of
the column REF.
}
\keyword{internal}
