% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{Collapse144to78}
\alias{Collapse144to78}
\title{Collapse a DNS 144 catalog matrix to a DNS 78 catalog matrix}
\usage{
Collapse144to78(catDNS144)
}
\arguments{
\item{catDNS144}{A DNS 144 catalog matrix whose row names indicate the 192
mutation types while its columns show the occurrences of each mutation type of
different samples.}
}
\value{
A DNS 78 catalog matrix whose row names indicate the 96 mutation
  types while its columns show the occurrences of each mutation type of different
  samples.
}
\description{
Collapse a DNS 144 catalog matrix to a DNS 78 catalog matrix
}
