% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNS_related_functions.R
\name{Collapse1536to96}
\alias{Collapse1536to96}
\title{Collapse a SNS 1536 catalog matrix to a 96 catalog matrix}
\usage{
Collapse1536to96(cat1536)
}
\arguments{
\item{cat1536}{A SNS 1536 catalog matrix whose row names indicate the 1536
mutation types while its columns show the occurrences of each mutation type
of different samples.}
}
\value{
A SNS 96 catalog matrix whose row names indicate the 96
  mutation types while its columns show the occurrences of
  each mutation type of different samples.
}
\description{
Collapse a SNS 1536 catalog matrix to a 96 catalog matrix
}
