% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNS_related_functions.R
\name{Collapse192to96}
\alias{Collapse192to96}
\title{Collapse a SNS 192 catalog matrix to a 96 catalog matrix}
\usage{
Collapse192to96(cat192)
}
\arguments{
\item{cat192}{A SNS 192 catalog matrix whose row names indicate the 192
mutation types while its columns show the occurrences of each mutation type of
different samples.}
}
\value{
A SNS 96 catalog matrix whose row names indicate the 96
}
\description{
Collapse a SNS 192 catalog matrix to a 96 catalog matrix
}
