% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{ConvertICAMSCatalogToSigProSBS96}
\alias{ConvertICAMSCatalogToSigProSBS96}
\title{Covert an ICAMS SBS96 Catalog to SigProfiler format}
\usage{
ConvertICAMSCatalogToSigProSBS96(input.catalog, file, sep = "\\t")
}
\arguments{
\item{input.catalog}{Either a character string, in which case this is the
path to a file containing a catalog in \code{\link[ICAMS]{ICAMS}}
format, or an in-memory \code{\link[ICAMS]{ICAMS}} catalog.}

\item{file}{The path of the file to be written.}

\item{sep}{Separator to use in the output file.}
}
\description{
Covert an ICAMS SBS96 Catalog to SigProfiler format
}
\keyword{internal}
