% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateCatalogAttribute}
\alias{CreateCatalogAttribute}
\title{Create attributes of a catalog}
\usage{
CreateCatalogAttribute(catalog, ref.genome, region, type)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}}.}

\item{ref.genome}{A genome argument as described in \code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}

\item{type}{A character string acting as a catalog type identifier, one of
"counts", "density", "counts.signature" and "density.signature".}
}
\value{
The original catalog with the following attributes added: ref.genome,
  region, type, abundance, class.
}
\description{
Create attributes of a catalog
}
