% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateCatalogClass}
\alias{CreateCatalogClass}
\title{Create the class attribute of a catalog}
\usage{
CreateCatalogClass(catalog)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}}.}
}
\value{
The original catalog with class attribute added.
}
\description{
Create the class attribute of a catalog
}
\keyword{internal}
