% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{CreateOneColDBSCatalog}
\alias{CreateOneColDBSCatalog}
\title{Create double base catalog for *one* sample from
a Variant Call Format (VCF) file}
\usage{
CreateOneColDBSCatalog(vcf, trans.ranges = NULL, sample.id = "count")
}
\arguments{
\item{vcf}{An in-memory VCF file annotated by the AddSeqContext and
AddTranscript functions. It must *not* contain indels and must
*not* contain SBS (single base substitutions), or triplet base
substitutions etc.}

\item{trans.ranges}{A \code{\link[data.table]{data.table}} which contains
transcript range and strand information. Please refer to
\code{\link{TranscriptRanges}} for more details.}

\item{sample.id}{Usually the sample id, but defaults to "count".}
}
\value{
A list of three matrices containing the DBS catalog: catDBS78,
  catDBS144, catDBS136 respectively. If trans.ranges = NULL, DBS 144 catalog
  will not be generated.
}
\description{
Create double base catalog for *one* sample from
a Variant Call Format (VCF) file
}
\note{
DBS 144 catalog only contains mutations in transcribed regions.
}
\keyword{internal}
