% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{CreateOneColDNSCatalog}
\alias{CreateOneColDNSCatalog}
\title{Create double nucleotide catalog for *one* sample from
a Variant Call Format (VCF) file}
\usage{
CreateOneColDNSCatalog(vcf, sample.id = "count")
}
\arguments{
\item{vcf}{An in-memory VCF file annotated by the AddSequence and
AddTranscript functions. It must *not* contain indels and must
*not* contain SNS (single nucleotide substituions), or triplet base
substitutions etc.}

\item{sample.id}{Usually the sample id, but defaults to "count".}
}
\value{
A list of three matrices containing the DNS catalog:
  catDNS78, catDNS144, catDNS136 respectively.
}
\description{
Create double nucleotide catalog for *one* sample from
a Variant Call Format (VCF) file
}
\keyword{internal}
