% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{CreateOneColIDCatalog}
\alias{CreateOneColIDCatalog}
\title{Create one column of an indel catalog from one VCF}
\usage{
CreateOneColIDCatalog(ID.vcf, SBS.vcf, trace = 0)
}
\arguments{
\item{ID.vcf}{An in-memory VCF as a data.frame annotated by the
  \code{\link{AddAndCheckSequenceID}} function. It must only
  contain indels and must \strong{not} contain SNSs
  (single nucleotide/base substituions), DBS, or triplet
  base substituions, etc.

  One design decision for variant callers is the representation of "complex
  indels", e.g. mutations e.g. CAT > GC. Some callers represent this as C>G,
  A>C, and T>_. Others might represent it as CAT > CG. Multiple issues can
  arise. In PCAWG, overlapping indel/SBS calls from different callers were
  included in the indel VCFs.}

\item{SBS.vcf}{An in-memory VCF as a data frame. Because we have to work with
some PCAWG data, we will look for neigboring indels and indels adjoining
SBS. That means this functions takes an SBS VCF and an ID VCF from the same
sample.}

\item{trace}{If > 0, various called functions cat information
useful for debugging and testing. The larger the number, the
more output.}
}
\value{
A list with two elemsents:
  ID.cat:   A 1-column matrix containing the mutation catalog information.
  problems: Locations of neighboring indels or indels neighboring SBS.
            In the future we might handle these depending on what we
            find in the indel calls from different variant callers.
TODO(steve) Is problems implemented?
}
\description{
Create one column of an indel catalog from one VCF
}
