% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_context_functions.R
\name{CreateOnePWMFromSBSVCF}
\alias{CreateOnePWMFromSBSVCF}
\title{Create position weight matrix (PWM) for *one* sample from
a Variant Call Format (VCF) file.}
\usage{
CreateOnePWMFromSBSVCF(vcf, ref.genome, seq.context.width)
}
\arguments{
\item{vcf}{One in-memory data frame of pure SBS mutations -- no DBS or 3+BS
mutations.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{seq.context.width}{The number of preceding and following bases to be
extracted around the mutated position from \code{ref.genome}.}
}
\value{
A position weight matrix (PWM) with each row representing the
  position weight of four bases.
}
\description{
Create position weight matrix (PWM) for *one* sample from
a Variant Call Format (VCF) file.
}
\keyword{internal}
