% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_context_functions.R
\name{CreatePPMFromSBSVCFs}
\alias{CreatePPMFromSBSVCFs}
\title{Create position probability matrices (PPM) from a list of SBS vcfs}
\usage{
CreatePPMFromSBSVCFs(list.of.SBS.vcfs, ref.genome, seq.context.width)
}
\arguments{
\item{list.of.SBS.vcfs}{List of in-memory data frames of pure SBS mutations
-- no DBS or 3+BS mutations.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{seq.context.width}{The number of preceding and following bases to be
extracted around the mutated position from \code{ref.genome}.}
}
\value{
A list of position probability matrices (PPM).
}
\description{
Create position probability matrices (PPM) from a list of SBS vcfs
}
\keyword{internal}
