% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{CreateSampleId}
\alias{CreateSampleId}
\title{CreateSampleId}
\usage{
CreateSampleId(filename, aliquot.id)
}
\arguments{
\item{filename}{TODO}

\item{aliquot.id}{TODO}
}
\value{
TODO
}
\description{
Get catalog sorting function from msigtools
}
