% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateTetranucAbundance}
\alias{CreateTetranucAbundance}
\title{Create tetranucleotide abundance}
\usage{
CreateTetranucAbundance(file)
}
\arguments{
\item{file}{Path to the file with the nucleotide abundance information with 4
base pairs.}
}
\value{
A numeric vector whose names indicate 136 different types of 4 base pairs
combinations while its values indicate the occurrences of each type.
}
\description{
Create tetranucleotide abundance
}
\keyword{internal}
