% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateTransRange}
\alias{CreateTransRange}
\title{Create a Transcript Range file from the raw GFF3 File}
\usage{
CreateTransRange(path)
}
\arguments{
\item{path}{The name/path of the raw GFF3 File, or a complete URL.}
}
\value{
A data frame which contains chromosome name, start, end position,
  strand information and gene name. Only the following four gene types are
  kept to facilitate transcriptional strand bias analysis: protein_coding,
  retained_intron, processed_transcript and nonsense_mediated_decay.
}
\description{
Create a Transcript Range file from the raw GFF3 File
}
\keyword{internal}
