% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateTransRanges}
\alias{CreateTransRanges}
\title{Create a transcript range file from the raw GFF3 File}
\usage{
CreateTransRanges(file)
}
\arguments{
\item{file}{The name/path of the raw GFF3 File, or a complete URL.}
}
\value{
A data table which contains chromosome name, start, end position,
strand information and gene name. It is keyed by chrom, start, and
end. Only genes that are associated with a CCDS ID are kept for
transcriptional strand bias analysis.
}
\description{
Create a transcript range file from the raw GFF3 File
}
\keyword{internal}
