% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{DirectoryOfSBSSimple2Catalog}
\alias{DirectoryOfSBSSimple2Catalog}
\title{DirectoryOfSBSSimple2Catalog}
\usage{
DirectoryOfSBSSimple2Catalog(path.to.dir, cts = empty.cats,
  verbose = 3)
}
\arguments{
\item{path.to.dir}{TODO}

\item{cts}{TODO}

\item{verbose}{TODO}
}
\value{
TODO
}
\description{
DirectoryOfSBSSimple2Catalog
}
