% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{.ReadAndSplitStrelkaSBSVCFs}
\alias{.ReadAndSplitStrelkaSBSVCFs}
\title{The argument updateProgress is to be used in ICAMS.shiny package.}
\usage{
.ReadAndSplitStrelkaSBSVCFs(files, names.of.VCFs = NULL, updateProgress = NULL)
}
\description{
The argument updateProgress is to be used in ICAMS.shiny package.
}
\keyword{internal}
