% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{.StrelkaSBSVCFFilesToCatalog}
\alias{.StrelkaSBSVCFFilesToCatalog}
\title{The argument updateProgress is to be used in ICAMS.shiny package.}
\usage{
.StrelkaSBSVCFFilesToCatalog(
  files,
  ref.genome,
  trans.ranges = NULL,
  region = "unknown",
  names.of.VCFs = NULL,
  updateProgress = NULL
)
}
\description{
The argument updateProgress is to be used in ICAMS.shiny package.
}
\keyword{internal}
