% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{.StrelkaSBSVCFFilesToZipFile}
\alias{.StrelkaSBSVCFFilesToZipFile}
\title{The argument updateProgress is to be used in ICAMS.shiny package.}
\usage{
.StrelkaSBSVCFFilesToZipFile(
  dir,
  zipfile,
  ref.genome,
  trans.ranges = NULL,
  region = "unknown",
  names.of.VCFs = NULL,
  base.filename = ""
)
}
\description{
The argument updateProgress is to be used in ICAMS.shiny package.
}
\keyword{internal}
