% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{.VCFsToDBSCatalogs}
\alias{.VCFsToDBSCatalogs}
\title{The argument updateProgress is to be used in ICAMS.shiny package.}
\usage{
.VCFsToDBSCatalogs(
  list.of.DBS.vcfs,
  ref.genome,
  trans.ranges = NULL,
  region = "unknown",
  updateProgress = NULL
)
}
\description{
The argument updateProgress is to be used in ICAMS.shiny package.
}
\keyword{internal}
