% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{.VCFsToIDCatalogs}
\alias{.VCFsToIDCatalogs}
\title{The argument updateProgress is to be used in ICAMS.shiny package.}
\usage{
.VCFsToIDCatalogs(
  list.of.vcfs,
  ref.genome,
  region = "unknown",
  updateProgress = NULL
)
}
\description{
The argument updateProgress is to be used in ICAMS.shiny package.
}
\keyword{internal}
