% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{GenerateEmptyKmerCounts}
\alias{GenerateEmptyKmerCounts}
\title{Generate an empty matrix of k-mer abundance}
\usage{
GenerateEmptyKmerCounts(k)
}
\arguments{
\item{k}{Length of k-mers (k>=2)}
}
\value{
An empty matrix of k-mer abundance
}
\description{
Generate an empty matrix of k-mer abundance
}
\keyword{internal}
