% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{GenerateKmer}
\alias{GenerateKmer}
\title{Generate permuation of k mers}
\usage{
GenerateKmer(k)
}
\arguments{
\item{k}{Length of kmers (k>=2)}
}
\value{
A vector of characters which have all the permuations of k mers
}
\description{
Generate permuation of k mers
}
\keyword{internal}
