% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_file_functions.R
\name{GenerateVCFsFromIndelSimpleFile1}
\alias{GenerateVCFsFromIndelSimpleFile1}
\title{Generate reconstructed VCFs from indel (small insertions and deletions) simple file}
\usage{
GenerateVCFsFromIndelSimpleFile1(file, output.dir, max.mc.cores = 1)
}
\arguments{
\item{file}{The name/path of the simple indel file, or a complete URL.}

\item{output.dir}{The directory where the reconstructed VCFs will be saved.}

\item{max.mc.cores}{The maximum number of cores to use. On Microsoft Windows
machines it is silently changed to 1.}
}
\description{
Generate reconstructed VCFs from indel (small insertions and deletions) simple file
}
\keyword{internal}
