% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_file_functions.R
\name{GenerateVCFsFromIndelSimpleFiles}
\alias{GenerateVCFsFromIndelSimpleFiles}
\title{Generate reconstructed VCFs from indel (small insertions and deletions) simple files}
\usage{
GenerateVCFsFromIndelSimpleFiles(
  files,
  output.dir,
  num.parallel.files = 1,
  mc.cores.per.file = 1
)
}
\arguments{
\item{files}{Character vector of file paths to the indel simple files.}

\item{output.dir}{The directory where the reconstructed VCFs will be saved.}

\item{num.parallel.files}{The (maximum) number of files to run in
parallel. On Microsoft Windows machines it is silently changed to 1. Each
file in turn can require multiple cores, as governed by
\code{mc.cores.per.file}.}

\item{mc.cores.per.file}{The maximum number of cores to use for each
file. On Microsoft Windows machines it is silently changed to 1.}
}
\description{
Generate reconstructed VCFs from indel (small insertions and deletions) simple files
}
\keyword{internal}
