% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{GetConsensusVAF}
\alias{GetConsensusVAF}
\title{Analogous to \code{\link{GetMutectVAF}}, calculating VAF and read depth
from PCAWG7 consensus vcfs}
\usage{
GetConsensusVAF(vcf)
}
\description{
Analogous to \code{\link{GetMutectVAF}}, calculating VAF and read depth
from PCAWG7 consensus vcfs
}
\keyword{internal}
