% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{GetGenomeKmers}
\alias{GetGenomeKmers}
\title{Generate kmer abundances from a given genome}
\usage{
GetGenomeKmers(k, genome, homopolymer.filter = FALSE)
}
\arguments{
\item{k}{Length of kmers (k>=2)}

\item{genome}{Name of a particular reference genome
(without quotations marks)}

\item{homopolymer.filter}{Optional. If TRUE, homopolymers will be masked from
genome(sequence)}
}
\value{
Matrix of the counts of each kmer across the genome
}
\description{
Generate kmer abundances from a given genome
}
\keyword{internal}
