% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{GetMutectVAF}
\alias{GetMutectVAF}
\title{Extract the VAFs (variant allele frequencies) from a VCF created by
    MuTect}
\usage{
GetMutectVAF(mutect.vcf)
}
\arguments{
\item{mutect.vcf}{said VCF as a data.frame}
}
\value{
A vector of VAFs, one for each row of mutect.vcf
}
\description{
Extract the VAFs (variant allele frequencies) from a VCF created by
    MuTect
}
