% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{GetSequenceKmerCounts}
\alias{GetSequenceKmerCounts}
\title{Generate k-mer abundance from given nucleotide sequences}
\usage{
GetSequenceKmerCounts(sequences, k)
}
\arguments{
\item{sequences}{A vector of nucleotide sequences}

\item{k}{Length of k-mers (k>=2)}
}
\value{
Matrix of the counts of each k-mer inside \code{sequences}
}
\description{
Generate k-mer abundance from given nucleotide sequences
}
\keyword{internal}
