% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{GetStrandedKmers}
\alias{GetStrandedKmers}
\title{Generate stranded kmer abundances from a given genome and gene annotation file}
\usage{
GetStrandedKmers(k, genome, trans.ranges, homopolymer.filter = FALSE)
}
\arguments{
\item{k}{Length of kmers (k>=2)}

\item{genome}{Name of a particular reference genome
(without quotations marks)}

\item{homopolymer.filter}{Optional. If TRUE, homopolymers will be masked from
genome(sequence)}

\item{trans.range}{A GFF3 trans.range.file}
}
\value{
Matrix of the counts of each kmer across the genome
}
\description{
Generate stranded kmer abundances from a given genome and gene annotation file
}
\keyword{internal}
