% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_related_functions.R
\name{GetStrelkaVAF}
\alias{GetStrelkaVAF}
\title{Extract the VAFs (variant allele frequencies) from a VAF created by
    Strelka version 1}
\usage{
GetStrelkaVAF(strelka.vcf)
}
\arguments{
\item{strelka.vcf}{said VCF as a data.frame}
}
\value{
A vector of VAFs, one for each row of strelka.vcf
}
\description{
Extract the VAFs (variant allele frequencies) from a VAF created by
    Strelka version 1
}
