% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICAMS.to.SigPro.ID}
\alias{ICAMS.to.SigPro.ID}
\title{Linker from PCAWG(ICAMS)-formatted to SigProExtractor-formatted indel names.}
\format{A 83*1 matrix. Its contents (first column) contain SigProExtractor
formatted indel names in SigProExtractor order. Its rownames refer to the 
corresponding PCAWG(ICAMS)-formatted indel names.}
\usage{
ICAMS.to.SigPro.ID
}
\description{
This data is designed for converting ICAMS-formatted indel names to
SigProExtractor-formatted indel names.
}
\keyword{datasets}
