% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{InferAbundance}
\alias{InferAbundance}
\title{Infer \code{abundance} given a matrix-like \code{object} and additional information.}
\usage{
InferAbundance(object, ref.genome, region, catalog.type)
}
\arguments{
\item{object}{A numeric matrix, numeric data frame, or \code{catalog}.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{A character string designating a genomic region;
see \code{\link{as.catalog}} and \code{\link{ICAMS}}.}

\item{catalog.type}{A character string for \code{catalog.type}
as described in \code{\link{ICAMS}}.}
}
\value{
A value that can be set as the abundance attribute of
a \code{catalog} (which may be \code{NULL} if no abundance
can be inferred).
}
\description{
Infer \code{abundance} given a matrix-like \code{object} and additional information.
}
\keyword{internal}
