% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_catalog_format.R
\name{InferCatalogInfo}
\alias{InferCatalogInfo}
\title{This function converts an data.table imported
from external catalog text file into ICAMS
internal catalog object of appropriate type.}
\usage{
InferCatalogInfo(object)
}
\description{
This function converts an data.table imported
from external catalog text file into ICAMS
internal catalog object of appropriate type.
}
\keyword{internal}
