% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{InferClassOfCatalogForRead}
\alias{InferClassOfCatalogForRead}
\title{Infer the class of catalog in a file.}
\usage{
InferClassOfCatalogForRead(file)
}
\arguments{
\item{file}{Path to a catalog on disk in the standardized format.}
}
\value{
A character string with class appropriate for the catalog on disk.
}
\description{
Infer the class of catalog in a file.
}
\keyword{internal}
