% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{InferRefGenomeName}
\alias{InferRefGenomeName}
\title{Infer reference genome name from a character string}
\usage{
InferRefGenomeName(ref.genome)
}
\arguments{
\item{ref.genome}{A character string indicating the reference genome.}
}
\value{
The inferred reference genome name.
}
\description{
Infer reference genome name from a character string
}
\keyword{internal}
