% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{InferRownames}
\alias{InferRownames}
\title{Infer the correct rownames for a matrix based on its number of rows}
\usage{
InferRownames(object)
}
\description{
Infer the correct rownames for a matrix based on its number of rows
}
\keyword{internal}
