% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{InitAliquotSample}
\alias{InitAliquotSample}
\title{InitAliquotSample}
\usage{
InitAliquotSample()
}
\value{
TODO
}
\description{
In InitAliquotSample, note the hack to add one more mapping. One aliquot ID
is not in PCAWG as per ZHANG Junjun's email
https://mail.google.com/mail/u/0/#search/Junjun.Zhang%40oicr.on.ca++pcawg_wgs/FMfcgxmTmlJdCfGhxbDHpsJCwknCNTDw
[1] "CMDI-UK::f92cf0a2-0172-a1cc-e040-11ac0c486b1a" >
setdiff(StripCaType(colnames(ofcat96)), StripCaType(ycol)) [1]
"CMDI-UK::SP116883" was the associated sample id. See
https://mail.google.com/mail/u/0/#search/SP116883/FMfcgxmTmlJdCfGhxbDHpsJCwknCNTDw
}
