% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{IsGRCh37}
\alias{IsGRCh37}
\title{Test if object is \code{BSgenome.Hsapiens.1000genome.hs37d5}.}
\usage{
IsGRCh37(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE if \code{x} is \code{BSgenome.Hsapiens.1000genome.hs37d5}.
}
\description{
Test if object is \code{BSgenome.Hsapiens.1000genome.hs37d5}.
}
\keyword{internal}
