% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{IsGRCh38}
\alias{IsGRCh38}
\title{Test if object is \code{BSgenome.Hsapiens.UCSC.hg38}.}
\usage{
IsGRCh38(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE if \code{x} is \code{BSgenome.Hsapiens.UCSC.hg38}.
}
\description{
Test if object is \code{BSgenome.Hsapiens.UCSC.hg38}.
}
\keyword{internal}
