% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{IsGRCm38}
\alias{IsGRCm38}
\title{Test if object is \code{BSgenome.Mmusculus.UCSC.mm10}.}
\usage{
IsGRCm38(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE if \code{x} is \code{BSgenome.Mmusculus.UCSC.mm10}.
}
\description{
Test if object is \code{BSgenome.Mmusculus.UCSC.mm10}.
}
\keyword{internal}
