% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{IsRefGenomeInstalled}
\alias{IsRefGenomeInstalled}
\title{Check whether the BSgenome package is installed}
\usage{
IsRefGenomeInstalled(ref.genome)
}
\arguments{
\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}
}
\value{
A logical value indicating whether the BSgenome package is installed.
}
\description{
Check whether the BSgenome package is installed
}
\keyword{internal}
