% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{IsTransformationfromCountsLegal}
\alias{IsTransformationfromCountsLegal}
\title{Source catalog type is counts or counts.signature}
\usage{
IsTransformationfromCountsLegal(s, t)
}
\description{
counts -> <anything>
counts.signature -> counts.signature, density.signature
}
\keyword{internal}
