% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{IsTransformationfromDensityLegal}
\alias{IsTransformationfromDensityLegal}
\title{density -> <anything>
density.signature -> density.signature, counts.signature}
\usage{
IsTransformationfromDensityLegal(s, t)
}
\description{
density -> <anything>
density.signature -> density.signature, counts.signature
}
\keyword{internal}
