% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justify_and_categorize_1_indel.R
\name{justify_and_categorize_1_indel}
\alias{justify_and_categorize_1_indel}
\title{Given a single insertion or deletion in context categorize it.}
\usage{
justify_and_categorize_1_indel(
  context,
  orig_ref,
  orig_alt,
  orig_pos,
  explain_indels = FALSE,
  regress = TRUE,
  remove_common_prefix = FALSE
)
}
\arguments{
\item{context}{Ample surrounding
sequence on each side of the insertion or deletion.}

\item{orig_ref}{The reference allele as a single string.}

\item{orig_alt}{The alternative allele as a single string}

\item{orig_pos}{The position of the inserted or deleted sequence in \code{context}.
If this is an insertion, it is the positon after the position of the insertion of \code{alt}.
If this is a deletion, it is the position in context at which \code{ref} starts.}

\item{explain_indels}{If TRUE then generate output explaining the processing}

\item{regress}{If TRUE than compare COSMIC-83 classification to results from
the previous ICAMS implementation}

\item{remove_common_prefix}{Set this to TRUE if the orig_alt and orig_ref
strings share the first character, as is common for most indel callers.
E.g. a the deletion ATGC to ATC show the ref allele as TG and the
mutated allele as T. T is the common prefix.
#}
}
\value{
A string that is the canonical representation
 of the type of the given
 insertion or deletion.
 Return \code{NA}
 and raise a warning if
 there is an un-normalized representation of
 the deletion of a repeat unit.
 See \code{FindDelMH} for details.
 (This seems to be very rare.)
}
\description{
Given a single insertion or deletion in context categorize it.
}
\keyword{internal}
