% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF_etc.R
\name{MakeDataFrameFromVCF}
\alias{MakeDataFrameFromVCF}
\title{Read in the data lines of a Variant Call Format (VCF) file}
\usage{
MakeDataFrameFromVCF(file)
}
\arguments{
\item{file}{The name/path of the VCF file, or a complete URL.}
}
\value{
A data frame storing mutation records of a VCF file.
}
\description{
Read in the data lines of a Variant Call Format (VCF) file
}
\keyword{internal}
