% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{MakeDataFrameFromVCFNew}
\alias{MakeDataFrameFromVCFNew}
\title{Read in the data lines of a Variant Call Format (VCF) file}
\usage{
MakeDataFrameFromVCFNew(file, suppress.discarded.variants.warnings = TRUE)
}
\arguments{
\item{file}{The name/path of the VCF file, or a complete URL.}

\item{suppress.discarded.variants.warnings}{Logical. Whether to suppress
warning messages showing information about the discarded variants. Default
is TRUE.}
}
\value{
A list of elements:
\itemize{
\item \code{df}: A data frame storing data lines from the original VCF file.
\item \code{discarded.variants}: \strong{Non-NULL only if} variants
were excluded.
}
}
\description{
Read in the data lines of a Variant Call Format (VCF) file
}
\keyword{internal}
