% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{NewTestMakeCatalogFromStrelkaSNSVCFs}
\alias{NewTestMakeCatalogFromStrelkaSNSVCFs}
\title{This function is to make catalogs from the sample VCF files
to compare with the expected catalog information.}
\usage{
NewTestMakeCatalogFromStrelkaSNSVCFs()
}
\description{
This function is to make catalogs from the sample VCF files
to compare with the expected catalog information.
}
