% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{NewTestStrelkaSNSCatalog}
\alias{NewTestStrelkaSNSCatalog}
\title{This function is to test whether the predefined functions
are working correctly to produce the desired SNS catalogs from Strelka VCF.}
\usage{
NewTestStrelkaSNSCatalog()
}
\description{
This function is to test whether the predefined functions
are working correctly to produce the desired SNS catalogs from Strelka VCF.
}
