% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{NormalizeAbundanceArg}
\alias{NormalizeAbundanceArg}
\title{Handle abundance (opportunity) specications uniformly.}
\usage{
NormalizeAbundanceArg(abundance, which.n)
}
\arguments{
\item{abundance}{Either an abunance variable or string specifying an abundance.}

\item{which.n}{The n for the n-mers, one of 2, 3, 4, 5 for 2-mers, 3-mers, etc.}
}
\description{
Handle abundance (opportunity) specications uniformly.
}
\keyword{internal}
