% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{OldTestStrelkaDNSCatalog}
\alias{OldTestStrelkaDNSCatalog}
\title{This function is to test whether the predefined functions
are working correctly to produce the desired DNS catalogs from Strelka VCF.}
\usage{
OldTestStrelkaDNSCatalog()
}
\description{
This function is to test whether the predefined functions
are working correctly to produce the desired DNS catalogs from Strelka VCF.
}
\keyword{internal}
