% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{OneSBSVCFAllCatalogs}
\alias{OneSBSVCFAllCatalogs}
\title{OneSBSVCFAllCatalogs}
\usage{
OneSBSVCFAllCatalogs(vcf, genome, sample.id = "count", verbose = 0)
}
\arguments{
\item{vcf}{TODO}

\item{genome}{TODO}

\item{sample.id}{TODO}

\item{verbose}{TODO}
}
\value{
TODO
}
\description{
TODO(steve) merge this in where needed
}
