% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPOSITE.R
\name{Plot96PartOfCompositeToPDF}
\alias{Plot96PartOfCompositeToPDF}
\title{Plot the SBS96 part of a SignatureAnalyzer COMPOSITE signature or catalog}
\usage{
Plot96PartOfCompositeToPDF(catalog, name, type = "density")
}
\arguments{
\item{catalog}{Catalog or signature matrix}

\item{name}{Name of file to print to.}

\item{type}{See \code{\link[ICAMS]{PlotCatalogToPdf}}.}
}
\description{
Plot the SBS96 part of a SignatureAnalyzer COMPOSITE signature or catalog
}
\keyword{internal}
