% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNS_catalog.R
\name{PlotCat1536}
\alias{PlotCat1536}
\title{Plot the pentanucleotide sequence contexts for one sample,
normalized by pentanucleotide occurrence in the genome.}
\usage{
PlotCat1536(catalog, id, scale = TRUE, abundance)
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the 1536 SNS mutation types
while its column contains the counts of each mutation type. The mutation
types are in six-letters like CATTAT, first 2-letters CA refers to (-2, -1)
position, third letter T refers to the base which has mutation, next second
2-letters TA refers to (+1, +2) position, last letter T refers to the base
after mutation.}

\item{id}{The id of the sample to be displayed on top of the graph.}

\item{scale}{A logical value indicating whether to do color scaling for all
mutation types.}

\item{abundance}{A matrix containing pentanucleotide abundance information.}
}
\value{
invisible(TRUE)
}
\description{
Plot the pentanucleotide sequence contexts for one sample,
normalized by pentanucleotide occurrence in the genome.
}
