% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNS_catalog.R
\name{PlotCat192Strand}
\alias{PlotCat192Strand}
\title{Plot the transcription strand bias graph of 6 SNS mutation types
("C>A", "C>G", "C>T", "T>A", "T>C", "T>G") in one sample}
\usage{
PlotCat192Strand(catalog, id, type = "counts", cex = 1,
  abundance = NULL)
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the 192 SNS mutation types
while its column contains the counts of each mutation type.}

\item{id}{The ID information of the sample which has mutations.}

\item{type}{A value indicating the type of the graph. If type = "counts", the
graph will plot the occurrences of the 6 SNS mutation types in the sample.
If type = "signature", the graph will plot mutation signatures of the 6 SNS
mutation types in the sample. If type = "density", the graph will plot the
rates of mutations per million trinucleotides for each of the 6 SNS
mutation types. The default value for type is "counts".}

\item{cex}{A numerical value giving the amount by which mutation class
labels, y axis labels, sample name and legend should be magnified relative
to the default.}

\item{abundance}{A matrix containing trinucleotide abundance and strand
information, to be used only when type = "density".}
}
\value{
invisible(TRUE)
}
\description{
Plot the transcription strand bias graph of 6 SNS mutation types
("C>A", "C>G", "C>T", "T>A", "T>C", "T>G") in one sample
}
