% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNS_catalog.R
\name{PlotCat96New}
\alias{PlotCat96New}
\title{Plot the SNS 96 mutation catalog of one sample}
\usage{
PlotCat96New(catalog, id, type = "density", abundance = NULL)
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the 96 SNS mutation types
while its columns contain the counts of each mutation type.}

\item{id}{The ID information of the sample which has mutations.}

\item{type}{A value indicating the type of the graph. If type = "density",
the graph will plot the rates of mutations per million trinucleotides for
each mutation type. If type = "counts", the graph will plot the occurrences
of the 96 mutation types in the sample. If type = "signature", the graph
will plot mutation signatures of the sample. The default value for type is
"density".}

\item{abundance}{A matrix containing trinucleotide abundance information. To
be used only when type = "density".}
}
\value{
invisible(TRUE)
}
\description{
Plot the SNS 96 mutation catalog of one sample
}
