% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog.DNS136}
\alias{PlotCatalog.DNS136}
\title{Plot the DNS 136 spectrum of \strong{one} sample}
\usage{
\method{PlotCatalog}{DNS136}(catalog)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{CreateCatalogAttribute}} for more details.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plotthetetranucleotidesequencecontextof10majorDNSmutationtypes("AC>NN",
"AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN", "TA>NN", "TC>NN", "TG>NN",
"TT>NN") for one sample.
}
\keyword{internal}
