% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog.DNS136Catalog}
\alias{PlotCatalog.DNS136Catalog}
\title{Plot the DNS 136 spectrum of \strong{one} sample}
\usage{
\method{PlotCatalog}{DNS136Catalog}(catalog)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the tetranucleotide sequence context of 10 major DNS mutation
types("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN", "TA>NN", "TC>NN",
"TG>NN", "TT>NN") for one sample.
}
\keyword{internal}
