% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog.DNS78}
\alias{PlotCatalog.DNS78}
\title{Plot the DNS 78 spectrum or signature of \strong{one} sample}
\usage{
\method{PlotCatalog}{DNS78}(catalog)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{CreateCatalogAttribute}} for more details.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the DNS 78 spectrum or signature of \strong{one} sample
}
\keyword{internal}
