% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog.DNSClassStrandBias}
\alias{PlotCatalog.DNSClassStrandBias}
\title{Plot the strand bias graph of DNS 144 spectrum or signature of \strong{one} sample}
\usage{
\method{PlotCatalog}{DNSClassStrandBias}(catalog, strandbias = TRUE,
  cex = 1)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{strandbias}{If TRUE, plot the transcription strand bias graph of 10
major DNS mutation types ("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN",
"GC>NN", "TA>NN", "TC>NN", "TG>NN", "TT>NN") in one sample.}

\item{cex}{A numerical value giving the amount by which mutation class labels,
mutation counts(if it exists), y axis and its labels, x axis labels and
its annotations(if it exists) sample name and legend(if it exists)
should be magnified relative to the default.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the strand bias graph of DNS 144 spectrum or signature of \strong{one} sample
}
\keyword{internal}
