% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog.ID}
\alias{PlotCatalog.ID}
\title{Plot the indel(insertion and deletion) spectrum or signature of \strong{one}
sample}
\usage{
\method{PlotCatalog}{ID}(catalog)
}
\arguments{
\item{catalog}{An S3 object with class "catalog".
See \code{\link{CreateCatalogAttribute}} for more details.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the indel(insertion and deletion) spectrum or signature of \strong{one}
sample
}
\note{
The sizes of repeats involved in deletions range from 0 to 5+ in the
  catalog rownames, but for plotting and end user documentation they ranges
  from 1 to 6+.
}
\keyword{internal}
