% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog.SNS1536}
\alias{PlotCatalog.SNS1536}
\title{Plot the SNS 1536 spectrum or signature of \strong{one} sample}
\usage{
\method{PlotCatalog}{SNS1536}(catalog)
}
\arguments{
\item{catalog}{An S3 object with class "catalog".
See \code{\link{CreateCatalogAttribute}} for more details.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot a 1536 mutation catalog to a PDF file. The mutation types are in six-
letters like CATTAT, first 2-letters CA refers to (-2, -1) position, third
letter T refers to the base which has mutation, next second 2-letters TA
refers to (+1, +2) position, last letter T refers to the base after mutation.
}
\keyword{internal}
