% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog.SNS192Catalog}
\alias{PlotCatalog.SNS192Catalog}
\title{Plot the SNS 192 spectrum or signature of \strong{one} sample}
\usage{
\method{PlotCatalog}{SNS192Catalog}(catalog, cex = 0.8)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{cex}{A numerical value giving the amount by which mutation class labels,
mutation counts(if it exists), y axis and its labels, x axis labels and
its annotations(if it exists) sample name and legend(if it exists)
should be magnified relative to the default.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the SNS 192 spectrum or signature of \strong{one} sample
}
\keyword{internal}
