% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog.SNS96Catalog}
\alias{PlotCatalog.SNS96Catalog}
\title{Plot the SNS 96 spectrum or signature of \strong{one} sample}
\usage{
\method{PlotCatalog}{SNS96Catalog}(catalog, cex = 0.8, grid = TRUE,
  upper = TRUE, xlabels = TRUE)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{cex}{A numerical value giving the amount by which mutation class labels,
mutation counts(if it exists), y axis and its labels, x axis labels and
its annotations(if it exists) sample name and legend(if it exists)
should be magnified relative to the default.}

\item{grid}{If TRUE, draw grid lines in the graph.}

\item{upper}{If TRUE, draw horizontal lines and the names of major mutation
class on top of graph.}

\item{xlabels}{If TRUE, draw x axis labels.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the SNS 96 spectrum or signature of \strong{one} sample
}
\keyword{internal}
