% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog.SNSClassStrandBias}
\alias{PlotCatalog.SNSClassStrandBias}
\title{Plot the strand bias graph of SNS 192 spectrum or signature of \strong{one} sample}
\usage{
\method{PlotCatalog}{SNSClassStrandBias}(catalog, strandbias = TRUE,
  cex = 1)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{strandbias}{If TRUE, Plot the transcription strand bias graph of 6 SNS
mutation types("C>A", "C>G", "C>T", "T>A", "T>C", "T>G") in one sample.}

\item{cex}{A numerical value giving the amount by which mutation class labels,
mutation counts(if it exists), y axis and its labels, x axis labels and
its annotations(if it exists) sample name and legend(if it exists)
should be magnified relative to the default.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the strand bias graph of SNS 192 spectrum or signature of \strong{one} sample
}
\keyword{internal}
