% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf.DNS136}
\alias{PlotCatalogToPdf.DNS136}
\title{Plot a DNS 136 catalog to a PDF file}
\usage{
\method{PlotCatalogToPdf}{DNS136}(catalog, filename)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{filename}{The name of the PDF file to be produced.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the tetranucleotide sequence contexts of 10 major DNS mutation types
("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN", "TA>NN", "TC>NN",
"TG>NN", "TT>NN") of various samples to a PDF file.
}
\keyword{internal}
