% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf.DNS78}
\alias{PlotCatalogToPdf.DNS78}
\title{Plot a DNS 78 catalog to a PDF file}
\usage{
\method{PlotCatalogToPdf}{DNS78}(catalog, filename)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{filename}{The name of the PDF file to be produced.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot a DNS 78 catalog to a PDF file
}
\keyword{internal}
