% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf.DNSClassStrandBias}
\alias{PlotCatalogToPdf.DNSClassStrandBias}
\title{Plot the strand bias graph of DNS 144 catalog to a PDF file}
\usage{
\method{PlotCatalogToPdf}{DNSClassStrandBias}(catalog, filename,
  strandbias = TRUE, cex = 1)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{filename}{The name of the PDF file to be produced.}

\item{strandbias}{If TRUE, plot the transcription strand bias graph of 10
major DNS mutation types("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN",
"GC>NN", "TA>NN", "TC>NN", "TG>NN", "TT>NN") of various samples to a PDF
file.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the strand bias graph of DNS 144 catalog to a PDF file
}
\keyword{internal}
