% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf.ID}
\alias{PlotCatalogToPdf.ID}
\title{Plot an indel(insertion and deletion) catalog to a PDF file}
\usage{
\method{PlotCatalogToPdf}{ID}(catalog, filename)
}
\arguments{
\item{catalog}{An S3 object with class "catalog".
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{filename}{The name of the PDF file to be produced.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot an indel(insertion and deletion) catalog to a PDF file
}
\note{
The sizes of repeats involved in deletions range from 0 to 5+ in the
  catalog rownames, but for plotting and end user documentation they ranges
  from 1 to 6+.
}
\keyword{internal}
