% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf.SNS1536Catalog}
\alias{PlotCatalogToPdf.SNS1536Catalog}
\title{Plot an SNS 1536 catalog to a PDF file}
\usage{
\method{PlotCatalogToPdf}{SNS1536Catalog}(catalog, filename)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{filename}{The name of the PDF file to be produced.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot a 1536 mutation catalog to a PDF file. The mutation types are in six-
letters like CATTAT, first 2-letters CA refers to (-2, -1) position, third
letter T refers to the base which has mutation, next second 2-letters TA
refers to (+1, +2) position, last letter T refers to the base after mutation.
}
\keyword{internal}
