% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf.SNS192}
\alias{PlotCatalogToPdf.SNS192}
\title{Plot an SNS 192 catalog to a PDF file}
\usage{
\method{PlotCatalogToPdf}{SNS192}(catalog, filename)
}
\arguments{
\item{catalog}{An S3 object with class "catalog".
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{filename}{The name of the PDF file to be produced.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot an SNS 192 catalog to a PDF file
}
\keyword{internal}
