% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf.SNS192Catalog}
\alias{PlotCatalogToPdf.SNS192Catalog}
\title{Plot an SNS 192 catalog to a PDF file}
\usage{
\method{PlotCatalogToPdf}{SNS192Catalog}(catalog, filename)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{filename}{The name of the PDF file to be produced.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot an SNS 192 catalog to a PDF file
}
\keyword{internal}
