% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf.SNS96Catalog}
\alias{PlotCatalogToPdf.SNS96Catalog}
\title{Plot an SNS 96 catalog to a PDF file}
\usage{
\method{PlotCatalogToPdf}{SNS96Catalog}(catalog, filename, grid = TRUE,
  upper = TRUE, xlabels = TRUE)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{filename}{The name of the PDF file to be produced.}

\item{grid}{If TRUE, draw grid lines in the graph.}

\item{upper}{If TRUE, draw horizontal lines and the names of major mutation
class on top of graph.}

\item{xlabels}{If TRUE, draw x axis labels.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot an SNS 96 catalog to a PDF file
}
\keyword{internal}
