% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf.SNSClassStrandBias}
\alias{PlotCatalogToPdf.SNSClassStrandBias}
\title{Plot the strand bias graph of SNS 192 catalog to a PDF file}
\usage{
\method{PlotCatalogToPdf}{SNSClassStrandBias}(catalog, filename,
  strandbias = TRUE)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{filename}{The name of the PDF file to be produced.}

\item{strandbias}{If TRUE, Plot the transcription strand bias graph of 6 SNS
mutation types ("C>A", "C>G", "C>T", "T>A", "T>C", "T>G") of various
samples to a PDF file.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the strand bias graph of SNS 192 catalog to a PDF file
}
\keyword{internal}
