% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_INDELS_catalog.R
\name{PlotCatID}
\alias{PlotCatID}
\title{Plot the insertion and deletion catalog of one sample.}
\usage{
PlotCatID(catalog, id, type = "counts")
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the insertion and deletion
mutation types while its column contains the counts of each mutation type.}

\item{id}{The ID information of the sample which has mutations.}

\item{type}{A value indicating the type of the graph. If type = "counts", the
graph will plot the occurrences of the insertion and deletion mutation
types in the sample. If type = "signature", the graph will plot mutation
signatures of the sample. The default value for type is "counts".}
}
\value{
invisible(TRUE)
}
\description{
Please take note that the deletions Repeat Size ranges from 0 to 5+
in the catalog, but for plotting and end user documentation
it ranges from 1 to 6+.
}
