% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_related_functions.R
\name{PlotExposure}
\alias{PlotExposure}
\title{Plot exposures in multiple plots each with a manageable number of samples}
\usage{
PlotExposure(
  exposure,
  samples.per.line = 30,
  plot.proportion = FALSE,
  cex.legend = 0.9,
  ...
)
}
\arguments{
\item{exposure}{Exposures as a numerical matrix (or data.frame) with
signatures in rows and samples in columns. Rownames are taken as the
signature names and column names are taken as the sample IDs. If you want
\code{exposure} sorted from largest to smallest, use
\code{\link{SortExposure}}. Do not use column names that start with
multiple underscores. The exposures will often be mutation counts, but
could also be e.g. mutations per megabase.}

\item{samples.per.line}{Number of samples to show in each plot.}

\item{plot.proportion}{Plot exposure proportions rather than counts.}

\item{...}{Other arguments passed to \code{\link{PlotExposure}}. If
\code{ylab} is not included, it defaults to a value depending on
\code{plot.proportion}. If \code{col} is not supplied the function tries to
do something reasonable.}
}
\description{
Plot exposures in multiple plots each with a manageable number of samples
}
\examples{
file <- system.file("extdata",
                    "synthetic.exposure.csv",
                    package = "ICAMS")
exposure <- ReadExposure(file)
PlotExposure(exposure[, 1:30])
}
