% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqcontext.to.deliver.R
\name{PlotPFMmatrix}
\alias{PlotPFMmatrix}
\title{Generate dot-line plot for sequence contest of 1bp indel.}
\usage{
PlotPFMmatrix(PFMmatrix, title)
}
\arguments{
\item{PFMmatrix}{An object return from \code{\link{GeneratePlotPFMmatrix}}.}

\item{title}{A string provides the title of the plot}
}
\value{
An \strong{invisible} list.
}
\description{
Generate dot-line plot for sequence contest of 1bp indel.
}
