% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_context_functions.R
\name{PlotPPMToPdf}
\alias{PlotPPMToPdf}
\title{Plot position probability matrices (PPM) to a PDF file}
\usage{
PlotPPMToPdf(list.of.ppm, file, titles = names(list.of.ppm))
}
\arguments{
\item{file}{The name of the PDF file to be produced.}

\item{titles}{A vector of titles on top of each PPM plot.}

\item{list.of.pwm}{A list of position probability matrices (PPM)}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot position probability matrices (PPM) to a PDF file
}
\keyword{internal}
