% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_context_functions.R
\name{PlotPWM}
\alias{PlotPWM}
\title{Plot position weight matrix (PWM) for *one* sample from a Variant Call Format
(VCF) file.}
\usage{
PlotPWM(pwm, title)
}
\arguments{
\item{pwm}{A position weight matrix (PWM) for *one* sample with each row
representing the position weight of four bases.}

\item{title}{The main title of the plot.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot position weight matrix (PWM) for *one* sample from a Variant Call Format
(VCF) file.
}
\keyword{internal}
