% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_context_functions.R
\name{PlotPWMToPdf}
\alias{PlotPWMToPdf}
\title{Plot position weight matrices (PWM) to a PDF file}
\usage{
PlotPWMToPdf(list.of.pwm, file, titles = names(list.of.pwm))
}
\arguments{
\item{list.of.pwm}{A list of position weight matrices (PWM)}

\item{file}{The name of the PDF file to be produced.}

\item{titles}{A vector of titles on top of each PWM plot.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot position weight matrices (PWM) to a PDF file
}
\keyword{internal}
