% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2TSS.R
\name{PlotTransBiasDist2TSS}
\alias{PlotTransBiasDist2TSS}
\title{Plot transcriptional strand bias with respect to distance to transcription
start site}
\usage{
PlotTransBiasDist2TSS(annotated.SBS.vcf, plot.type)
}
\arguments{
\item{annotated.SBS.vcf}{An annotated SBS VCF as a data.table which contains
sequence context and transcript information. Please refer to
\code{\link{AnnotateSBSVCF}} for more details.}

\item{plot.type}{A character string indicating one mutation type to be
plotted. It should be one of "C>A", "C>G", "C>T", "T>A", "T>C", "T>G".}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot transcriptional strand bias with respect to distance to transcription
start site
}
\examples{
file <- c(system.file("extdata",
                      "Strelka.SBS.GRCh37.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadAndSplitStrelkaSBSVCFs(file)
SBS.vcf <- list.of.vcfs$SBS.vcfs[[1]]             
if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
  annotated.SBS.vcf <- AnnotateSBSVCF(SBS.vcf, ref.genome = "hg19",
                                      trans.ranges = trans.ranges.GRCh37)
  PlotTransBiasDist2TSS(annotated.SBS.vcf, plot.type = c("C>A", "C>T"))
}
}
