% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2TSS.R
\name{PlotTransBiasDist2TSSToPDF}
\alias{PlotTransBiasDist2TSSToPDF}
\title{Plot transcriptional strand bias with respect to distance to transcription
start site to a PDF file}
\usage{
PlotTransBiasDist2TSSToPDF(annotated.SBS.vcf, plot.type, file)
}
\arguments{
\item{annotated.SBS.vcf}{An annotated SBS VCF as a data.table which contains
sequence context and transcript information. Please refer to
\code{\link{AnnotateSBSVCF}} for more details.}

\item{plot.type}{A vector of character indicating types to be plotted. It
should be within "C>A", "C>G", "C>T", "T>A", "T>C", "T>G".}

\item{file}{The name of output file}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot transcriptional strand bias with respect to distance to transcription
start site to a PDF file
}
\examples{
file <- c(system.file("extdata",
                      "Strelka.SBS.GRCh37.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadAndSplitStrelkaSBSVCFs(file)
SBS.vcf <- list.of.vcfs$SBS.vcfs[[1]]             
if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
  annotated.SBS.vcf <- AnnotateSBSVCF(SBS.vcf, ref.genome = "hg19",
                                      trans.ranges = trans.ranges.GRCh37)
  PlotTransBiasDist2TSSToPDF(annotated.SBS.vcf = annotated.SBS.vcf, 
                             plot.type = c("C>A","C>G","C>T","T>A","T>C"),
                             file = file.path(tempdir(), "test.pdf"))
}
}
