% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandbias_functions.R
\name{PlotTransBiasDistToTSS}
\alias{PlotTransBiasDistToTSS}
\title{Plot transcription strand bias with respect to distance to transcription
start site.}
\usage{
PlotTransBiasDistToTSS(annotated.SBS.vcf, plot.type, ymax = NULL)
}
\arguments{
\item{annotated.SBS.vcf}{An SBS VCF annotated by \code{\link{AnnotateSBSVCF}}.}

\item{plot.type}{A character string indicating one mutation type to be
plotted. It should be one of "C>A", "C>G", "C>T", "T>A", "T>C", "T>G".}

\item{ymax}{Limit for the y axis. If not specified, it defaults to NULL and
the y axis limit equals 1.2 times of the maximum mutation counts in a
specific mutation type.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot transcription strand bias with respect to distance to transcription
start site.
}
\examples{
file <- c(system.file("extdata",
                      "Strelka.SBS.GRCh37.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadAndSplitStrelkaSBSVCFs(file)
SBS.vcf <- list.of.vcfs$SBS.vcfs[[1]]             
if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
  annotated.SBS.vcf <- AnnotateSBSVCF(SBS.vcf, ref.genome = "hg19",
                                      trans.ranges = trans.ranges.GRCh37)
  PlotTransBiasDistToTSS(annotated.SBS.vcf, plot.type = "C>T")
}
}
\references{
Hu, J., Adar, S., Selby, C. P., Lieb, J. D. & Sancar, A.
  Genome-wide analysis of human global and transcription-coupled excision
  repair of UV damage at single-nucleotide resolution. \emph{Genes Dev}. 29,
  948–960 (2015), https://doi.org/10.1101/gad.261271.115
}
