% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandbias_functions.R
\name{PlotTransBiasDistToTSSToPdf}
\alias{PlotTransBiasDistToTSSToPdf}
\title{Plot transcription strand bias with respect to distance to transcription
start site to a PDF file.}
\usage{
PlotTransBiasDistToTSSToPdf(annotated.SBS.vcf, file, plot.type = c("C>A",
  "C>G", "C>T", "T>A", "T>C", "T>G"))
}
\arguments{
\item{annotated.SBS.vcf}{An SBS VCF annotated by \code{\link{AnnotateSBSVCF}}.}

\item{file}{The name of output file.}

\item{plot.type}{A vector of character indicating types to be plotted. It
can be one or more types from "C>A", "C>G", "C>T", "T>A", "T>C", "T>G".
The default is to print all the six mutation types.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot transcription strand bias with respect to distance to transcription
start site to a PDF file.
}
\examples{
file <- c(system.file("extdata",
                      "Strelka.SBS.GRCh37.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadAndSplitStrelkaSBSVCFs(file)
SBS.vcf <- list.of.vcfs$SBS.vcfs[[1]]             
if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
  annotated.SBS.vcf <- AnnotateSBSVCF(SBS.vcf, ref.genome = "hg19",
                                      trans.ranges = trans.ranges.GRCh37)
  PlotTransBiasDistToTSSToPdf(annotated.SBS.vcf = annotated.SBS.vcf, 
                             file = file.path(tempdir(), "test.pdf"))
}
}
\references{
Hu, J., Adar, S., Selby, C. P., Lieb, J. D. & Sancar, A.
  Genome-wide analysis of human global and transcription-coupled excision
  repair of UV damage at single-nucleotide resolution. \emph{Genes Dev}. 29,
  948–960 (2015), https://doi.org/10.1101/gad.261271.115
}
