% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{PreserveCatalogAttribute}
\alias{PreserveCatalogAttribute}
\title{Preserve attributes of the input catalog}
\usage{
PreserveCatalogAttribute(pre.catalog, new.catalog)
}
\arguments{
\item{pre.catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{new.catalog}{A new catalog which needs to inherit the necessary attributes from
\code{pre.catalog}.}
}
\value{
The new catalog that has inherited the necessary attributes from
\code{pre.catalog}
}
\description{
Preserve attributes of the input catalog
}
