% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ReadAbundance3Bp}
\alias{ReadAbundance3Bp}
\title{Read data from a nucleotide abundance file with 3 base pairs}
\usage{
ReadAbundance3Bp(path)
}
\arguments{
\item{path}{Path to the file with the nucleotide abundance information with 3
base pairs.}
}
\value{
A matrix whose row names indicate 32 different types of 3 base pairs
  combinations while its column contains the occurrences of each type.
}
\description{
Read data from a nucleotide abundance file with 3 base pairs
}
