% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ReadAbundance4Bp}
\alias{ReadAbundance4Bp}
\title{Read data from a nucleotide abundance file with 4 base pairs}
\usage{
ReadAbundance4Bp(path)
}
\arguments{
\item{path}{Path to the file with the nucleotide abundance information with 4
base pairs.}
}
\value{
A matrix whose row names indicate 10 different types of 2 base pairs
  combinations while its column contains the occurrences of each type.
}
\description{
Read data from a nucleotide abundance file with 4 base pairs
}
\keyword{internal}
