% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{ReadAndSplitStrelkaSBSVCFs}
\alias{ReadAndSplitStrelkaSBSVCFs}
\title{Read and split Strelka SBS VCF files}
\usage{
ReadAndSplitStrelkaSBSVCFs(files, names.of.VCFs = NULL)
}
\arguments{
\item{files}{Character vector of file paths to the Strelka SBS VCF files.}

\item{names.of.VCFs}{Optional. Character vector of names of the VCF files.
The order of names in \code{names.of.VCFs} should match the order of VCF
file paths in \code{files}. If \code{NULL}(default), this function will
remove all of the path up to and including the last path separator (if any)
in \code{files} and file paths without extensions (and the leading dot)
will be used as the names of the VCF files.}
}
\description{
The function will find and merge adjacent SBS pairs into DBS if their VAFs
are very similar. The default threshold value for VAF is 0.02.
}
\section{Value}{

A list of elements as follows:
\itemize{
\item \code{SBS.vcfs}: List of data.frames of pure SBS mutations -- no DBS or
3+BS mutations.
\item \code{DBS.vcfs}: List of data.frames of pure DBS mutations -- no SBS or
3+BS mutations.
\item \code{ThreePlus}: List of data tables with the key CHROM, LOW.POS,
HIGH.POS and additional information (reference sequence, alternative
sequence, context, etc.) which contain rows in the input that did not
represent SBSs or DBSs. Additional information not fully implemented at
this point because of limited immediate biological interest.
\item \code{multiple.alt}: Rows with multiple alternate alleles (removed from
\code{SBS.vcfs} etc.)
}
}

\examples{
file <- c(system.file("extdata/Strelka-SBS-vcf",
                      "Strelka.SBS.GRCh37.s1.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadAndSplitStrelkaSBSVCFs(file)
}
\seealso{
\code{\link{StrelkaSBSVCFFilesToCatalog}}
}
