% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadAndSplitStrelkaSNSVCFs}
\alias{ReadAndSplitStrelkaSNSVCFs}
\title{Read and split Strelka SNS VCF files.}
\usage{
ReadAndSplitStrelkaSNSVCFs(vector.of.file.paths)
}
\arguments{
\item{vector.of.file.paths}{Character vector of
file paths to the VCF files.}
}
\value{
A list of 3 in-memory objects with the elements:
   SNS.vcfs:  List of Data frames of pure SNS mutations -- no DNS or 3+BS mutations
   DNS.vcfs:  List of Data frames of pure DNS mutations -- no SNS or 3+BS mutations
   ThreePlus: List of Data tables with the key CHROM, LOW.POS, HIGH.POS and additional
   information (reference sequence, alternative sequence, context, etc.)
   Additional information not fully implemented at this point because of
   limited immediate biological interest.
}
\description{
Read and split Strelka SNS VCF files.
}
\seealso{
\code{\link{StrelkaSNSVCFFilesToCatalog}}
}
