% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{ReadCat96}
\alias{ReadCat96}
\title{Read 96-channel spectrum or signatures in PCAWG7 format}
\usage{
ReadCat96(path, strict = TRUE)
}
\arguments{
\item{path}{Path to a catalog on disk in the "PCAWG7" format.}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}
}
\value{
A catalog in canonical in-memory format.
}
\description{
Read 96-channel spectrum or signatures in PCAWG7 format
}
