% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPOSITE.R, R/read_write_catalog.R
\name{ReadCatalog.COMPOSITECatalog}
\alias{ReadCatalog.COMPOSITECatalog}
\alias{ReadCatalog.Methods}
\alias{ReadCatalog.SBS96Catalog}
\alias{ReadCatalog.SBS192Catalog}
\alias{ReadCatalog.SBS1536Catalog}
\alias{ReadCatalog.DBS78Catalog}
\alias{ReadCatalog.DBS144Catalog}
\alias{ReadCatalog.DBS136Catalog}
\alias{ReadCatalog.IndelCatalog}
\title{ReadCatalog.Methods}
\usage{
ReadCatalog.COMPOSITECatalog(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  strict = TRUE,
  stop.on.error = TRUE
)

ReadCatalog.SBS96Catalog(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  strict = TRUE,
  stop.on.error = TRUE
)

ReadCatalog.SBS192Catalog(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  strict = TRUE,
  stop.on.error = TRUE
)

ReadCatalog.SBS1536Catalog(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  strict = TRUE,
  stop.on.error = TRUE
)

ReadCatalog.DBS78Catalog(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  strict = TRUE,
  stop.on.error = TRUE
)

ReadCatalog.DBS144Catalog(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  strict = TRUE,
  stop.on.error = TRUE
)

ReadCatalog.DBS136Catalog(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  strict = TRUE,
  stop.on.error = TRUE
)

ReadCatalog.IndelCatalog(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  strict = TRUE,
  stop.on.error = TRUE
)
}
\arguments{
\item{file}{Path to a catalog on disk in the standardized format.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{region A character string designating a genomic region;
see \code{\link{as.catalog}} and \code{\link{ICAMS}}.}

\item{catalog.type}{One of "counts", "density", "counts.signature",
"density.signature".}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}

\item{stop.on.error}{If TRUE, call \code{stop} on error; otherwise
return a 1-column matrix of NA's with the attribute "error"
containing error information. The number of rows may not
be the correct number for the expected catalog type.}
}
\description{
Read catalog methods
}
\details{
See the generic function \code{\link{ReadCatalog}}
}
