% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{ReadCatalogErrReturn}
\alias{ReadCatalogErrReturn}
\title{Get error message and either stop or create a null error output for read catalog}
\usage{
ReadCatalogErrReturn(err.info, nrow, stop.on.error, do.message = TRUE)
}
\arguments{
\item{err.info}{The information passed to the \code{tryCatch} \code{error}
function argument.}

\item{nrow}{The number of rows to put in the 1-column NA return matrix.}

\item{stop.on.error}{If \code{TRUE} then call \code{stop()}.}

\item{do.message}{If \code{TRUE} then \code{message} the error information.}
}
\description{
Get error message and either stop or create a null error output for read catalog
}
\keyword{internal}
