% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{ReadCatDNS144}
\alias{ReadCatDNS144}
\title{Read a 144 DNS catalog from path}
\usage{
ReadCatDNS144(path, strict = TRUE)
}
\arguments{
\item{path}{Path to a catalog on disk in the "PCAWG7" format.}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}
}
\value{
A catalog in canonical in-memory format.
}
\description{
Read a 144 DNS catalog from path
}
