% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{ReadCatID}
\alias{ReadCatID}
\title{Read a ID (insertion/deletion) catalog from path}
\usage{
ReadCatID(path, strict = TRUE)
}
\arguments{
\item{path}{Path to a catalog on disk in the "PCAWG7" format.}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}
}
\value{
A catalog in canonical in-memory format.
}
\description{
Please take note that the deletions Repeat Size ranges from 0 to 5+
in the catalog, but for plotting and end user documentation
it ranges from 1 to 6+.
}
