% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{ReadDukeNUSCat192}
\alias{ReadDukeNUSCat192}
\title{Read a 192-channel spectra (or signaure) catalog in Duke-NUS format.}
\usage{
ReadDukeNUSCat192(file, ref.genome = NULL, region = "unknown",
  catalog.type = "counts", abundance = NULL)
}
\description{
The file needs to have the column names Before	Ref	After	Var
in the first 4 columns
}
\keyword{internal}
