% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{ReadIDSimpleAsVCF}
\alias{ReadIDSimpleAsVCF}
\title{ReadIDSimpleAsVCF}
\usage{
ReadIDSimpleAsVCF(path, verbose = 0)
}
\arguments{
\item{path}{TODO}

\item{verbose}{TODO}
}
\value{
TODO
}
\description{
ReadIDSimpleAsVCF
}
