% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadListOfMutectVCFs}
\alias{ReadListOfMutectVCFs}
\title{Read a list of Mutect VCF files from path}
\usage{
ReadListOfMutectVCFs(vector.of.file.paths)
}
\arguments{
\item{vector.of.file.paths}{A vector containing the paths of the VCF files.}
}
\value{
A list of vcfs from vector.of.file.paths.
}
\description{
Read a list of Mutect VCF files from path
}
