% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{ReadSimpleAsVCF}
\alias{ReadSimpleAsVCF}
\title{ReadSimpleAsVCF}
\usage{
ReadSimpleAsVCF(path, verbose = 0)
}
\arguments{
\item{path}{TODO}

\item{verbose}{TODO}
}
\value{
TODO
}
\description{
ReadSimpleAsVCF
}
