% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{ReadStapleGT96SBS}
\alias{ReadStapleGT96SBS}
\title{Read a 96-channel spectra (or signature) catalog where rownames are e.g. "A[C>A]T".}
\usage{
ReadStapleGT96SBS(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  abundance = NULL,
  sep = "\\t"
)
}
\description{
The file needs to have the rownames in the first column.
}
\keyword{internal}
