% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadStrelkaSNSVCF}
\alias{ReadStrelkaSNSVCF}
\title{Read in the data lines of an SNS VCF created by Strelka version 1}
\usage{
ReadStrelkaSNSVCF(path)
}
\arguments{
\item{path}{The name/path of the VCF file, or a complete URL.}
}
\value{
A data frame storing mutation records of a VCF file.
}
\description{
Read in the data lines of an SNS VCF created by Strelka version 1
}
\keyword{internal}
