% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadStrelkaSNSVCFs}
\alias{ReadStrelkaSNSVCFs}
\title{Read Strelka SNS (single nucleotide substitutions) VCF files.}
\usage{
ReadStrelkaSNSVCFs(vector.of.file.paths)
}
\arguments{
\item{vector.of.file.paths}{Character vector of
file paths to VCF files.}
}
\value{
A list of vcfs from vector.of.file.paths.
}
\description{
Read Strelka SNS (single nucleotide substitutions) VCF files.
}
\keyword{internal}
