% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadStrelkaVCF}
\alias{ReadStrelkaVCF}
\title{Read in the data lines of a Variant Call Format (VCF) file}
\usage{
ReadStrelkaVCF(path)
}
\arguments{
\item{path}{The name/path of the VCF file, or a complete URL.}
}
\value{
A data frame storing mutation records of a VCF file.
}
\description{
Read in the data lines of a Variant Call Format (VCF) file
}
\keyword{internal}
