% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ReadTranscriptRanges}
\alias{ReadTranscriptRanges}
\title{Read transcript ranges and strands from a gff3 format file.
Use this one for the new, cut down gff3 file (2018 11 24)}
\usage{
ReadTranscriptRanges(path)
}
\arguments{
\item{path}{Path to the file with the transcript information with 1-based
start end positions of genomic ranges.}
}
\value{
A data.table keyed by chrom, chromStart, and chromEnd.
}
\description{
Read transcript ranges and strands from a gff3 format file.
Use this one for the new, cut down gff3 file (2018 11 24)
}
