% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{RemoveTransRangesOnBothStrand}
\alias{RemoveTransRangesOnBothStrand}
\title{Remove ranges that have transcripts on both strands from a trans.ranges file}
\usage{
RemoveTransRangesOnBothStrand(trans.ranges)
}
\arguments{
\item{trans.ranges}{A GFF3 trans.range.file}
}
\value{
A data table which has removed ranges that have transcripts on both
strands from the input \code{trans.ranges}.
}
\description{
Remove ranges that have transcripts on both strands from a trans.ranges file
}
\keyword{internal}
