% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF_etc.R
\name{RenameColumnsWithNameEnd}
\alias{RenameColumnsWithNameEnd}
\title{Is there any column in \code{df} with name "end"?
If there is, change its name to "end_old" so that it will
conflict with code in other parts of ICAMS package.}
\usage{
RenameColumnsWithNameEnd(df)
}
\description{
Is there any column in \code{df} with name "end"?
If there is, change its name to "end_old" so that it will
conflict with code in other parts of ICAMS package.
}
\keyword{internal}
