% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{Restaple1536}
\alias{Restaple1536}
\title{Convert 1536-channel mutation-type identifiers like this \code{"ACCGTA" -> "AC[C>A]GT"}}
\usage{
Restaple1536(c1)
}
\arguments{
\item{c1}{A vector of character strings with the first 5 characters
being the source trinucleotide and the last character being the
mutated (center) nucleotide. E.g. \code{ACCGTA} means a mutation from
\code{ACCGT > ACAGT}.}
}
\description{
This is an internal function needed for generating
"non-canonical" row number formats for catalogs.
}
\keyword{internal}
