% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{Restaple96}
\alias{Restaple96}
\title{Convert 96-channel mutation-type identifiers like this \code{"ACTA" -> "A[C>A]T"}}
\usage{
Restaple96(c1)
}
\arguments{
\item{c1}{A vector of character strings with the first 3 characters
being the source trinucleotide and the last character being the
mutated (center) nucleotide. E.g. \code{ACTA} means a mutation from
\code{ACT > AAT}.}
}
\description{
This is an internal function needed for generating
"non-canonical" row number formats for catalogs.
}
\keyword{internal}
