% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{RevcSBS96}
\alias{RevcSBS96}
\title{Reverse complement strings that represent stranded SBSs}
\usage{
RevcSBS96(mutstring)
}
\arguments{
\item{mutstring}{A vector of 4-character strings representing
stranded SBSs in trinucleotide context,
for example "AATC" represents AAT > ACT mutations.}
}
\value{
Return the vector of
reverse complements of the first 3 characters
concatenated with the reverse complement of the
last character, e.g. "AATC" returns "ATTG".
}
\description{
Reverse complement strings that represent stranded SBSs
}
\keyword{internal}
