% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.gene.expression.value.GRCh37}
\alias{sample.gene.expression.value.GRCh37}
\title{Sample gene expression values for human GRCh37.}
\format{A data frame with two columns. The first column contains the gene
  symbol, the second columns contains the gene expression values
  TPM (Transcripts Per Kilobase Million).}
\usage{
sample.gene.expression.value.GRCh37
}
\description{
This data is designed to be used as an example in function
\code{\link{PlotTransBiasExp}} and \code{\link{PlotTransBiasExpToPdf}}.
}
\keyword{datasets}
