% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{SampleID2AliquotID}
\alias{SampleID2AliquotID}
\title{SampleID2AliquotID}
\usage{
SampleID2AliquotID(sample.id)
}
\arguments{
\item{sample.id}{TODO}
}
\value{
TODO
}
\description{
SampleID2AliquotID
}
