% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_indel.R
\name{show_indel}
\alias{show_indel}
\title{Print an indel with a dotted deletion marker.}
\usage{
show_indel(long_str, short_str, pos, ins_or_del)
}
\arguments{
\item{long_str}{Character string containing the pre-edit sequence.}

\item{short_str}{Character string containing the post-edit sequence.}

\item{pos}{One-based position of the edit within `long_str`.}

\item{ins_or_del}{Single character, `"i"` for insertion or `"d"` for deletion.}
}
\description{
Print an indel with a dotted deletion marker.
}
\keyword{internal}
