% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNS_related_functions.R
\name{SNSVCFsToCatalogs}
\alias{SNSVCFsToCatalogs}
\title{Create a list of 3 catalogs (one each for 96, 192, 1536)
out of the contents of the VCFs in list.of.vcfs}
\usage{
SNSVCFsToCatalogs(list.of.vcfs, genome, trans.ranges)
}
\arguments{
\item{list.of.vcfs}{List vector of in-memory VCFs. The list names will be the
sample ids in the output catalog.}

\item{genome}{Name of a particular reference genome (without quotations
marks).}

\item{trans.ranges}{A data frame containing transcript ranges.}
}
\value{
A list of 3 catalogs, one each for 96, 192, 1536:
  cat96
  cat192
  cat1536
}
\description{
Create a list of 3 catalogs (one each for 96, 192, 1536)
out of the contents of the VCFs in list.of.vcfs
}
