% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{SplitListOfStrelkaVCFs}
\alias{SplitListOfStrelkaVCFs}
\title{Split a list of in-memory Strelka VCF into SNS, DNS, and variants involving
> 2 consecutive bases}
\usage{
SplitListOfStrelkaVCFs(list.of.vcfs)
}
\arguments{
\item{list.of.vcfs}{A list of in-memory data frame containing Strelka VCF file contents.}
}
\value{
A list of 3 in-memory objects with the elements:
}
\description{
SNSs are single nucleotide substitutions,
eg C>T, A<G,....  DNSs are double nucleotide substitutions,
eg CC>TT, AT>GG, ...  Variants involving > 2 consecutive
bases are rare, so this function just records them. These
would be variants such ATG>CCT, AGAT > TCTA, ...
}
