% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{SplitMutectVCFs}
\alias{SplitMutectVCFs}
\title{Split each Mutect VCF into SBS, DBS, and ID VCFs (plus two left-over data.frames)}
\usage{
SplitMutectVCFs(list.of.vcfs)
}
\arguments{
\item{list.of.vcfs}{List of VCFs as in-memory data.frames}
}
\value{
A list with 5 elements, as folows:
}
\description{
Split each Mutect VCF into SBS, DBS, and ID VCFs (plus two left-over data.frames)
}
