% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{SplitStrelkaSNSVCF}
\alias{SplitStrelkaSNSVCF}
\title{Split an in-memory Strelka VCF into SNS, DNS, and variants involving
> 2 consecutive bases}
\usage{
SplitStrelkaSNSVCF(vcf.df, max.vaf.diff = 0.02)
}
\arguments{
\item{vcf.df}{An in-memory data frame containing a Strelka VCF file contents.}

\item{max.vaf.diff}{The maximum difference of VAF, default value is 0.02.}
}
\value{
A list of 3 in-memory objects with the elements:
}
\description{
SNSs are single nucleotide substitutions,
eg C>T, A<G,....  DNSs are double nucleotide substitutions,
eg CC>TT, AT>GG, ...  Variants involving > 2 consecutive
bases are rare, so this function just records them. These
would be variants such ATG>CCT, AGAT > TCTA, ...
}
\keyword{internal}
