% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_name_functions.R
\name{StandardChromName}
\alias{StandardChromName}
\title{Standardize the chromosome name annotations for a data frame.}
\usage{
StandardChromName(df)
}
\arguments{
\item{df}{A data frame whose first column contains the Chromosome name}
}
\value{
A data frame whose Chromosome names are only in the form of 1:22, "X"
and "Y".
}
\description{
Standardize the chromosome name annotations for a data frame.
}
\keyword{internal}
