% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{StopIfCatalogTypeIllegal}
\alias{StopIfCatalogTypeIllegal}
\title{Stop if \code{catalog.type} is illegal.}
\usage{
StopIfCatalogTypeIllegal(catalog.type)
}
\arguments{
\item{catalog.type}{Character string to check.}
}
\description{
Stop if \code{catalog.type} is illegal.
}
\keyword{internal}
