% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{StopIfNrowIllegal}
\alias{StopIfNrowIllegal}
\title{Stop if the number of rows in \code{object} is illegal}
\usage{
StopIfNrowIllegal(object)
}
\arguments{
\item{object}{A \code{catalog}, numeric \code{matrix}, or numeric \code{data.fram}}
}
\description{
Stop if the number of rows in \code{object} is illegal
}
\keyword{internal}
