% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{StopIfRegionIllegal}
\alias{StopIfRegionIllegal}
\title{Stop if \code{region} is illegal.}
\usage{
StopIfRegionIllegal(region)
}
\arguments{
\item{region}{Character string to check.}
}
\description{
Stop if \code{region} is illegal.
}
\keyword{internal}
