% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{StopIfTranscribedRegionIllegal}
\alias{StopIfTranscribedRegionIllegal}
\title{Stop if \code{region} is illegal for an in-transcript catalogs}
\usage{
StopIfTranscribedRegionIllegal(region)
}
\arguments{
\item{region}{The region to test (a character string)}
}
\description{
Stop if \code{region} is illegal for an in-transcript catalogs
}
\keyword{internal}
