% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{StrelkaSNSVCFFilesToCatalog}
\alias{StrelkaSNSVCFFilesToCatalog}
\title{Create SNS and DNS catalogs from Strelka SNS VCF files}
\usage{
StrelkaSNSVCFFilesToCatalog(vector.of.file.paths, genome, trans.ranges)
}
\arguments{
\item{vector.of.file.paths}{A vector containing the paths of the Strelka SNS VCF files.}

\item{genome}{A particular reference genome(without quotation marks). Use
\link[BSgenome]{available.genomes} to get the list of "BSgenome data
packages" currently available. There are 2 types of predefined reference
genome which are incorporated in this function. User can invoke a
predefined human GRCh38/hg38 BSgenome data package by typing \code{genome =
"GRCh38"} or \code{genome = "hg38"}. User can invoke a predefined human
GRCh37/hg19 BSgenome data package by typing \code{genome = "GRCh37"} or
\code{genome = "hg19"}.}

\item{trans.ranges}{A data.table which contains transcript range and
strand information.}
}
\value{
A list of 3 SNS catalogs (one each for 96, 192, and 1536)
  and 3 DNS catalogs (one each for 78, 136, and 144)
}
\description{
Create 3 SNS catalogs (96, 192, 1536) and 3 DNS catalogs (78, 136, 144)
from the Strelka SNS VCFs specified by vector.of.file.paths
}
\details{
This function calls \code{\link{VCFsToSNSCatalogs}} and
\code{\link{VCFsToDNSCatalogs}}
}
\note{
SNS 192 and DNS 144 catalog only contains mutations in transcribed regions.
}
