% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{StrelkaVCFsToSNSCatalogs}
\alias{StrelkaVCFsToSNSCatalogs}
\title{Create SNS catalogs from Strelka VCFs}
\usage{
StrelkaVCFsToSNSCatalogs(list.of.SNS.vcfs, genome, trans.ranges)
}
\arguments{
\item{list.of.SNS.vcfs}{List of in-memory data frames of pure SNS mutations
-- no DNS or 3+BS mutations. The list names will be the sample ids in the
output catalog.}

\item{genome}{Name of a particular reference genome (without quotations
marks).}

\item{trans.ranges}{A data frame containing transcript ranges.}
}
\value{
A list of 3 catalogs, one each for 96, 192, 1536:
  cat96
  cat192
  cat1536
}
\description{
Create a list of 3 catalogs (one each for 96, 192, 1536)
out of the contents in list.of.SNS.vcfs
}
\keyword{internal}
