% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TCFromCouSigCou}
\alias{TCFromCouSigCou}
\title{Source catalog type is counts or counts.signature}
\usage{
TCFromCouSigCou(s, t)
}
\description{
counts.signature -> density.signature, counts.signature
counts -> anything
}
\keyword{internal}
