% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{TestDNSCatalog}
\alias{TestDNSCatalog}
\title{This function is to test whether the predefined functions
are working correctly to produce the desired DNS catalogs}
\usage{
TestDNSCatalog(vcf.df)
}
\arguments{
\item{vcf.df}{An in-memory data frame containing a VCF file contents.}
}
\description{
This function is to test whether the predefined functions
are working correctly to produce the desired DNS catalogs
}
\keyword{internal}
