% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{TestMakeCatalogFromMutectVCFs}
\alias{TestMakeCatalogFromMutectVCFs}
\title{This function is to make catalogs from the sample Mutect VCF file
to compare with the expected catalog information.}
\usage{
TestMakeCatalogFromMutectVCFs()
}
\description{
This function is to make catalogs from the sample Mutect VCF file
to compare with the expected catalog information.
}
\keyword{internal}
