% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{TestMakeCatalogFromSNSVCFs}
\alias{TestMakeCatalogFromSNSVCFs}
\title{This function is to make catalogs from the sample VCF files
to compare with the expected catalog information.}
\usage{
TestMakeCatalogFromSNSVCFs()
}
\description{
This function is to make catalogs from the sample VCF files
to compare with the expected catalog information.
}
