% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{TestMakeCatalogFromStrelkaIDVCFs}
\alias{TestMakeCatalogFromStrelkaIDVCFs}
\title{This function is to make catalogs from the sample Strelka ID VCF files
to compare with the expected catalog information.}
\usage{
TestMakeCatalogFromStrelkaIDVCFs()
}
\description{
This function is to make catalogs from the sample Strelka ID VCF files
to compare with the expected catalog information.
}
