% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{TestMakeCatalogFromStrelkaSBSVCFs}
\alias{TestMakeCatalogFromStrelkaSBSVCFs}
\title{This function is to make catalogs from the sample Strelka SBS VCF files
to compare with the expected catalog information.}
\usage{
TestMakeCatalogFromStrelkaSBSVCFs()
}
\description{
This function is to make catalogs from the sample Strelka SBS VCF files
to compare with the expected catalog information.
}
\keyword{internal}
