% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{TestMakeCatalogFromStrelkaSNSVCFs}
\alias{TestMakeCatalogFromStrelkaSNSVCFs}
\title{This function is to make catalogs from the sample Strelka SNS VCF files
to compare with the expected catalog information.}
\usage{
TestMakeCatalogFromStrelkaSNSVCFs()
}
\description{
This function is to make catalogs from the sample Strelka SNS VCF files
to compare with the expected catalog information.
}
\keyword{internal}
