% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{TestMutectVCFToCatalog}
\alias{TestMutectVCFToCatalog}
\title{test \code{SplitListOfMutectVCFs} and functions to create catalogs.}
\usage{
TestMutectVCFToCatalog()
}
\description{
test \code{SplitListOfMutectVCFs} and functions to create catalogs.
}
\details{
Stop if the catalogs created do not match the expected values.
}
