% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPOSITE.R
\name{TestPlotCatCOMPOSITE}
\alias{TestPlotCatCOMPOSITE}
\title{Plot the a SignatureAnalyzer COMPOSITE signature or catalog into separate pdfs}
\usage{
TestPlotCatCOMPOSITE(catalog, filename.header, type,
  id = colnames(catalog))
}
\arguments{
\item{catalog}{Catalog or signature matrix}

\item{filename.header}{Contain path and the beginning part of the file name.
The name of the pdf files will be:
\emph{filename.header}\code{.SBS.96.pdf}
\emph{filename.header}\code{.SBS.1536.pdf}
\emph{filename.header}\code{.DBS.78.pdf}
\emph{filename.header}\code{.ID.83.pdf}}

\item{type}{See \code{\link[ICAMS]{PlotCatalogToPdf}}.}

\item{id}{A vector containing the identifiers of the samples
or signatures in \code{catalog}.}
}
\description{
Plot the a SignatureAnalyzer COMPOSITE signature or catalog into separate pdfs
}
\keyword{internal}
