% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{TestSNSandDNSCat}
\alias{TestSNSandDNSCat}
\title{This function is to test whether the predefined functions
are working correctly to produce the desired SNS and DNS catalogs}
\usage{
TestSNSandDNSCat()
}
\description{
This function is to test whether the predefined functions
are working correctly to produce the desired SNS and DNS catalogs
}
\keyword{internal}
