% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{TestStrelkaDNSCatalog}
\alias{TestStrelkaDNSCatalog}
\title{This function is to test whether the predefined functions
are working correctly to produce the desired DNS catalogs.}
\usage{
TestStrelkaDNSCatalog(vcf.df)
}
\arguments{
\item{vcf.df}{This function is to test whether the predefined functions
are working correctly to produce the desired DNS catalogs.}
}
\description{
This function is to test whether the predefined functions
are working correctly to produce the desired DNS catalogs.
}
\keyword{internal}
