% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TransformSpectra}
\alias{TransformSpectra}
\alias{TransDinucSpectra}
\alias{TransTrinucSpectra}
\alias{TransTetranucSpectra}
\alias{TransPentanucSpectra}
\title{Transform nucleotide spectra functions}
\usage{
TransDinucSpectra(catalog, source.abundance, target.abundance)

TransTrinucSpectra(catalog, source.abundance, target.abundance)

TransTetranucSpectra(catalog, source.abundance, target.abundance)

TransPentanucSpectra(catalog, source.abundance, target.abundance)
}
\arguments{
\item{catalog}{A matrix of mutation counts. Rownames indicate the mutation
types. Each column contains the mutation counts for one sample.}

\item{source.abundance}{An abundance matrix specified by the user, which
can be created using functions \code{\link{CreateDinucAbundance}},
\code{\link{CreateTrinucAbundance}}, \code{\link{CreateTetranucAbundance}},
\code{\link{CreatePentanucAbundance}}.
There are 6 types of predefined abundance matrix which are incorporated
in this function ("GRCh37.genome", "GRCh37.exome", "GRCh38.genome",
"GRCh38.exome", "GRCm38.genome", "GRCm38.exome").
User can invoke a specific predefined abundance matrix by typing its name,
e.g. source.abundance = "GRCh37.genome".}

\item{target.abundance}{An abundance matrix specified by the user, which
can be created using functions \code{\link{CreateDinucAbundance}},
\code{\link{CreateTrinucAbundance}}, \code{\link{CreateTetranucAbundance}},
\code{\link{CreatePentanucAbundance}}.
There are 6 types of predefined abundance matrix which are incorporated
in this function ("GRCh37.genome", "GRCh37.exome", "GRCh38.genome",
"GRCh38.exome", "GRCm38.genome", "GRCm38.exome").
User can invoke a specific predefined abundance matrix by typing its name,
e.g. target.abundance = "GRCm38.genome".}
}
\value{
A matrix of inferred mutation counts. Rownames indicate the mutation
  types which are the same as those in \code{catalog}.
  Each column contains the inferred mutation counts for one sample based on
  \code{target.abundance}.
}
\description{
Transform count spectra from a particular organism region to an inferred
count spectra based on the target nucleotide abundance.
}
