% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{TransRownames.ID.PCAWG.SigPro}
\alias{TransRownames.ID.PCAWG.SigPro}
\title{For indels, convert ICAMS/PCAWG7 rownames into SigProfiler rownames}
\usage{
TransRownames.ID.PCAWG.SigPro(vector.of.rownames)
}
\description{
For indels, convert ICAMS/PCAWG7 rownames into SigProfiler rownames
}
\examples{
ICAMS:::TransRownames.ID.PCAWG.SigPro("DEL:C:1:0") # 1:Del:C:0;
ICAMS:::TransRownames.ID.PCAWG.SigPro("INS:repeat:2:5+") # 2:Ins:R:5

}
\keyword{internal}
