% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{Unstaple1536}
\alias{Unstaple1536}
\title{Convert SBS1536-channel mutations-type identifiers like this \code{"AC[C>A]GT" -> "ACCGTA"}}
\usage{
Unstaple1536(c1)
}
\arguments{
\item{c1}{A vector of character strings with the mutation indicated by
e.g. \code{[C>A]} in the middle.}
}
\description{
Convert SBS1536-channel mutations-type identifiers like this \code{"AC[C>A]GT" -> "ACCGTA"}
}
\keyword{internal}
