% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{Unstaple78}
\alias{Unstaple78}
\title{Convert DBS78-channel mutations-type identifiers like this \code{"AC>GA" -> "ACGA"}}
\usage{
Unstaple78(c1)
}
\arguments{
\item{c1}{A vector of character strings with a \code{>} sign separating
reference and variant context e.g. \code{AC>GA}.}
}
\description{
Convert DBS78-channel mutations-type identifiers like this \code{"AC>GA" -> "ACGA"}
}
\keyword{internal}
