% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{Unstaple96}
\alias{Unstaple96}
\title{Convert 96-channel mutations-type identifiers like this \code{"A[C>A]T" -> "ACTA"}}
\usage{
Unstaple96(c1)
}
\arguments{
\item{c1}{A vector of character strings with the mutation indicated by
e.g. \code{[C>A]} in the middle.}
}
\description{
Convert 96-channel mutations-type identifiers like this \code{"A[C>A]T" -> "ACTA"}
}
\keyword{internal}
