% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{VCFFilesToCatalog}
\alias{VCFFilesToCatalog}
\title{Create SNS and DNS catalogs from VCF files}
\usage{
VCFFilesToCatalog(vector.of.file.paths, genome, trans.ranges)
}
\arguments{
\item{vector.of.file.paths}{A vector containing the paths of the VCF files.}

\item{genome}{Name of a particular reference genome
(without quotations marks).}

\item{trans.ranges}{A data.table which contains transcript range and
strand information.}
}
\value{
A list of 3 SNS catalogs (one each for 96, 192, and 1536)
  and 3 DNS catalogs (one each for 78, 136, and 144)
}
\description{
Create 3 SNS catalogs (96, 192, 1536) and 3 DNS catalogs (78, 136, 144)
from the VCFs specified by vector.of.file.paths
}
\details{
This function calls \code{\link{VCFsToSNSCatalogs}} and \code{\link{VCFsToDNSCatalogs}}
}
