% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{VCFsToZipFileXtra}
\alias{VCFsToZipFileXtra}
\title{Analogous to \code{\link{VCFsToZipFile}}, also generates density CSV and PDF files in the zip
archive.}
\usage{
VCFsToZipFileXtra(
  dir,
  zipfile,
  ref.genome,
  variant.caller = "unknown",
  num.of.cores = 1,
  trans.ranges = NULL,
  region = "unknown",
  names.of.VCFs = NULL,
  tumor.col.names = NA,
  filter.status = NULL,
  get.vaf.function = NULL,
  ...,
  max.vaf.diff = 0.02,
  base.filename = "",
  return.annotated.vcfs = FALSE,
  suppress.discarded.variants.warnings = TRUE
)
}
\description{
Analogous to \code{\link{VCFsToZipFile}}, also generates density CSV and PDF files in the zip
archive.
}
\keyword{internal}
