% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCat}
\alias{WriteCat}
\title{Write a matrix of mutation catalog to a file on disk}
\usage{
WriteCat(ct, path, num.row, row.order, row.header, strict)
}
\arguments{
\item{ct}{A matrix of mutation catalog.}

\item{path}{The path of the file to be written on disk.}

\item{num.row}{The number of rows in the file to be written.}

\item{row.order}{The row order to be used for writing the file.}

\item{row.header}{The row header to be used for writing the file.}

\item{strict}{If TRUE, do additional checks on the input,
and stop if the checks fail.}
}
\description{
Write a matrix of mutation catalog to a file on disk
}
\keyword{internal}
