% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCat192}
\alias{WriteCat192}
\title{Write a SNS 192 mutation catalog to a file on disk}
\usage{
WriteCat192(ct, path, strict = TRUE)
}
\arguments{
\item{ct}{A matrix of SNS 192 mutation catalog.}

\item{path}{The path of the file to be written on disk.}

\item{strict}{If TRUE, do additional checks on the input,
and stop if the checks fail.}
}
\description{
Write a SNS 192 mutation catalog to a file on disk
}
