% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{WriteCatalogIndelSigPro}
\alias{WriteCatalogIndelSigPro}
\title{Write Indel Catalogs in SigProExtractor format}
\usage{
WriteCatalogIndelSigPro(catalog, file, strict = TRUE, sep = "\\t")
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}};
see also \code{\link{as.catalog}}.}

\item{file}{The path to the file to be created.}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}

\item{sep}{Separator to use in the output file. In older version
SigProfiler read comma-separated files; as of May 2020 it
reads tab-separated files.}
}
\description{
Write Indel Catalogs in SigProExtractor format to a file.
}
\note{
In ID (small insertions and deletions) catalogs in SigProExtractor format,
deletion repeat sizes range from 0 to 5, rather than 0 to 5+.
}
\keyword{internal}
