% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCatDNS144}
\alias{WriteCatDNS144}
\title{Write a DNS 144 mutation catalog to a file on disk}
\usage{
WriteCatDNS144(ct, path, strict = TRUE)
}
\arguments{
\item{ct}{A matrix of DNS 144 mutation catalog.}

\item{path}{The path of the file to be written on disk.}

\item{strict}{If TRUE, do additional checks on the input,
and stop if the checks fail.}
}
\description{
Write a DNS 144 mutation catalog to a file on disk
}
