% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCatID}
\alias{WriteCatID}
\title{Write a ID (insertion/deletion) catalog to a file on disk}
\usage{
WriteCatID(ct, path, strict = TRUE)
}
\arguments{
\item{ct}{A matrix of ID (insertion/deletion) catalog.}

\item{path}{The path of the file to be written on disk.}

\item{strict}{If TRUE, do additional checks on the input,
and stop if the checks fail.}
}
\description{
Please take note that the deletions Repeat Size ranges from 0 to 5+
in the catalog, but for plotting and end user documentation
it ranges from 1 to 6+.
}
