% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCatQUAD136}
\alias{WriteCatQUAD136}
\title{Write a QUAD 136 catalog to a file on disk}
\usage{
WriteCatQUAD136(ct, path, strict = TRUE)
}
\arguments{
\item{ct}{A matrix of QUAD 136 catalog.}

\item{path}{The path of the file to be written on disk.}

\item{strict}{If TRUE, do additional checks on the input,
and stop if the checks fail.}
}
\description{
Write a QUAD 136 catalog to a file on disk
}
