% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_related_functions.R
\name{WriteExposure}
\alias{WriteExposure}
\title{Write an exposure matrix to a file}
\usage{
WriteExposure(exposure, file)
}
\arguments{
\item{exposure}{Exposures as a numerical matrix (or data.frame) with
signatures in rows and samples in columns. Rownames are taken as the
signature names and column names are taken as the sample IDs.}

\item{file}{File to which to write the exposure matrix (as a CSV file).}
}
\description{
Write an exposure matrix to a file
}
\examples{
file <- system.file("extdata",
                    "synthetic.exposure.csv",
                    package = "ICAMS")
exposure <- ReadExposure(file)
WriteExposure(exposure, file = file.path(tempdir(), "synthetic.exposure.csv"))
}
