% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAWG7_interaction_functions.R
\name{WriteMinimalVCFList}
\alias{WriteMinimalVCFList}
\title{WriteMinimalVCFList}
\usage{
WriteMinimalVCFList(list.of.VCFs, path)
}
\arguments{
\item{list.of.VCFs}{TODO}

\item{path}{TODO}
}
\value{
TODO
}
\description{
Each VCF in list.of.VCFs needs to have columns
CHROM, POS, REF, ALT, VAF, STRAND ??????.
They will be extracted and re-ordered if necessary
to be consistent. The header line will begin with
#CHROM
}
