% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{AddAndCheckSequenceID}
\alias{AddAndCheckSequenceID}
\title{Add sequence context to a data frame with ID (insertion/deletion) mutation records,
 and confirm that they match the given reference genome.}
\usage{
AddAndCheckSequenceID(df, genome, flag.mismatches = FALSE)
}
\arguments{
\item{df}{A data frame storing mutation records of a VCF file
 containing only insertions and deletions. This function expects that
 there is a "context base" to the left, for example REF = ACG, ALT = A
(deletion of CG) or REF = A, ALT = ACC (insertion of CC).}

\item{genome}{A particular reference genome(without quotation marks). Use
\code{BSgenome::available.genomes()} to get the list of "BSgenome data
packages" curently available. There are 2 types of predefined reference
genome which are incorporated in this function. User can invoke a
predefined human GRCh38/hg38 BSgenome data package by typing \code{genome =
"GRCh38"} or \code{genome = "hg38"}. User can invoke a predefined human
GRCh37/hg19 BSgenome data package by typing \code{genome = "GRCh37"} or
\code{genome = "hg19"}.}

\item{flag.mismatches}{If > 0, if there are mismatches to references, print
out the first \code{flag.mismatches} rows and continue.  Otherwise \code{stop}.}
}
\value{
A data frame with 2 new columns added to the input data frame:
\enumerate{
 \item \code{seq.context} The sequence embedding the variant.

 \item \code{seq.context.width} The width of \code{seq.context} to the left
    of the variant. Does not include the "context base". TODO(Steve): do we need
    to modify this function so that it can handle indel callers that do
    not provide the "context base"?
}
}
\description{
Add sequence context to a data frame with ID (insertion/deletion) mutation records,
 and confirm that they match the given reference genome.
}
\keyword{internal}
