% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{AddDBSClass}
\alias{AddDBSClass}
\title{Add DBS mutation class to an annotated DBS VCF}
\usage{
AddDBSClass(vcf)
}
\arguments{
\item{vcf}{An in-memory VCF file annotated with sequence context and
transcript information by function \code{\link{AnnotateDBSVCF}}. It must
*not* contain indels and must *not* contain SBS (single base
substitutions), or triplet base substitutions etc.}
}
\value{
The original \code{vcf} with three additional columns
  \code{DBS78.class}, \code{DBS136.class} and \code{DBS144.class} added.
}
\description{
Add DBS mutation class to an annotated DBS VCF
}
\keyword{internal}
