% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{AddSequence}
\alias{AddSequence}
\title{Add sequence context to a data frame with mutation records}
\usage{
AddSequence(df, seq = BSgenome.Hsapiens.1000genomes.hs37d5)
}
\arguments{
\item{df}{An input data frame storing mutation records of a VCF file.}

\item{seq}{A particular reference genome.}
}
\value{
A data frame with a new column added to the input data frame,
    which contains sequence context information.
}
\description{
Add sequence context to a data frame with mutation records
}
\keyword{internal}
