% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all.abundance}
\alias{all.abundance}
\title{K-mer abundances.}
\format{An object of class \code{list} of length 2.}
\usage{
all.abundance
}
\description{
An R list with one element each for 
\code{BSgenome.Hsapiens.1000genomes.hs37d5},
and \code{BSgenome.Hsapiens.UCSC.hg38}.
Each element is in turn a sub-list keyed by 
\code{exome}, \code{transcript}, 
and \code{genome}. Each element of the sub list
is keyed by the number of rows in the catalog class (as a string, e.g. 
\code{"78"}, not \code{78}. The keys are:
78 (\code{DBS78Catalog}), 96 (\code{SBS96Catalog}), 136 (\code{DBS136Catalog}),
144 (\code{DBS144Catalog}), 192 (\code{SBS192Catalog}),
 and 1536 (\code{SBS1536Catalog}). So, for example to get the exome
 abundances for SBS96 catalogs for \code{BSgenome.Hsapiens.UCSC.hg38} exomes
 one would reference 
 \code{all.abundance[["BSgenome.Hsapiens.UCSC.hg38"]][["exome"]]["96"]}
 or \code{all.abundance$BSgenome.Hsapiens.UCSC.hg38$exome$"96"}.
 The value of the abundance is an integer vector with the the K-mers
 as names and the values being the number of the K-mer.
}
\keyword{datasets}
