% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{AnnotateSBSVCF}
\alias{AnnotateSBSVCF}
\title{Add sequence and transcript information to an in-memory SBS VCF.}
\usage{
AnnotateSBSVCF(SBS.vcf, ref.genome, trans.ranges = NULL)
}
\arguments{
\item{SBS.vcf}{An in-memory SBS VCF as a \code{data.frame}.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{trans.ranges}{a \code{\link[data.table]{data.table}} which contains
transcript range and strand information. Please refer to
\code{\link{TranscriptRanges}} for more details. 
If \code{is.null(trans.ranges)} do not add transcript range
information.}
}
\value{
An in-memory SBS VCF as a \code{data.table}.
This has been annotated with the sequence context
(column name \code{seq.21context})
and with transcript information in the form of a gene symbol
(e.g. \code{"TP53"}) and transcript strand. This information is
in the columns transcript range
   \code{trans.start.pos},
  \code{trans.end.pos}, \code{trans.gene.symbol}, or
  \code{trans.strand} columns to the output. These columns
  are not added if \code{is.null(trans.ranges)}.
}
\description{
Add sequence and transcript information to an in-memory SBS VCF.
}
