% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{as.catalog}
\alias{as.catalog}
\title{Create a catalog from a matrix or matrix-like object}
\usage{
as.catalog(catalog, ref.genome, region, catalog.type)
}
\arguments{
\item{catalog}{A numeric matrix or <Nanhai fill in>.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome", or "transcript".}

\item{catalog.type}{One of "counts", "density", "counts.signature",
"density.signature".}
}
\value{
A catalog as described in \code{\link{ICAMS}}.
}
\description{
TODO(Nanhai): change name of this argument and handle
numeric data frames as well.
}
