% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{as.catalog}
\alias{as.catalog}
\title{Create a catalog from a numeric \code{matrix} or numeric \code{data.frame}.}
\usage{
as.catalog(object, ref.genome = NULL, region = "unknown",
  catalog.type = "counts", abundance = NULL)
}
\arguments{
\item{object}{A numeric \code{matrix} or numeric \code{data.frame}. 
This object must have
 rownames to denote the mutation types. See \code{\link{CatalogRowOrder}}
 for more details.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{A character string designating a region, one of
\code{genome}, \code{transcript}, \code{exome}, \code{unknown};
see \code{\link{ICAMS}}.}

\item{catalog.type}{One of "counts", "density", "counts.signature",
"density.signature".}

\item{abundance}{Optional, only needed when \code{ref.genome}
is not one of
the reference genomes known to ICAMS (see \code{\link{ICAMS}}.
The abundance should
contain the counts of different source sequences for mutations. \cr
See \code{\link{all.abundance}}.}
}
\value{
A catalog as described in \code{\link{ICAMS}}.
}
\description{
Create a catalog from a numeric \code{matrix} or numeric \code{data.frame}.
}
