% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{as.catalog}
\alias{as.catalog}
\title{Create a catalog from a matrix or matrix-like object}
\usage{
as.catalog(object, ref.genome, region, catalog.type, abundance)
}
\arguments{
\item{object}{A numeric matrix or numeric data frame. This object must have
rownames to denote the mutation types. See \code{\link{CatalogRowOrder}}
for more details.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome", or "transcript".}

\item{catalog.type}{One of "counts", "density", "counts.signature",
"density.signature".}

\item{abundance}{Optional, only needed when \code{ref.genome} does not belong
to the two human reference genomes supported by ICAMS. The abundance should
contain the counts of different source sequences for mutations. See
\code{ICAMS:::abundance.3bp.exome.unstranded.GRCh37} for an example.}
}
\value{
A catalog as described in \code{\link{ICAMS}}.
}
\description{
TODO(Nanhai): change name of this argument and handle
numeric data frames as well.
}
