% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{Canonicalize1DEL}
\alias{Canonicalize1DEL}
\title{Canonicalize1DEL}
\usage{
Canonicalize1DEL(ref, alt, context)
}
\arguments{
\item{ref}{TODO}

\item{alt}{TODO}

\item{context}{TODO}
}
\value{
TODO
}
\description{
Canonicalize1DEL
}
\keyword{internal}
