% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{Canonicalize1INS}
\alias{Canonicalize1INS}
\title{Canonicalize1INS}
\usage{
Canonicalize1INS(ref, alt, context)
}
\arguments{
\item{ref}{TODO}

\item{alt}{TODO}

\item{context}{TODO}
}
\value{
TODO
}
\description{
Canonicalize1INS
}
