% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{Canonicalize1INS}
\alias{Canonicalize1INS}
\title{Given an insertion and its sequence context, categorize it.}
\usage{
Canonicalize1INS(context, ins.sequence, pos, trace = 0)
}
\arguments{
\item{context}{The deleted sequence plus ample surrounding
sequence on each side (at least as long as \code{ins.sequence} * 6).}

\item{ins.sequence}{The deleted sequence in \code{context}.}

\item{pos}{The position of \code{ins.sequence} in \code{context}.}

\item{trace}{If > 0, then generate
messages tracing how the computation is carried out.}
}
\value{
A string that is the canonical representation of the given insertion type
}
\description{
Given an insertion and its sequence context, categorize it.
}
\keyword{internal}
