% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNS_related_functions.R
\name{CanonicalizeDNS}
\alias{CanonicalizeDNS}
\title{CanonicalizeDNS}
\usage{
CanonicalizeDNS(ref.vec, alt.vec)
}
\arguments{
\item{ref.vec}{TODO}

\item{alt.vec}{TODO}
}
\value{
TODO
}
\description{
CanonicalizeDNS
}
\keyword{internal}
