% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{CanonicalizeID}
\alias{CanonicalizeID}
\title{Determine the mutation types of insertions and deletions.}
\usage{
CanonicalizeID(context, ref, alt, pos)
}
\arguments{
\item{context}{A vector of ample surrounding
sequence on each side the variants}

\item{ref}{Vector of reference alleles}

\item{alt}{Vector of alternative alleles}

\item{pos}{Vector of the positions of the insertions and deletions in
\code{context}.}
}
\value{
A vector of strings that are the canonical representations
 of the given insertions and deletions.
}
\description{
Determine the mutation types of insertions and deletions.
}
\keyword{internal}
