% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CatalogRowOrder}
\alias{CatalogRowOrder}
\alias{catalog.row.order.SNS.96}
\alias{catalog.row.order.SNS.192}
\title{canonical order of row names in a SNS 96 catalog}
\format{A string of characters}
\usage{
catalog.row.order.SNS.96

catalog.row.order.SNS.192
}
\description{
canonical order of row names in a SNS 96 catalog
}
\keyword{datasets}
