% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CatalogRowOrder}
\alias{CatalogRowOrder}
\alias{catalog.row.order}
\title{Standard order of row names in a catalog.}
\format{A list of character vectors indicating the standard
  orders of row names in catalogs.}
\usage{
catalog.row.order
}
\description{
This data is designed for those, including extenders of this package,
who need to create their own catalogs from formats not
supported by this package.
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  is in the range from 0 to 5+, but for plotting
  and end user documentation it ranges
  from 1 to 6+.
}
\keyword{datasets}
