% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CatalogRowOrder}
\alias{CatalogRowOrder}
\alias{catalog.row.order}
\title{Canonical order of row names in a catalog}
\format{A list which containts string of characters indicating the canonical
  order of row names in a catalog.}
\usage{
catalog.row.order
}
\description{
Canonical order of row names in a catalog
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
\keyword{datasets}
