% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CatalogRowOrder}
\alias{CatalogRowOrder}
\alias{catalog.row.order.SNS.96}
\alias{catalog.row.order.SNS.192}
\alias{catalog.row.order.SNS.1536}
\alias{catalog.row.order.DNS.78}
\alias{catalog.row.order.DNS.144}
\alias{catalog.row.order.DNS.136}
\alias{catalog.row.order.ID}
\title{Canonical order of row names in a catalog}
\format{A string of characters indicating the canonical order of row names in
  a catalog.}
\usage{
catalog.row.order.SNS.96

catalog.row.order.SNS.192

catalog.row.order.SNS.1536

catalog.row.order.DNS.78

catalog.row.order.DNS.144

catalog.row.order.DNS.136

catalog.row.order.ID
}
\description{
Canonical order of row names in a catalog
}
\note{
In the ID (insertion and deletion) catalog, the deletions repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
\keyword{datasets}
