% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CatalogRowOrder}
\alias{CatalogRowOrder}
\alias{catalog.row.order}
\alias{catalog.row.order.sp}
\title{Standard order of row names in a catalog.}
\format{
A list of character vectors indicating the standard
  orders of row names in catalogs.

An object of class \code{list} of length 8.

An object of class \code{list} of length 4.
}
\usage{
catalog.row.order

catalog.row.order.sp
}
\description{
This data is designed for those
who need to create their own catalogs from formats not
supported by this package. The rownames denote the mutation
types.  For example, for SBS96 catalogs, the rowname
 AGAT represents a mutation from AGA > ATA.
}
\note{
In ID (small insertion and deletion) catalogs, deletion repeat sizes
  range from 0 to 5+, but for plotting and end-user documentation deletion
  repeat sizes range from 1 to 6+. In ID83 catalogs, deletion repeat sizes
  range from 0 to 5.
}
\examples{
catalog.row.order$SBS96
# "ACAA" "ACCA" "ACGA" "ACTA" "CCAA" "CCCA" "CCGA" "CCTA" ...
# There are altogether 96 row names to denote the mutation types
# in SBS96 catalog.

catalog.row.order.sp$ID83
# "DEL:C:1:0" "DEL:C:1:1" "DEL:C:1:2" "DEL:C:1:3" ...
# There are altogether 83 row names to denote the mutation types
# in ID83 catalog.
}
\keyword{datasets}
