% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_1_justified_indel.R
\name{categorize_1_justified_indel}
\alias{categorize_1_justified_indel}
\title{Given a indel and its sequence context, categorize it}
\usage{
categorize_1_justified_indel(context, ins_or_del, ins_or_del_seq, pos)
}
\arguments{
\item{context}{The sequence surrounding the indel PRIOR to the insertion
or deletion.}

\item{ins_or_del}{A singgle character, with "i" denoting
an insertion and "d" denotine an deletion.}

\item{ins_or_del_seq}{The the sequence that was inserted or deleted.}

\item{pos}{For deletions, the 1-based position of the start of the
deleted sequence; for insertions, the position immediately to the right
of where the insertion occurs.}
}
\value{
A string that is the canonical representation
of the given deletion type. Return \code{NA}
and raise a warning if
there is an un-normalized representation of
the deletion of a repeat unit.
See \code{FindDelMH} for details.
(This seems to be very rare.)
}
\description{
This function is primarily for internal use, but we export it
to document the underlying logic.
}
\details{
See \url{https://github.com/steverozen/ICAMS/blob/v3.0.9-branch/data-raw/PCAWG7_indel_classification_2021_09_03.xlsx}
for additional information on deletion mutation classification.

This function first handles deletions in homopolymers, then
handles deletions in simple repeats with
longer repeat units, (e.g. \code{CACACACA}, see
\code{\link{FindMaxRepeatDel}}),
and if the deletion is not in a simple repeat,
looks for microhomology (see \code{\link{FindDelMH}}).
}
\examples{
categorize_1_justified_indel("GGAAAGG", "d", ins_or_del_seq = "A", pos = 3) # "DEL:T:1:2"
categorize_1_justified_indel("GGAAAGG", "d", ins_or_del_seq = "A", pos = 4) # "DEL:T:1:2"
categorize_1_justified_indel("TTATT", "d", ins_or_del_seq = "A", pos = 3)   # "DEL:T:1:0"

}
