% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_name_functions.R
\name{CheckAndFixChrNamesForTransRanges}
\alias{CheckAndFixChrNamesForTransRanges}
\title{Check and, if possible, correct the chromosome names in a trans.ranges \code{data.table}}
\usage{
CheckAndFixChrNamesForTransRanges(
  trans.ranges,
  vcf.df,
  ref.genome,
  name.of.VCF = NULL
)
}
\arguments{
\item{trans.ranges}{A \code{\link[data.table]{data.table}} which contains
transcript range and strand information. Please refer to
\code{\link{TranscriptRanges}} for more details.}

\item{vcf.df}{A VCF as a \code{data.frame}. Check the names in column
\code{CHROM}.}

\item{ref.genome}{The reference genome with the chromosome names to check
\code{vcf.df$CHROM} against; must be a Bioconductor 
\code{\link[BSgenome]{BSgenome}}, e.g.
\code{\link[BSgenome.Hsapiens.UCSC.hg38]{BSgenome.Hsapiens.UCSC.hg38}}.}

\item{name.of.VCF}{Name of the VCF file.}
}
\value{
If the \code{vcf.df$CHROM} values are correct or can be corrected,
  then a vector of chromosome names that can be used as a replacement for
  \code{trans.ranges$chrom}. If the names in \code{vcf.df$CHROM} cannot be
  made to be consistent with the chromosome names in
  \code{trans.ranges$chrom}, then \code{stop}.
}
\description{
Check and, if possible, correct the chromosome names in a trans.ranges \code{data.table}
}
\keyword{internal}
