% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{CheckAndReturnDBSCatalogs}
\alias{CheckAndReturnDBSCatalogs}
\title{Check and return DBS catalogs}
\usage{
CheckAndReturnDBSCatalogs(
  catDBS78,
  catDBS136,
  catDBS144 = NULL,
  discarded.variants,
  annotated.vcfs
)
}
\arguments{
\item{catDBS78}{An DBS78 catalog.}

\item{catDBS136}{An DBS136 catalog.}

\item{catDBS144}{An DBS144 catalog.}

\item{discarded.variants}{A list of discarded variants.}

\item{annotated.vcfs}{A list of annotated VCFs.}
}
\value{
A list of DBS catalogs. Also return the discarded variants and
  annotated VCFs if they exit.
}
\description{
Check and return DBS catalogs
}
\keyword{internal}
