% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{CheckAndReturnSBSMatrix}
\alias{CheckAndReturnSBSMatrix}
\title{Check and return the SBS mutation matrix}
\usage{
CheckAndReturnSBSMatrix(
  vcf,
  discarded.variants,
  mat96,
  mat1536,
  mat192 = NULL,
  return.annotated.vcf = FALSE,
  sample.id = "counts"
)
}
\arguments{
\item{vcf}{An in-memory VCF file annotated with sequence context and
transcript information by function \code{\link{AnnotateSBSVCF}}. It must
*not* contain indels and must *not* contain DBS (double base
substitutions), or triplet base substitutions etc., even if encoded as
neighboring SBS.}

\item{discarded.variants}{A \code{data.frame} which contains rows of SBS
variants whose pentanucleotide context contains "N".}

\item{mat96}{The SBS96 mutation count matrix.}

\item{mat1536}{The SBS1536 mutation count matrix.}

\item{mat192}{The SBS192 mutation count matrix.}

\item{return.annotated.vcf}{Whether to return the annotated VCF with
additional columns showing the mutation class for each variant. Default is
FALSE.}

\item{sample.id}{Usually the sample id, but defaults to "count".}
}
\description{
Check and return the SBS mutation matrix
}
\section{Value}{
 A list of three 1-column matrices with the names
  \code{catSBS96}, \code{catSBS192}, \code{catSBS1536}. If transcript
  information is not available in \code{vcf}, \code{catSBS192} is not
  generated. Do not rely on the order of elements in the list. If
  \code{return.annotated.vcf} = TRUE, another element \code{annotated.vcf}
  will appear in the list. If there are SBS variants whose pentanucleotide
  context contains "N", they will be excluded in the analysis and an
  additional element \code{discarded.variants} will appear in the return
  list.
}

\keyword{internal}
