% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CheckCatalogAttribute}
\alias{CheckCatalogAttribute}
\title{Check attributes of catalog specified by user}
\usage{
CheckCatalogAttribute(ref.genome, region, type)
}
\arguments{
\item{ref.genome}{A genome argument as described in \code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}

\item{type}{A character string acting as a catalog type identifier, one of
"counts", "density", "signature".}
}
\value{
TRUE
}
\description{
Check attributes of catalog specified by user
}
\keyword{internal}
