% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CollapseCatalog}
\alias{CollapseCatalog}
\alias{Collapse192To96}
\alias{Collapse1536To96}
\alias{Collapse144To78}
\title{"Collapse" a catalog.}
\usage{
Collapse192To96(catalog)

Collapse1536To96(catalog)

Collapse144To78(catalog)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}}.}
}
\value{
A catalog as defined in \code{\link{ICAMS}}.
}
\description{
"Collapse" a catalog. Do not use this function for
signature catalogs.
}
\details{
\code{Collapse192To96} Collapse an SNS 192 catalog
to an SNS 96 catalog.

\code{Collapse1536To96} Collapse an SNS 1536 catalog
 to an SNS 96 catalog.

\code{Collapse144To78} Collapse a DNS 144 catalog
to a DNS 78 catalog.
}
