% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CollapseCatalog}
\alias{CollapseCatalog}
\alias{Collapse192To96}
\alias{Collapse1536To96}
\alias{Collapse144To78}
\title{Collapse Catalog Functions}
\usage{
Collapse192To96(catalog)

Collapse1536To96(catalog)

Collapse144To78(catalog)
}
\arguments{
\item{catalog}{A catalog matrix to be collapsed whose row names indicate the
mutation types while its columns show the occurrences of each mutation
type of different samples.}
}
\value{
A canonical catalog matrix whose row names indicate the mutation
  types while its columns show the occurrences of each mutation type of
  different samples.
}
\description{
Collapse a catalog matrix to a canonical one
}
\details{
\code{Collapse192To96} Collapse a SNS 192 catalog matrix to a SNS 96 catalog matrix.

\code{Collapse1536To96} Collapse a SNS 1536 catalog matrix to a SNS 96 catalog matrix.

\code{Collapse144To78} Collapse a DNS 144 catalog matrix to a DNS 78 catalog matrix.
}
