% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateCatalogAbundance}
\alias{CreateCatalogAbundance}
\title{Create the abundance attribute of a catalog}
\usage{
CreateCatalogAbundance(catalog, ref.genome, region)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}}.}

\item{ref.genome}{A genome argument as described in \code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}
}
\value{
The original catalog with abundance attribute added.
}
\description{
Create the abundance attribute of a catalog
}
\keyword{internal}
