% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateCatalogAbundance}
\alias{CreateCatalogAbundance}
\title{Create the abundance attribute of a catalog}
\usage{
CreateCatalogAbundance(object, ref.genome, region, catalog.type)
}
\arguments{
\item{object}{A numeric matrix or numeric data frame. This object must have
rownames to denote the mutation types. See \code{\link{CatalogRowOrder}}
for more details.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}

\item{catalog.type}{One of "counts", "density", "counts.signature",
"density.signature".}
}
\value{
The original catalog with abundance attribute added.
}
\description{
Create the abundance attribute of a catalog
}
\keyword{internal}
