% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateCatalogAttribute}
\alias{CreateCatalogAttribute}
\title{Create an S3 object of class "catalog"}
\usage{
CreateCatalogAttribute(catalog, ref.genome, region, type)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}}.}

\item{ref.genome}{A reference genome as described in \code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}

\item{type}{A character string acting as a catalog type identifier, one of
"counts", "density", "signature".}
}
\value{
An S3 object of class "catalog".
}
\description{
Create an S3 object of class "catalog"
}
