% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateCatalogClass}
\alias{CreateCatalogClass}
\title{Create the additional class attribute of a catalog}
\usage{
CreateCatalogClass(object)
}
\arguments{
\item{object}{A numeric matrix or numeric data frame. This object must have
rownames to denote the mutation types. See \code{\link{CatalogRowOrder}}
for more details.}
}
\value{
The original object with additional class attribute added.
}
\description{
Create the additional class attribute of a catalog
}
\keyword{internal}
