% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateDinucAbundance}
\alias{CreateDinucAbundance}
\title{Create dinucleotide abundance}
\usage{
CreateDinucAbundance(path)
}
\arguments{
\item{path}{Path to the file with the nucleotide abundance information with 4
base pairs.}
}
\value{
An integer vector whose names indicate 10 different types of 2 base pairs
  combinations while its values indicate the occurrences of each type.
}
\description{
Create dinucleotide abundance
}
\keyword{internal}
