% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateExomeStrandedRanges}
\alias{CreateExomeStrandedRanges}
\title{Create exome transcriptionally stranded regions}
\usage{
CreateExomeStrandedRanges(path, trans.ranges)
}
\arguments{
\item{path}{Path to a SureSelect BED file which contains unstranded exome
ranges.}

\item{trans.ranges}{A data.table which contains transcript range and strand
information. Please refer to \code{\link{TranscriptRanges}} for more
details.}
}
\value{
A data table which contains chromosome name, start, end position,
  strand information. It is keyed by chrom, start, and
  end.
}
\description{
Create exome transcriptionally stranded regions
}
\keyword{internal}
