% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{CreateOneColDBSCatalog}
\alias{CreateOneColDBSCatalog}
\title{Create double base catalog for *one* sample from
a Variant Call Format (VCF) file}
\usage{
CreateOneColDBSCatalog(vcf, sample.id = "count")
}
\arguments{
\item{vcf}{An in-memory VCF file annotated by the AddSequence and
AddTranscript functions. It must *not* contain indels and must
*not* contain SBS (single base substitutions), or triplet base
substitutions etc.}

\item{sample.id}{Usually the sample id, but defaults to "count".}
}
\value{
A list of three matrices containing the DBS catalog:
  catDBS78, catDBS144, catDBS136 respectively.
}
\description{
Create double base catalog for *one* sample from
a Variant Call Format (VCF) file
}
\note{
DBS 144 catalog only contains mutations in transcribed regions.
}
\keyword{internal}
