% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{CreateOneColIDCatalog}
\alias{CreateOneColIDCatalog}
\title{Create an indel (ID) mutation catalog for *one* sample from a Variant Call Format (VCF)
file}
\usage{
CreateOneColIDCatalog(ID.vcf, SBS.vcf)
}
\arguments{
\item{ID.vcf}{An in-memory VCF as a data.frame annotated by the AddSequence
  and AddTranscript functions. It must only contain indels and must *not*
  contain SNS (single nucleotide substituions), DNS, or triplet base substitutions
  etc.

  * Sequence must already have been added to ID.vcf

  One design decision for variant callers is the representation of "complex
  indels", e.g. mutations e.g. CAT > GC. Some callers represent this as C>G,
  A>C, and T>_. Others might represent it as CAT > CG. Multiple issues can
  arise. In PCAWG, overlapping indel/SBS calls from different callers were
  included in the indel VCFs.}

\item{SBS.vcf}{An in-memory VCF as a data frame. Because we have to work with
some PCAWG data, we will look for neigboring indels and indels adjoining
SBS. That means this functions takes an SBS VCF and an ID VCF from the same
sample.}
}
\value{
A list with two elemsents:
  ID.cat:   A 1-column matrix containing the mutation catalog information.
  problems: Locations of neighboring indels or indels neighboring SBS.
            In the future we might handle these depending on what we
            find in the indel calls from different variant callers.
TODO(steve) Is problems implemented?
}
\description{
Create an indel (ID) mutation catalog for *one* sample from a Variant Call Format (VCF)
file
}
\keyword{internal}
