% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{CreateOneColSBSCatalog}
\alias{CreateOneColSBSCatalog}
\title{Create single base mutation catalog for *one* sample from
a Variant Call Format (VCF) file.}
\usage{
CreateOneColSBSCatalog(vcf, sample.id = "count")
}
\arguments{
\item{vcf}{An in-memory VCF file annotated by the AddSeqContext and
AddTranscript functions. It must *not* contain indels and must *not*
contain DBS (double base substitutions), or triplet base substitutions
etc., even if encoded as neighboring SBS.}

\item{sample.id}{Usually the sample id, but defaults to "count".}
}
\value{
A list of three matrices containing the SBS mutation catalog:
  96, 192, 1536 catalog respectively.
}
\description{
Create single base mutation catalog for *one* sample from
a Variant Call Format (VCF) file.
}
\note{
catSBS192 only contains mutations in transcribed regions.
}
\keyword{internal}
