% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_related_functions.R
\name{CreateOneColSNSCatalog}
\alias{CreateOneColSNSCatalog}
\title{Create single nucleotide mutation catalog for *one* sample from
a Variant Call Format (VCF) file.}
\usage{
CreateOneColSNSCatalog(vcf, sample.id = "count")
}
\arguments{
\item{vcf}{An in-memory VCF file annotated by the AddSequence and
AddTranscript functions. It must *not* contain indels and must *not*
contain DNS (double nucleotide substituions), or triplet base substituions
etc., even if encoded as neighboring SNS.}

\item{sample.id}{Usually the sample id, but defaults to "count".}
}
\value{
A list of three matrices containing the SNS mutation catalog:
  96, 192, 1536 catalog respectively.
}
\description{
Create single nucleotide mutation catalog for *one* sample from
a Variant Call Format (VCF) file.
}
