% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreatePentanucAbundance}
\alias{CreatePentanucAbundance}
\title{Create pentanucleotide abundance}
\usage{
CreatePentanucAbundance(path)
}
\arguments{
\item{path}{Path to the file with the nucleotide abundance information
with 5 base pairs.}
}
\value{
A numeric vector whose names indicate 512 different types of 5 base
  pairs combinations while its values indicate the occurrences of each type.
}
\description{
Create pentanucleotide abundance
}
\keyword{internal}
