% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreatePentanucAbundance}
\alias{CreatePentanucAbundance}
\title{Create pentanucleotide abundance file}
\usage{
CreatePentanucAbundance(path)
}
\arguments{
\item{path}{Path to the file with the nucleotide abundance information
with 5 base pairs.}
}
\value{
A matrix whose row names indicate 512 different types of 5 base
  pairs combinations while its column contains the occurrences of each type.
}
\description{
Create pentanucleotide abundance file
}
