% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateStrandedTrinucAbundance}
\alias{CreateStrandedTrinucAbundance}
\title{Create stranded trinucleotide abundance}
\usage{
CreateStrandedTrinucAbundance(path)
}
\arguments{
\item{path}{Path to the file with the nucleotide abundance information with 3
base pairs.}
}
\value{
A numeric vector whose names indicate 64 different types of 3 base pairs
  combinations while its values indicate the occurrences of each type.
}
\description{
Create stranded trinucleotide abundance
}
\keyword{internal}
