% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{CreateTetranucAbundance}
\alias{CreateTetranucAbundance}
\title{Create tetranucleotide abundance file}
\usage{
CreateTetranucAbundance(path)
}
\arguments{
\item{path}{Path to the file with the nucleotide abundance information with 4
base pairs.}
}
\value{
A matrix whose row names indicate 136 different types of 4 base pairs
  combinations while its column contains the occurrences of each type.
}
\description{
Create tetranucleotide abundance file
}
