% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{FindDelMH}
\alias{FindDelMH}
\title{FindDelMH}
\usage{
FindDelMH(context, q, pos)
}
\arguments{
\item{context}{TODO}

\item{q}{TODO}

\item{pos}{TODO}
}
\value{
TODO
}
\description{
Microhomology can be alligned in multiple equivalent ways.
Example:
}
\details{
GGCTAGTT aligned to

GGCTAGAACTAGTT
GG------CTAGTT GGCTAGTT GG[CTAGAA]CTAGTT
                           ----   ----
GGC------TAGTT GGCTAGTT GGC[TAGAAC]TAGTT
                          * ---  * ---
GGCT------AGTT GGCTAGTT
GGCTA------GTT GGCTAGTT
GGCTAG------TT GGCTAGTT

All the same pairs of sequence, aligned 5 different ways.
4 bp of microhomology.

Need to find:

(1) The maxium match of undeleted sequence on left that is
identical to the right end of deleted sequence, and

(2) The maxium match of undeleted sequence on right that is
identical to the left end of deleted sequence.

The microhomology sequence is the concatenation of items
(1) and (2).
}
