% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{FindMaxRepeatDel}
\alias{FindMaxRepeatDel}
\title{FindMaxRepeatDel}
\usage{
FindMaxRepeatDel(context, q, pos)
}
\arguments{
\item{context}{TODO}

\item{q}{TODO}

\item{pos}{TODO}
}
\value{
TODO
}
\description{
q is a substring of context at pos to pos + len(q) - 1
}
\details{
e.g. q = ac
pos = 3
context = xyaczt
pos         ^

for deletion, if substr(context, pos, pos + len(q) - 1) != q
there is an error.
}
