% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{FindMaxRepeatDel}
\alias{FindMaxRepeatDel}
\title{Return the number of repeat units in which a deletion
is embedded. TODO(Steve): check this statement; what
if there is no repeat?}
\usage{
FindMaxRepeatDel(context, q, pos)
}
\arguments{
\item{context}{A string that embeds \code{q} at position \code{pos}}

\item{q}{A substring of \code{context} at \code{pos} to \code{pos + nchar(q) - 1}}

\item{pos}{The position of \code{q}}
}
\value{
The number of repeat units in which \code{q} is embedded.
}
\description{
e.g. q = ac
pos = 3
context = xyaczt
pos         ^
Return 1
}
\details{
If \code{substr(context, pos, pos + nchar(q) - 1) != q} then stop
}
