% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{FindMaxRepeatIns}
\alias{FindMaxRepeatIns}
\title{FindMaxRepeatIns}
\usage{
FindMaxRepeatIns(context, q, pos)
}
\arguments{
\item{context}{TODO}

\item{q}{TODO}

\item{pos}{TODO}
}
\value{
TODO
}
\description{
If q is an insertion into context between pos and pos+1
if q is repeated in context it might start at pos+1:
}
\details{
e.g. q = ac
pos = 4
context = abxyac
 pos         ^
 start        ^

or q might start at pos + 1 - len(q)

e.g. q = ac
pos = 4
context = xyaczz
 pos         ^
 start      ^
}
