% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.expression.level.example.GRCh37}
\alias{gene.expression.level.example.GRCh37}
\title{Example gene expression level values for human GRCh37.}
\format{A \code{data.frame} which contains the transcription level of genes.}
\usage{
gene.expression.level.example.GRCh37
}
\description{
This data is designed to be used as an example in function
\code{\link{PlotTransBiasExp}} and \code{\link{PlotTransBiasExpToPdf}}.
}
\keyword{datasets}
