% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneExpressionData}
\alias{GeneExpressionData}
\alias{gene.expression.data.HepG2}
\alias{gene.expression.data.MCF10A}
\title{Example gene expression data from two cell lines}
\format{
A \code{\link{data.table}} which contains the expression values of genes.

An object of class \code{data.table} (inherits from \code{data.frame}) with 57736 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 57736 rows and 4 columns.
}
\usage{
gene.expression.data.HepG2

gene.expression.data.MCF10A
}
\description{
This data is designed to be used as an example in function \cr
\code{\link{PlotTransBiasGeneExp}} and \code{\link{PlotTransBiasGeneExpToPdf}}.
}
\examples{
gene.expression.data.HepG2
# Ensembl.gene.ID  gene.symbol  counts            TPM
# ENSG00000000003       TSPAN6    6007   33.922648455
# ENSG00000000005         TNMD       0    0.000000000
# ENSG00000000419         DPM1    4441   61.669371091
# ENSG00000000457        SCYL3    1368    3.334619195
# ENSG00000000460     C1orf112     916    2.416263423
#             ...          ...     ...            ...
}
\keyword{datasets}
