% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{GenerateKmer}
\alias{GenerateKmer}
\title{Generate all possible k-mers of length k.}
\usage{
GenerateKmer(k)
}
\arguments{
\item{k}{Length of kmers (k>=2)}
}
\value{
Character vector containing all possible k-mers.
}
\description{
Generate all possible k-mers of length k.
}
\keyword{internal}
