% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{GetConsensusVAF}
\alias{GetConsensusVAF}
\title{Analogous to \code{\link{GetMutectVAF}}, calculating VAF and read depth
from PCAWG7 consensus vcfs}
\usage{
GetConsensusVAF(vcf, mc.cores = 1)
}
\arguments{
\item{vcf}{An in-memory VCF data frame.}

\item{mc.cores}{The number of cores to use. Not available on Windows
unless \code{mc.cores = 1}.}
}
\description{
Analogous to \code{\link{GetMutectVAF}}, calculating VAF and read depth
from PCAWG7 consensus vcfs
}
\keyword{internal}
