% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{GetMutationLoadsFromMutectVCFs}
\alias{GetMutationLoadsFromMutectVCFs}
\title{Get mutation loads information from Mutect VCF files}
\usage{
GetMutationLoadsFromMutectVCFs(list)
}
\arguments{
\item{list}{A list generated by calling function
\code{\link{ReadAndSplitMutectVCFs}} to Mutect VCF files.}
}
\value{
A list containing the total mutation loads of Mutect VCF files
  according to mutation type:
  
\enumerate{

 \item \code{SBS} Single base substitution.

 \item \code{DBS} Double base substitution.

 \item \code{ID} Small insertion and deletion.

 \item \code{excluded.variants} Other types of mutations which
 are excluded in the analysis in \code{\link{ICAMS}}.

}
}
\description{
Get mutation loads information from Mutect VCF files
}
\keyword{internal}
