% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{GetMutationLoadsFromMutectVCFs}
\alias{GetMutationLoadsFromMutectVCFs}
\title{Get mutation loads information from Mutect VCF files.}
\usage{
GetMutationLoadsFromMutectVCFs(catalogs)
}
\arguments{
\item{catalogs}{A list generated by calling function
\code{\link{MutectVCFFilesToCatalog}} to Mutect VCF files.}
}
\value{
A list containing mutation loads information from Mutect VCF files:

\enumerate{
 \item \code{total.variants} Total number of mutations.

 \item \code{SBS} Number of single base substitutions.

 \item \code{DBS} Number of double base substitutions.

 \item \code{ID} Number of small insertions and deletions.

 \item \code{discarded.variants} Number of other types of mutations which are
 excluded in the analysis in \code{\link{ICAMS}}.

}
}
\description{
Get mutation loads information from Mutect VCF files.
}
\keyword{internal}
