% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{GetMutationLoadsFromStrelkaSBSVCFs}
\alias{GetMutationLoadsFromStrelkaSBSVCFs}
\title{Get mutation loads information from Strelka SBS VCF files.}
\usage{
GetMutationLoadsFromStrelkaSBSVCFs(split.vcfs)
}
\arguments{
\item{split.vcfs}{A list generated by calling function
\code{\link{ReadAndSplitStrelkaSBSVCFs}} to Strelka SBS VCF files.}
}
\value{
A list containing mutation loads information from Strelka SBS VCF
  files:
  
\enumerate{
 \item \code{total.variants} Total number of mutations.
 
 \item \code{SBS} Number of single base substitutions.

 \item \code{DBS} Number of double base substitutions.

 \item \code{ID} Number of small insertions and deletions.

 \item \code{excluded.variants} Number of other types of mutations which are
 excluded in the analysis in \code{\link{ICAMS}}.

}
}
\description{
Get mutation loads information from Strelka SBS VCF files.
}
\keyword{internal}
