% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{GetStrandedKmerCounts}
\alias{GetStrandedKmerCounts}
\title{Generate stranded k-mer abundance from a given genome and gene annotation file}
\usage{
GetStrandedKmerCounts(k, ref.genome, trans.ranges, homopolymer.filter.path)
}
\arguments{
\item{k}{Length of k-mers (k>=2)}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{homopolymer.filter.path}{If given, homopolymers will be masked from
genome(sequence). Only simplerepeat masking is accepted now.}

\item{trans.range}{A GFF3 trans.range.file}
}
\value{
Matrix of the counts of each stranded k-mer across the \code{ref.genome}
}
\description{
Generate stranded k-mer abundance from a given genome and gene annotation file
}
\keyword{internal}
