% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{GetVAF}
\alias{GetVAF}
\alias{GetStrelkaVAF}
\alias{GetMutectVAF}
\title{Extract the VAFs (variant allele frequencies) from a VCF file.}
\usage{
GetStrelkaVAF(vcf)

GetMutectVAF(vcf)
}
\arguments{
\item{vcf}{Said VCF as a data.frame.}
}
\value{
A vector of VAFs, one for each row of \code{vcf}.
}
\description{
Extract the VAFs (variant allele frequencies) from a VCF file.
}
