% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICAMS.R
\docType{package}
\name{ICAMS}
\alias{ICAMS}
\alias{ICAMS-package}
\title{ICAMS: In-depth Characterization and Analysis of Mutational Signatures}
\description{
A toolkit for analysis and visualization
of experimentally elucidated mutational signatures -- the kind of analysis
and visualization presented in Boot et al., "In-depth characterization of
the cisplatin mutational signature in human cell lines and in esophageal and
liver tumors", *Genome Research,* 2018 [https://genome.cshlp.org/content/28/5/654.short].
}
\details{
ICAMS can read in variant call files (VCFs)
generated by Strelka or Mutect, and collate the
mutations  into "catalogs" of mutational
spectra. ICAMS can plot the catalogs of mutational spectra
and mutational signatures.

ICAMS can build and plot catalogs of mutational spectra
for single nucleotide
substitutions (SNS), double nucleotide substitutions (DNS),
and small insertions and deletions (ID). It can also read and
write these catalogs.
}
\section{Creating catalogs from from variant call files (VCF files)}{

\enumerate{
\item \code{\link{StrelkaSNSVCFFilesToCatalog}}, which creates 3 SNS catalogs (96,
192, 1536) and 3 DNS catalogs (78, 136, 144) from the Strelka SNS VCFs.

\item \code{\link{StrelkaIDVCFFilesToCatalog}}, which creates ID (indels) catalog
from the Strelka ID VCFs.

\item \code{\link{MutectVCFFilesToCatalog}}, which creates 3 SNS catalogs (96,
192, 1536), 3 DNS catalogs (78, 136, 144) and ID (indels) catalog from the
Mutect VCFs.
}
}

\section{Plotting catalogs}{

Functions for plotting catalogs of mutational spectra or
of mutational signatures to a PDF file.
\code{\link{PlotCatalogToPdf}}
}

\section{Writing catalogs}{

Functions for writing a catalog mutational spectra or of
mutational signatures to a file on disk.
\code{\link{WriteCatalog}}
}

\section{Reading catalogs}{

Functions for reading files that contain catalogs of mutational
spectra or signatures in standardized format. These
also work for reading mutational signature profiles.
\code{\link{ReadCatalog}}
}

\section{Transforming catalogs}{
 TODO(Steve): not done.
Functions for transforming count spectra from a particular organism region to
an inferred count spectra based on the target nucleotide abundance.
\code{\link{TransformSpectra}}
}

\section{Collapsing catalogs}{

Functions for collapsing a mutational spectrum or signature catalog
based on a more fined-grained set of features (for example, single-nucleotide
substitutions in the context of the preceding and following 2 bases)
to a catalog or spectrum based on a coarser-grained set of features
(for eample, single-nucleotide substitutions in the context of the
immediately precedig and following bases).
\code{\link{CollapseCatalog}}
}

\section{Reading and splitting variant call (VCF) files}{

These are lower level functions for reading in VCF files
and splitting them into VCFs for SNSs, DNSs, and IDs.
*TODO(steve): why do we need this / or we need to provide the
functions that take the split VCFs files and annotate them
and make catalogs, right?*
\enumerate{

\item \code{\link{ReadAndSplitStrelkaSNSVCFs}} Read and split Strelka single
nucleotide substitution (SNS) VCFs (not Strelka indel VCFS).

\item \code{\link{ReadStrelkaIDVCFs}} Read Strelka indel (ID) VCFs (not Strelka
SNS VCFS).

\item \code{\link{ReadAndSplitMutectVCFs}} Read and split Mutect VCFs, which
contain indels and double nucleotide substitutions (DNSs) as well and SNSs.
}
}

\section{Exported data}{
 \enumerate{

\item \code{\link{CatalogRowOrder}} Canonical order of row names in a catalog.

\item \code{\link{TranscriptRanges}} Transcript ranges and strand information
for a particular organism. }
}

