% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICAMS.R
\docType{package}
\name{ICAMS}
\alias{ICAMS}
\alias{ICAMS-package}
\title{ICAMS: In Depth Characterization and Analysis of Mutational Signatures}
\description{
This package provides ...
}
\section{Plotting Signatures}{

Functions for comparing mutational signatures and
sets of mutational signatures:
\code{\link{PlotCat96}}
}

\section{Generating mutational signature catalogs from VCF files}{

fwriteDataFrame
}

\keyword{internal}
