% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICAMS.R
\docType{package}
\name{ICAMS}
\alias{ICAMS}
\alias{ICAMS-package}
\title{ICAMS: In-depth Characterization and Analysis of Mutational Signatures}
\description{
This package has functions to read in VCF files from Strelka and GATK,
create SNS, DNS, ID catalogs and do different types of plotting.
}
\section{Plotting catalogs}{

Functions for plotting the mutation catalog of one sample:
\code{\link{PlotCatalog}}

Functions for plotting mutation catalog of different samples to a PDF file:
\code{\link{CatalogToPdf}}
}

\section{Reading catalogs}{

Functions for reading a catalog in PCAWG7 format from path:
\code{\link{ReadCatalog}}
}

\section{Writing catalogs}{

Functions for writting a mutation catalog to a file on disk:
\code{\link{WriteCatalog}}
}

