% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICAMS.R
\docType{package}
\name{ICAMS}
\alias{ICAMS}
\alias{ICAMS-package}
\title{ICAMS: In-depth Characterization and Analysis of Mutational Signatures}
\description{
This package has functions to read in VCF files from Strelka and Mutect (in
the Broad GATK package), create, read, and write single nucleotide
substitutions (SNS), double nucleotide substitutions (DNS), insertions and
deletions (ID) catalogs and do different types of plotting.
}
\details{
This alpha version only works with VCFs for human GRCh37, but will work for
arbitrary \strong{human} catalogs (assuming no major change in "opportunities"
between GRCh37 and GRCh38).
}
\section{Reading VCF files}{
 \enumerate{

\item \code{\link{ReadListOfStrelkaSNSVCFs}} Read Strelka single nucleotide
substitution (SNS) VCFs (not Strelka indel VCFS).

\item \code{\link{ReadListOfStrelkaIDVCFs}} Read Strelka indel (ID) VCFs (not Strelka
SNS VCFS).

\item \code{\link{ReadListOfMutectVCFs}} Read Mutect VCFs, which contain indels and double
nucleotide substitutions (DNSs) as well and SNSs.
}
}

\section{Splitting in-memory VCFs}{

\enumerate{
\item \code{\link{SplitListOfStrelkaSNSVCFs}} Split Strelka SNS VCFs
into pairs of SNS and inferred DNS VCFs. This is necessary because Strelka
does not call DNSs directly. Instead this functions merges adjacent SNSs
into DNSs provided their VAFs (variant allele frequencies) are similar.

\item \code{\link{SplitListOfMutectVCFs}} Split Mutect VCFs into
 SNS, DNS, and indel components. (Mutect VCFS include SNSs, DNSs, and indels,
 as well as a few other, extremely rare mutation types, such as triplet
 nucleotide substitutions).
}
}

\section{Creating catalogs from VCF files}{

\enumerate{
\item \code{\link{StrelkaSNSVCFFilesToCatalog}}, which creates 3 SNS catalogs (96,
192, 1536) and 3 DNS catalogs (78, 136, 144) from the Strelka SNS VCFs.\cr
\item \code{\link{StrelkaIDVCFFilesToCatalog}}, which creates ID (indels) catalog
from the Strelka ID VCFs.\cr
\item \code{\link{MutectVCFFilesToCatalog}}, which creates 3 SNS catalogs (96,
192, 1536), 3 DNS catalogs (78, 136, 144) and ID (indels) catalog from the
Mutect VCFs.
}
}

\section{Reading catalogs}{

Functions for reading files that contain mutational
spectrum catalogs in standardized format. These
also work for reading mutational signature profiles.
\code{\link{ReadCatalog}}
}

\section{Writing catalogs}{

Functions for writing a mutational spectrum catalog to a file on disk.
These also work for writing mutationl signature profiles.
\code{\link{WriteCatalog}}
}

\section{Collapsing catalogs}{

Functions for collapsing a mutation catalog:
\code{\link{CollapseCatalog}}
}

\section{Plotting catalogs}{

Functions for plotting mutation spectrum catalogs
to a PDF file. These also work for plotting
mutational signature profiles.
\code{\link{PlotCatalogToPdf}}
}

