% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICAMS.R
\docType{package}
\name{ICAMS}
\alias{ICAMS}
\alias{ICAMS-package}
\title{ICAMS: In-depth Characterization and Analysis of Mutational Signatures}
\description{
A toolkit for analysis and visualization
of experimentally elucidated mutational signatures -- the kind of analysis
and visualization presented in Boot et al., "In-depth characterization of
the cisplatin mutational signature in human cell lines and in esophageal and
liver tumors", \emph{Genome Research}, 2018,
https://genome.cshlp.org/content/28/5/654.short.
}
\details{
ICAMS can read in variant call files (VCFs)
generated by Strelka or Mutect, and collate the
mutations  into "catalogs" of mutational
spectra.
ICAMS can create and plot catalogs of mutational spectra
or signatures for single nucleotide
substitutions (SNS), double nucleotide substitutions (DNS),
and small insertions and deletions (ID). It can also read and
write these catalogs.
}
\section{Catalogs}{

A key data type in \code{ICAMS} is a "catalog" of mutation counts, of
mutation densities, or of mutational signatures.

A catalog has one of the following types:
\enumerate{
\item Matrix of mutation counts (one column per sample), representing
(count-based) mutational spectra.

\item Matrix of mutation densities, i.e. mutations per occurrences
 of source sequences (one column per sample), representing
 (density-based) mutational spectra.

\item Matrix of mutational signatures, which
are similar to spectra. However where spectra consist of
counts or densities of mutations in each mutation class
(e.g. ACA > AAA, ACA > AGA, ACA > ATA, ACC > AAC, ...),
signatures consist of
the proportions of mutations in each class (with all the
proportions summing to 1). A mutational signature can be based
on either:
\itemize{
  \item 3.1 mutation counts (a "count-based mutational signature"), or
  \item 3.2 mutation densities (a "density-based mutational signature").
}
}

Many functions take the argument \code{catalog.type}, with possible
values \code{"counts"}, \code{"density"}, \code{"counts.signature"},
or \code{"density.signature"}, corresponding to the types of catalogs
in items 1, 2, 3.1, and 3.2, above.

Catalogs are implemented as S3 objects of class \code{matrix} and
one of the ICAMS classes \code{SNS96Catalog}, \code{SNS192Catalog},
\code{SNS1536Catalog}, \code{DNS78Catalog}, \code{DNS144Catalog},
\code{DNS136Catalog}, \code{IndelCatalog}.

If you need to create a catalog from a source other than
this package (i.e. other than with
\code{\link{ReadCatalog}}
or \code{\link{StrelkaSNSVCFFilesToCatalog}},
\code{\link{MutectVCFFilesToCatalog}}, etc.), then use
\code{\link{as.catalog}}.
}

\section{Creating catalogs from variant call files (VCF files)}{

\enumerate{
\item \code{\link{StrelkaSNSVCFFilesToCatalog}} creates 3 SNS catalogs (96,
192, 1536) and 3 DNS catalogs (78, 136, 144) from the Strelka SNS VCFs.

\item \code{\link{StrelkaIDVCFFilesToCatalog}} creates ID (indel) catalog
from the Strelka ID VCFs.

\item \code{\link{MutectVCFFilesToCatalog}} creates 3 SNS catalogs (96,
192, 1536), 3 DNS catalogs (78, 136, 144) and ID (indel) catalog from the
Mutect VCFs.
}
}

\section{The \code{ref.genome} argument}{

Many functions take the argument \code{ref.genome}.
This can be either
\enumerate{
  \item A variable from the Bioconductor \link[BSgenome]{BSgenome} package
  that contains a particular reference genome, for example
  \code{BSgenome.Hsapiens.1000genomes.hs37d5}.
  \code{BSgenome::available.genomes()} returns the available
  genomes.

 \item The strings \code{"hg38"} or \code{"GRCh38"}
 are shorthand for
 \code{BSgenome.Hsapiens.UCSC.hg38},
 and the strings \code{"hg19"} or \code{"GRCh37"}
 are shorthand for
 \code{BSgenome.Hsapiens.1000genomes.hs37d5}.
 }
}

\section{The Bioconductor \code{BSgenome} package}{

 This package will be installed automatically if \code{\link{ICAMS}}
 is installed with \code{devtools::install_local} or
 with \code{devtools::install_github}.
 Otherwise you must
 manually install \code{BSgenome} and the
 necessary genomes, e.g. \preformatted{
 BSgenome.Hsapiens.1000genomes.hs37d5.}
 See instructions at \preformatted{
 https://bioconductor.org/packages/release/bioc/html/BSgenome.html.}
 Genomes other than the two human genomes mentioned above must
 be installed manually.

 Use \code{\link[BSgenome]{available.genomes}} to get the list of
 available genomes.
}

\section{Plotting catalogs}{

The \code{\link{PlotCatalog}} functions plot mutational spectra
for one sample or plot one mutational signature.

The \code{\link{PlotCatalogToPdf}}
functions plot catalogs of mutational spectra or
of mutational signatures to a PDF file.
}

\section{Writing catalogs}{

The \code{\link{WriteCatalog}} functions
write a catalog of mutational spectra or of
mutational signatures to a file.
}

\section{Reading catalogs}{

The \code{\link{ReadCatalog}} functions
read a file that contains a catalog of mutational
spectra or of signatures in standardized format.
}

\section{Transforming catalogs}{

The \code{\link{TransformCatalog}}
function transforms catalogs of mutational spectra or
signatures to account for differing abundances of the source
sequence of the mutations in the genome.

For example, mutations from
ACG are much rarer in the human genome than mutations from ACC
simply because CG dinucleotides are rare in the genome.
Consequently, there are two possible representations of
mutational spectra or signatures. One representation is
based on mutation counts as observed in a given genome,
and this approach is widely used, as, for example, at
https://cancer.sanger.ac.uk/cosmic/signatures, which
presents signatures based on observed mutation counts
in the human genome. We call these "count-based spectra"
or "count-based signatures".

Alternatively,
mutational spectra or signatures can be represented as
mutations per source sequence, for example
the number of ACT > AGT mutations occurring at all
ACT 3-mers in a genome. We call these "density-based
spectra" or "density-based signatures".

This function can also transform spectra
based on observed genome-wide counts to "density"-based
catalogs. In density-based catalogs
mutations are expressed as mutations per
source sequences. For example,
a density-based catalog represents
the proportion of ACCs mutated to
ATCs, the proportion of ACGs mutated to ATGs, etc.
This is
different from count-based catalogs, which
contain the number of ACC > ATC mutations, the number of
ACG > ATG mutations, etc.

This function can also transform observed-count based
spectra or signatures from genome to exome based counts,
or between different species (since the abundances of
source sequences vary between genome and exome and between
species).
}

\section{Collapsing catalogs}{

The \code{\link{CollapseCatalog}} functions
\enumerate{
\item take a mutational spectrum or signature catalog
that is based on a fined-grained set of features (for example, single-nucleotide
substitutions in the context of the preceding and following 2 bases), and

\item collapse it to a catalog based on a coarser-grained set of features
(for example, single-nucleotide substitutions in the context of the
immediately preceding and following bases).
}
}

\section{Data}{

 \enumerate{

\item \code{\link{CatalogRowOrder}} Standard order of rownames in a catalog.
The rownames of encode the type of each mutation. The rownames denote the mutation
types.  For example, for SNS96 catalogs, the rowname
 AGAT represents a mutation from AGA > ATA.

\item \code{\link{TranscriptRanges}} Transcript ranges and strand information
for a particular reference genome.

 }
}

