% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justify_id_vcf.R
\name{justify_id_vcf}
\alias{justify_id_vcf}
\title{Add sequence context and transcript information to an in-memory ID
(insertion/deletion) VCF, and confirm that they match the given reference
genome}
\usage{
justify_id_vcf(
  ID.vcf,
  ref.genome,
  name.of.VCF = NULL,
  suppress.discarded.variants.warnings = TRUE,
  explain_indels = 1,
  context_width_multiplier = 20L
)
}
\arguments{
\item{ID.vcf}{An in-memory ID (insertion/deletion) VCF as a
\code{data.frame}. This function expects that there is a "context base" to
the left, for example REF = ACG, ALT = A (deletion of CG) or REF = A, ALT =
ACC (insertion of CC).}

\item{ref.genome}{Can be a string or a BSgenome. If a string, it should
a well-known name for reference genome in BSgenome}

\item{name.of.VCF}{Name of the VCF file.}

\item{suppress.discarded.variants.warnings}{If TRUE, do warn when variants
that cannot be processed are discarded.}

\item{explain_indels}{If 0, do not explain, if 1, explain indels that were
justifed (via messages), if 2, geenrate messages for all indels.}

\item{context_width_multiplier}{Used to guess how much sequence on each side
of an indel is needed to categorize it.}
}
\value{
A list of elements:
\itemize{
\item \code{annotated.vcf}: The original VCF data
frame in which \code{POS}, \code{REF} and \code{ALT} with new columns added to the input data frame:
\itemize{
\item \code{seq.context}: The sequence embedding the variant.
\item \code{seq.context.width}: The width of \code{seq.context} to the left of the indel
\item \code{pos_shift}
}
\item \code{discarded.variants}: \strong{Non-NULL only if} there are variants
that were excluded from the analysis. See the added extra column
\code{discarded.reason} for more details.
}
}
\description{
Add sequence context and transcript information to an in-memory ID
(insertion/deletion) VCF, and confirm that they match the given reference
genome
}
\examples{
file <- c(system.file("tests/testthat/testdata/",
                      "vcf_to_test_justify_indels_in_id_vcf_with_contexts.csv",
                       package = "ICAMS"))

stopifnot(file.exists(file))

read.csv(file) |> dplyr::select(-seq.context.width, -seq.context) -> vcf_no_contexts

if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
    justify_id_vcf(vcf_no_contexts, ref.genome = "hg19", explain_indels = 2)}

}
