% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justify_indel.R
\name{justify_indel}
\alias{justify_indel}
\title{Move the notional position of a deletion or insertion as far left as possible.}
\usage{
justify_indel(long_str, short_str, pos, expected_delta = NULL)
}
\arguments{
\item{long_str}{A single character string}

\item{short_str}{A single character string.}

\item{pos}{An integer; see the description}

\item{expected_delta}{A single string; see the desctiption}
}
\value{
A list with elements leftmost_pos and del_string
}
\description{
short_string should be generated by a deletion in long-str
at (1-based) position pos. If supplied, the expected_delta
should be the deleted sequence. If expected_delta is not
NULL, the function checks this and aborts if this is not
the case.

Dually, pos is the 1-based position in short string
in front o which one can
make an insertion to get long_str.

This funcition moves pos as far to left as possible so that a deletion
at that position still results in an edit of long_str
to generate short_str.
}
