% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{MakeVCFDNSdf}
\alias{MakeVCFDNSdf}
\title{Take DNS ranges and the original VCF and generate a VCF with
dinucleotide REF and ALT alleles. The output VCF has minimal columns:
just CHROM, POS, ID, REF, ALT.}
\usage{
MakeVCFDNSdf(DNS.range.df, SNS.vcf.dt)
}
\arguments{
\item{DNS.range.df}{Data frame with columns CHROM, LOW, HIGH}

\item{SNS.vcf.dt}{TODO}
}
\value{
TODO
}
\description{
Take DNS ranges and the original VCF and generate a VCF with
dinucleotide REF and ALT alleles. The output VCF has minimal columns:
just CHROM, POS, ID, REF, ALT.
}
