% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{MutectVCFFilesToCatalog}
\alias{MutectVCFFilesToCatalog}
\title{Create SBS, DBS and Indel catalogs from Mutect VCF files}
\usage{
MutectVCFFilesToCatalog(
  files,
  ref.genome,
  trans.ranges = NULL,
  region = "unknown",
  names.of.VCFs = NULL,
  tumor.col.names = NA,
  updateProgress = NULL
)
}
\arguments{
\item{files}{Character vector of file paths to the Mutect VCF files.}

\item{updateProgress}{Optional. Currently only used in ICAMS.shiny package to
update the progress indicator.}
}
\value{
A list of 3 SBS catalogs (one each for 96, 192, and 1536), 3 DBS
  catalogs (one each for 78, 136, and 144) and ID catalog. If trans.ranges =
  NULL, SBS 192 and DBS 144 catalog will not be generated. Each catalog has
  attributes added. See \code{\link{as.catalog}} for more details.
}
\description{
Create 3 SBS catalogs (96, 192, 1536), 3 DBS catalogs (78, 136, 144) and
Indel catalog from the Mutect VCFs specified by \code{files}
}
\details{
This function calls \code{\link{VCFsToSBSCatalogs}},
\code{\link{VCFsToDBSCatalogs}} and \code{\link{VCFsToIDCatalogs}}
}
\note{
SBS 192 and DBS 144 catalogs include only mutations in transcribed
  regions. In ID (small insertion and deletion) catalogs, deletion repeat sizes
  range from 0 to 5+, but for plotting and end-user documentation deletion
  repeat sizes range from 1 to 6+.
}
\section{Comments}{

To add or change attributes of the catalog, you can use function
\code{\link[base]{attr}}. \cr For example, \code{attr(catalog, "abundance")
<- custom.abundance}.
}

\examples{
file <- c(system.file("extdata/Mutect-vcf",
                      "Mutect.GRCh37.vcf",
                      package = "ICAMS"))
if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
  catalogs <- MutectVCFFilesToCatalog(file, ref.genome = "hg19", 
                                      trans.ranges = trans.ranges.GRCh37,
                                      region = "genome")}
}
