% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{MutectVCFFilesToCatalog}
\alias{MutectVCFFilesToCatalog}
\title{Create SBS, DBS and Indel catalogs from Mutect VCF files}
\usage{
MutectVCFFilesToCatalog(files, ref.genome, trans.ranges, region)
}
\arguments{
\item{files}{Character vector of file paths to the Mutect VCF files.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{trans.ranges}{a \code{\link[data.table]{data.table}} which contains
transcript range and strand information. Please refer to
\code{\link{TranscriptRanges}} for more details.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}
}
\value{
A list of 3 SBS catalogs (one each for 96, 192, and 1536), 3 DBS
  catalogs (one each for 78, 136, and 144) and ID catalog. Each catalog has
  attributes added. See \code{\link{as.catalog}} for more
  details.
}
\description{
Create 3 SBS catalogs (96, 192, 1536), 3 DBS catalogs (78, 136, 144) and
Indel catalog from the Mutect VCFs specified by \code{files}
}
\details{
This function calls \code{\link{VCFsToSBSCatalogs}},
\code{\link{VCFsToDBSCatalogs}} and \code{\link{VCFsToIDCatalogs}}
}
\note{
SBS 192 and DBS 144 catalogs include only mutations in transcribed regions.
}
