% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{MutectVCFFilesToCatalogAndPlotToPdf}
\alias{MutectVCFFilesToCatalogAndPlotToPdf}
\title{Create SBS, DBS and Indel catalogs from Mutect VCF files and plot them to PDF}
\usage{
MutectVCFFilesToCatalogAndPlotToPdf(vector.of.file.paths, ref.genome,
  trans.ranges, region, file, no.context)
}
\arguments{
\item{vector.of.file.paths}{Character vector of file paths to the Mutect VCF
files.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{trans.ranges}{A data.table which contains transcript range and strand
information. Please refer to \code{\link{TranscriptRanges}} for more
details.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}

\item{file}{The name of the PDF file to be produced.}

\item{no.context}{A logical value indicating whether there is preceding and
following base context for the plot. Only implemented for SBS192Catalog.}
}
\value{
A list of 3 SBS catalogs (one each for 96, 192, and 1536), 3 DBS
  catalogs (one each for 78, 136, and 144), Indel catalog and their plots
  outputed to PDF with specified file name. Each catalog has attributes
  added. See \code{\link{as.catalog}} for more details.
}
\description{
Create 3 SBS catalogs (96, 192, 1536), 3 DBS catalogs (78, 136, 144) and
Indel catalog from the Mutect VCFs specified by vector.of.file.paths and plot them
to PDF
}
\details{
This function calls \code{\link{MutectVCFFilesToCatalog}} and
\code{\link{PlotCatalogToPdf}}
}
\note{
SBS 192 and DBS 144 catalogs include only mutations in transcribed regions.
}
