% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{NormalizeGenomeArg}
\alias{NormalizeGenomeArg}
\title{Take strings representing a genome and return the \code{\link[BSgenome]{BSgenome}} object.}
\usage{
NormalizeGenomeArg(ref.genome)
}
\arguments{
\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}
}
\value{
If \code{ref.genome} is 
a \code{\link[BSgenome]{BSgenome}} object, return it.
Otherwise return the \code{\link[BSgenome]{BSgenome}} object identified by the
string \code{ref.genome}.
}
\description{
Take strings representing a genome and return the \code{\link[BSgenome]{BSgenome}} object.
}
\keyword{internal}
