% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{NormalizeGenomeArg}
\alias{NormalizeGenomeArg}
\title{Take strings representing a genome and return the \link[BSgenome] object.}
\usage{
NormalizeGenomeArg(genome)
}
\arguments{
\item{genome}{Either a variable containing a BSgenome object or
a character string acting as a genome identifier.}
}
\value{
If \code{genome} is \code{BSgenome} object, return it.
Otherwise return the \code{BSgenome} object identified by the
string \code{genome}.
}
\description{
Take strings representing a genome and return the \link[BSgenome] object.
}
\keyword{internal}
