% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog}
\alias{PlotCatalog}
\alias{PlotCatalog.SBS96Catalog}
\alias{PlotCatalog.SBS192Catalog}
\alias{PlotCatalog.SBS192CatalogNoContext}
\alias{PlotCatalog.SBS1536Catalog}
\alias{PlotCatalog.DBS78Catalog}
\alias{PlotCatalog.DBS144Catalog}
\alias{PlotCatalog.DBS136Catalog}
\alias{PlotCatalog.IndelCatalog}
\title{Plot \strong{one} spectrum or signature.}
\usage{
PlotCatalog(catalog, no.context = FALSE, ...)

\method{PlotCatalog}{SBS96Catalog}(catalog, cex = 0.8, grid = TRUE,
  upper = TRUE, xlabels = TRUE)

\method{PlotCatalog}{SBS192Catalog}(catalog, cex = 0.8)

\method{PlotCatalog}{SBS192CatalogNoContext}(catalog, no.context = TRUE,
  cex = 1)

\method{PlotCatalog}{SBS1536Catalog}(catalog)

\method{PlotCatalog}{DBS78Catalog}(catalog)

\method{PlotCatalog}{DBS144Catalog}(catalog, cex = 1)

\method{PlotCatalog}{DBS136Catalog}(catalog)

\method{PlotCatalog}{IndelCatalog}(catalog)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{no.context}{If TRUE, no preceding and following base context for
SBS192 catalog plot.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the spectrum of \strong{one} sample or plot \strong{one} signature.
}
\note{
The sizes of repeats involved in deletions range from 0 to 5+ in the
  mutational-spectra and signature catalog rownames, but for plotting and
  end-user documentation they range from 1 to 6+.
}
