% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog}
\alias{PlotCatalog}
\alias{PlotCatalog.SNS96Catalog}
\alias{PlotCatalog.SNS192Catalog}
\alias{PlotCatalog.SNS192CatalogNoContext}
\alias{PlotCatalog.SNS1536Catalog}
\alias{PlotCatalog.DNS78Catalog}
\alias{PlotCatalog.DNS144Catalog}
\alias{PlotCatalog.DNS136Catalog}
\alias{PlotCatalog.IndelCatalog}
\title{Plot \strong{one} spectrum or signature.}
\usage{
PlotCatalog(catalog, no.context = FALSE, ...)

\method{PlotCatalog}{SNS96Catalog}(catalog, cex = 0.8, grid = TRUE,
  upper = TRUE, xlabels = TRUE)

\method{PlotCatalog}{SNS192Catalog}(catalog, cex = 0.8)

\method{PlotCatalog}{SNS192CatalogNoContext}(catalog, no.context = TRUE,
  cex = 1)

\method{PlotCatalog}{SNS1536Catalog}(catalog)

\method{PlotCatalog}{DNS78Catalog}(catalog)

\method{PlotCatalog}{DNS144Catalog}(catalog, cex = 1)

\method{PlotCatalog}{DNS136Catalog}(catalog)

\method{PlotCatalog}{IndelCatalog}(catalog)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{no.context}{If TRUE, no preceding and following base context for
SNS192 catalog plot.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the spectrum of \strong{one} sample or plot \strong{one} signature.
}
\note{
The sizes of repeats involved in deletions range from 0 to 5+ in the
  mutational-specta and signature catalog rownames, but for plotting and
  end-user documentation they range from 1 to 6+.
}
