% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SNS_catalog.R, R/plot_DNS_catalog.R,
%   R/plot_INDELS_catalog.R
\name{PlotCatalog}
\alias{PlotCatalog}
\alias{PlotCat96}
\alias{PlotCat192}
\alias{PlotCat192Strand}
\alias{PlotCat1536}
\alias{PlotCatDNS78}
\alias{PlotCatDNS144}
\alias{PlotQUAD136}
\alias{PlotCatID}
\title{Plot Catalog Functions}
\usage{
PlotCat96(catalog, id, type = "density", cex = 0.8, grid = TRUE,
  upper = TRUE, xlabels = TRUE, abundance = NULL)

PlotCat192(catalog, id, type = "counts", cex = 0.8, abundance = NULL)

PlotCat192Strand(catalog, id, type = "counts", cex = 1,
  abundance = NULL)

PlotCat1536(catalog, id, abundance)

PlotCatDNS78(catalog, id, type = "density", abundance = NULL)

PlotCatDNS144(catalog, id, type = "counts", cex = 1,
  abundance = NULL)

PlotQUAD136(catalog, id = colnames(catalog), type = "density",
  abundance = .abundance.4bp)

PlotCatID(catalog, id, type = "counts")
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the mutation types
while its columns contain the counts of each mutation type.}

\item{id}{The ID information of the sample which has mutations.}

\item{type}{A value indicating the type of graph. If type = "counts", the
graph will plot the occurrences of the mutation types in the sample. If
type = "signature", the graph will plot mutation signatures of the sample.
If type = "density", the graph will plot the rates of mutations per million
nucleotides for each mutation type. (Please take note there is no "density"
type for PlotCatID function and the option of type = "density" is not
implemented for function PlotCat192, PlotCat192Strand and PlotCatDNS144 at
the current stage.)}

\item{cex}{A numerical value giving the amount by which mutation class labels,
mutation counts(if there exists), y axis and its labels, x axis labels and
its annotations(if there exists) sample name and legend(if there exists)
should be magnified relative to the default.}

\item{grid}{A logical value indicating whether to draw the grid lines in the
graph.}

\item{upper}{A logical value indicating whether to draw horizontal lines and
names of major mutation class on top of graph.}

\item{xlabels}{A logical value indicating whether to draw x axis labels.}

\item{abundance}{A matrix containing nucleotide abundance information and
strand information(if there exists), to be used only when type = "density".}
}
\value{
invisible(TRUE)
}
\description{
Plot the mutation catalog of one sample
}
\details{
\code{PlotCat96} Plot the SNS 96 mutation catalog of one sample.

\code{PlotCat192} Plot the SNS 192 mutation catalog of one sample.

\code{PlotCat192Strand} Plot the transcription strand bias graph of 6 SNS
mutation types ("C>A", "C>G", "C>T", "T>A", "T>C", "T>G") in one sample.

\code{PlotCat1536} Plot the pentanucleotide sequence contexts for one sample,
normalized by pentanucleotide occurrence in the genome. The mutation types
are in six-letters like CATTAT, first 2-letters CA refers to (-2, -1)
position, third letter T refers to the base which has mutation, next second
2-letters TA refers to (+1, +2) position, last letter T refers to the base
after mutation.

\code{PlotCatDNS78} Plot the DNS 78 mutation catalog of one sample.

\code{PlotCatDNS144} Plot the transcription strand bias graph of 10 major DNS
mutation types ("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN",
"TA>NN", "TC>NN", "TG>NN", "TT>NN") in one sample.

\code{PlotQUAD136} Plot the tetranucleotide sequence contexts of 10 major DNS
mutation types ("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN",
"TA>NN", "TC>NN", "TG>NN", "TT>NN") for one sample, normalized by
tetranucleotide occurrence in the genome.

\code{PlotCatID} Plot the insertion and deletion catalog of one sample.
(Please take note that the deletions Repeat Size ranges from 0 to 5+ in the
catalog, but for plotting and end user documentation it ranges from 1 to 6+.)
}
