% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog}
\alias{PlotCatalog}
\title{Plot \strong{one} spectrum or signature.}
\usage{
PlotCatalog(catalog, strandbias = FALSE, ...)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{strandbias}{If TRUE, plot strand bias graph for SNS192 or DNS144
catalog. Leave out this parameter if you don't intend to plot strand bias
graph.}

\item{...}{Arguments to be passed to methods.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot the spectrum of \strong{one} sample or plot \strong{one} signature.
}
