% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalog}
\alias{PlotCatalog}
\alias{PlotCatSNS96}
\alias{PlotCatSNS192}
\alias{PlotCatSNS192Strand}
\alias{PlotCatSNS1536}
\alias{PlotCatDNS78}
\alias{PlotCatDNS144}
\alias{PlotCatDNS136}
\alias{PlotCatID}
\title{Plot catalog functions}
\usage{
PlotCatSNS96(catalog, id = colnames(catalog), type = "density",
  cex = 0.8, grid = TRUE, upper = TRUE, xlabels = TRUE,
  abundance = NULL)

PlotCatSNS192(catalog, id = colnames(catalog), type = "counts",
  cex = 0.8, abundance = NULL)

PlotCatSNS192Strand(catalog, id = colnames(catalog), type = "counts",
  cex = 1, abundance = NULL)

PlotCatSNS1536(catalog, abundance, id = colnames(catalog))

PlotCatDNS78(catalog, id = colnames(catalog), type = "density",
  abundance = NULL)

PlotCatDNS144(catalog, id = colnames(catalog), type = "counts",
  cex = 1, abundance = NULL)

PlotCatDNS136(catalog, id = colnames(catalog), type = "density",
  abundance = NULL)

PlotCatID(catalog, id = colnames(catalog), type = "counts")
}
\arguments{
\item{catalog}{A matrix whose rownames indicate the mutation type/sequence
context(CatDNS136) while its columns contain the counts of each mutation
type/sequence context(CatDNS136).}

\item{id}{The identifier of the sample which has mutations.}

\item{type}{A value indicating the type of graph. If type = "counts", the
graph will plot the occurrences of the mutation types in the sample. If
type = "signature", the graph will plot mutation signatures of the sample.
If type = "density", the graph will plot the rates of mutations per million
nucleotides for each mutation type. (Please take note there is no "density"
type for PlotCatID function and the option of type = "density" is not
implemented for function PlotCatSNS192, PlotCatSNS192Strand and PlotCatDNS144 at
the current stage.)}

\item{cex}{A numerical value giving the amount by which mutation class labels,
mutation counts(if it exists), y axis and its labels, x axis labels and
its annotations(if it exists) sample name and legend(if it exists)
should be magnified relative to the default.}

\item{grid}{If TRUE, draw grid lines in the graph.}

\item{upper}{If TRUE, draw horizontal lines and the names of major mutation
class on top of graph.}

\item{xlabels}{If TRUE, draw x axis labels.}

\item{abundance}{a named integer vector, see \link{Abundance}, used only
when \code{type = "density"}.}
}
\value{
invisible(TRUE)
}
\description{
Plot the catalog of one sample which has mutations
}
\details{
\code{PlotCatSNS96} Plot the SNS 96 mutation catalog of one sample.

\code{PlotCatSNS192} Plot the SNS 192 mutation catalog of one sample.

\code{PlotCatSNS192Strand} Plot the transcription strand bias graph of 6 SNS
mutation types ("C>A", "C>G", "C>T", "T>A", "T>C", "T>G") in one sample.

\code{PlotCatSNS1536} Plot the pentanucleotide sequence contexts for one sample,
normalized by pentanucleotide occurrence in the genome. The mutation types
are in six-letters like CATTAT, first 2-letters CA refers to (-2, -1)
position, third letter T refers to the base which has mutation, next second
2-letters TA refers to (+1, +2) position, last letter T refers to the base
after mutation.

\code{PlotCatDNS78} Plot the DNS 78 mutation catalog of one sample.

\code{PlotCatDNS144} Plot the transcription strand bias graph of 10 major DNS
mutation types ("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN",
"TA>NN", "TC>NN", "TG>NN", "TT>NN") in one sample.

\code{PlotCatDNS136} Plot the tetranucleotide sequence context of 10 major DNS
mutation types ("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN",
"TA>NN", "TC>NN", "TG>NN", "TT>NN") for one sample.

\code{PlotCatID} Plot the insertion and deletion catalog of one sample.
(Please take note that deletion repeat size ranges from 0 to 5+ in the
catalog, but for plotting and end user documentation it ranges from 1 to 6+.)
}
