% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf}
\alias{PlotCatalogToPdf}
\alias{PlotCatalogToPdf.SNS96Catalog}
\alias{PlotCatalogToPdf.SNS192Catalog}
\alias{PlotCatalogToPdf.SNSClassStrandBias}
\alias{PlotCatalogToPdf.SNS1536Catalog}
\alias{PlotCatalogToPdf.DNS78Catalog}
\alias{PlotCatalogToPdf.DNSClassStrandBias}
\alias{PlotCatalogToPdf.DNS136Catalog}
\alias{PlotCatalogToPdf.IndelCatalog}
\title{Plot catalogs to a PDF file.}
\usage{
PlotCatalogToPdf(catalog, file, strandbias = FALSE, ...)

\method{PlotCatalogToPdf}{SNS96Catalog}(catalog, file, grid = TRUE,
  upper = TRUE, xlabels = TRUE)

\method{PlotCatalogToPdf}{SNS192Catalog}(catalog, file)

\method{PlotCatalogToPdf}{SNSClassStrandBias}(catalog, file,
  strandbias = TRUE)

\method{PlotCatalogToPdf}{SNS1536Catalog}(catalog, file)

\method{PlotCatalogToPdf}{DNS78Catalog}(catalog, file)

\method{PlotCatalogToPdf}{DNSClassStrandBias}(catalog, file,
  strandbias = TRUE, cex = 1)

\method{PlotCatalogToPdf}{DNS136Catalog}(catalog, file)

\method{PlotCatalogToPdf}{IndelCatalog}(catalog, file)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{file}{The name of the PDF file to be produced.}

\item{strandbias}{If TRUE, plot strand bias graph for SNS192 or DNS144
catalog. Leave out this parameter if you don't intend to plot strand bias
graph.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot catalogs to a PDF file.
}
\note{
The sizes of repeats involved in deletions range from 0 to 5+ in the
  mutational-specta and signature catalog rownames, but for plotting and
  end-user documentation they range from 1 to 6+.
}
