% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf}
\alias{PlotCatalogToPdf}
\title{Plot catalogs to a PDF file.}
\usage{
PlotCatalogToPdf(catalog, filename, strandbias = FALSE, ...)
}
\arguments{
\item{catalog}{An S3 object with class "catalog".
See \code{\link{CreateCatalogAttribute}} for more details.}

\item{filename}{The name of the PDF file to be produced.}

\item{strandbias}{If TRUE, plot strand bias graph for SNS192 or DNS144
catalog. Leave out this parameter if you don't intend to plot strand bias
graph.}

\item{...}{Arguments to be passed to methods.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot catalogs to a PDF file.
}
