% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf}
\alias{PlotCatalogToPdf}
\alias{PlotCatalogToPdf.SNS96Catalog}
\alias{PlotCatalogToPdf.SNS192Catalog}
\alias{PlotCatalogToPdf.SNS192CatalogNoContext}
\alias{PlotCatalogToPdf.SNS1536Catalog}
\alias{PlotCatalogToPdf.DNS78Catalog}
\alias{PlotCatalogToPdf.DNS144Catalog}
\alias{PlotCatalogToPdf.DNS136Catalog}
\alias{PlotCatalogToPdf.IndelCatalog}
\title{Plot catalogs to a PDF file.}
\usage{
PlotCatalogToPdf(catalog, file, no.context = FALSE, ...)

\method{PlotCatalogToPdf}{SNS96Catalog}(catalog, file, grid = TRUE,
  upper = TRUE, xlabels = TRUE)

\method{PlotCatalogToPdf}{SNS192Catalog}(catalog, file)

\method{PlotCatalogToPdf}{SNS192CatalogNoContext}(catalog, file,
  no.context = TRUE)

\method{PlotCatalogToPdf}{SNS1536Catalog}(catalog, file)

\method{PlotCatalogToPdf}{DNS78Catalog}(catalog, file)

\method{PlotCatalogToPdf}{DNS144Catalog}(catalog, file, cex = 1)

\method{PlotCatalogToPdf}{DNS136Catalog}(catalog, file)

\method{PlotCatalogToPdf}{IndelCatalog}(catalog, file)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{file}{The name of the PDF file to be produced.}

\item{no.context}{If TRUE, no preceding and following base context for
SNS192 catalog plot.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot catalogs to a PDF file.
}
\note{
The sizes of repeats involved in deletions range from 0 to 5+ in the
  mutational-specta and signature catalog rownames, but for plotting and
  end-user documentation they range from 1 to 6+.
}
