% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf}
\alias{PlotCatalogToPdf}
\alias{PlotCatSNS96ToPdf}
\alias{PlotCatSNS192ToPdf}
\alias{PlotSNSClassStrandBiasToPdf}
\alias{PlotCatSNS1536ToPdf}
\alias{PlotCatDNS78ToPdf}
\alias{PlotDNSClassStrandBiasToPdf}
\alias{PlotCatDNS136ToPdf}
\alias{PlotCatIDToPdf}
\title{Plot catalogs to a PDF file.}
\usage{
PlotCatSNS96ToPdf(catalog, name, type, id = colnames(catalog),
  grid = TRUE, upper = TRUE, xlabels = TRUE)

PlotCatSNS192ToPdf(catalog, name, id = colnames(catalog),
  type = "counts", cex = 0.8)

PlotSNSClassStrandBiasToPdf(catalog, name, type, id = colnames(catalog),
  cex = 1)

PlotCatSNS1536ToPdf(catalog, name, type, id = colnames(catalog))

PlotCatDNS78ToPdf(catalog, name, type, id = colnames(catalog))

PlotDNSClassStrandBiasToPdf(catalog, name, type, id = colnames(catalog),
  cex = 1)

PlotCatDNS136ToPdf(catalog, name, type, id = colnames(catalog))

PlotCatIDToPdf(catalog, name, type, id = colnames(catalog))
}
\arguments{
\item{catalog}{A catalog as described in \code{\link{ICAMS}}.}

\item{name}{The name of the PDF file to be produced.}

\item{type}{A string specifying the type of the input catalog, one of:
\enumerate{

  \item "counts", show the counts of each mutation type.

  \item "density", show the rates of mutations
  per million source n-mers for each mutation type;
  not supported
  for \code{\link{PlotCatIDToPdf}},
  \code{\link{PlotCatSNS192ToPdf}},
  \code{\link{PlotSNSClassStrandBiasToPdf}}, and
  \code{\link{PlotDNSClassStrandBiasToPdf}}.
  \item "signature", show the proportions of each
  mutation type; not supported for
  \code{\link{PlotCatDNS136ToPdf}}.

}}

\item{id}{A vector containing the identifiers of the samples
or signatures in \code{catalog}.}

\item{grid}{If TRUE, draw grid lines in the graph.}

\item{upper}{If TRUE, draw horizontal lines and the names of major mutation
class on top of graph.}

\item{xlabels}{If TRUE, draw x axis labels.}

\item{cex}{A numerical value giving the amount by which mutation class labels,
y axis labels, sample name and legend (if it exists) should be magnified
relative to the default.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot catalogs to a PDF file.
}
\details{
\code{PlotCatSNS96ToPdf} Plot an SNS 96 catalog
to a PDF file.

\code{PlotCatSNS192ToPdf} Plot an SNS 192 catalog
to a PDF file.

\code{PlotSNSClassStrandBiasToPdf} Plot the transcription strand bias graph of
6 SNS mutation types ("C>A", "C>G", "C>T", "T>A", "T>C", "T>G")
of various samples to a PDF file.

\code{PlotCatSNS1536ToPdf} Plot a 1536 mutation catalog to a PDF
file. The mutation types are in six-letters like CATTAT, first 2-letters CA
refers to (-2, -1) position, third letter T refers to the base which has
mutation, next second 2-letters TA refers to (+1, +2) position, last letter T
refers to the base after mutation.

\code{PlotCatDNS78ToPdf} Plot a DNS 78 mutation catalog
to a PDF file.

\code{PlotDNSClassStrandBiasToPdf} Plot the transcription strand bias graph of
10 major DNS mutation types ("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN",
"GC>NN", "TA>NN", "TC>NN", "TG>NN", "TT>NN") of various samples
to a PDF file.

\code{PlotCatDNS136ToPdf} Plot the tetranucleotide sequence contexts of 10 major
DNS mutation types ("AC>NN", "AT>NN", "CC>NN", "CG>NN", "CT>NN", "GC>NN",
"TA>NN", "TC>NN", "TG>NN", "TT>NN") of various samples to a PDF file.

\code{PlotCatIDToPdf} Plot a insertion and deletion catalog to a PDF file.
(Note that sizes of repeats involved in deletions
range from 0 to 5+ in the catalog rownames,
but for plotting and end user
documentation they ranges from 1 to 6+.)
}
