% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotCatalogToPdf}
\alias{PlotCatalogToPdf}
\alias{PlotCatalogToPdf.SBS96Catalog}
\alias{PlotCatalogToPdf.SBS192Catalog}
\alias{PlotCatalogToPdf.SBS192CatalogNoContext}
\alias{PlotCatalogToPdf.SBS1536Catalog}
\alias{PlotCatalogToPdf.DBS78Catalog}
\alias{PlotCatalogToPdf.DBS144Catalog}
\alias{PlotCatalogToPdf.DBS136Catalog}
\alias{PlotCatalogToPdf.IndelCatalog}
\title{Plot catalogs to a PDF file.}
\usage{
PlotCatalogToPdf(catalog, file, no.context = FALSE, ...)

\method{PlotCatalogToPdf}{SBS96Catalog}(catalog, file, grid = TRUE,
  upper = TRUE, xlabels = TRUE)

\method{PlotCatalogToPdf}{SBS192Catalog}(catalog, file)

\method{PlotCatalogToPdf}{SBS192CatalogNoContext}(catalog, file,
  no.context = TRUE)

\method{PlotCatalogToPdf}{SBS1536Catalog}(catalog, file)

\method{PlotCatalogToPdf}{DBS78Catalog}(catalog, file)

\method{PlotCatalogToPdf}{DBS144Catalog}(catalog, file, cex = 1)

\method{PlotCatalogToPdf}{DBS136Catalog}(catalog, file)

\method{PlotCatalogToPdf}{IndelCatalog}(catalog, file)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{file}{The name of the PDF file to be produced.}

\item{no.context}{If TRUE, no preceding and following base context for
SBS192 catalog plot.}

\item{...}{Additional arguments to be passed to methods.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot catalogs to a PDF file.
}
\note{
The sizes of repeats involved in deletions range from 0 to 5+ in the
  mutational-spectra and signature catalog rownames, but for plotting and
  end-user documentation they range from 1 to 6+.
}
