% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_related_functions.R
\name{PlotExposureInternal}
\alias{PlotExposureInternal}
\title{Plot a single exposure plot}
\usage{
PlotExposureInternal(
  exposure,
  plot.proportion = FALSE,
  plot.legend = TRUE,
  cex.legend = 0.9,
  ...
)
}
\arguments{
\item{exposure}{Exposures as a numerical matrix (or data.frame) with
signatures in rows and samples in columns. Rownames are taken as the
signature names and column names are taken as the sample IDs. If you want
\code{exp} sorted from largest to smallest, use \code{\link{SortExposure}}.
Do not use column names that start with multiple underscores. The exposures
will often be mutation counts, but could also be e.g. mutations per
megabase.}

\item{plot.proportion}{Plot exposure proportions rather than counts.}

\item{plot.legend}{If \code{TRUE}, plot a legend.}

\item{cex.legend}{A numerical value giving the amount by which legend
plotting text and symbols should be magnified relative to the default.}

\item{...}{Parameters passed to \code{\link[graphics]{barplot}}.}
}
\value{
An invisible numeric vector giving the coordinates of all the bar
  midpoints drawn, useful for adding to the graph.
}
\description{
Plot a single exposure plot
}
\keyword{internal}
