% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_context_functions.R
\name{PlotPPM}
\alias{PlotPPM}
\title{Plot position probability matrix (PPM) for *one* sample from a Variant Call Format
(VCF) file.}
\usage{
PlotPPM(ppm, title)
}
\arguments{
\item{title}{The main title of the plot.}

\item{pwm}{A position probability matrix (PPM) for *one* sample.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot position probability matrix (PPM) for *one* sample from a Variant Call Format
(VCF) file.
}
\keyword{internal}
