% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.strandbias.expresslevel.R
\name{PlotTransBiasExpToPdf}
\alias{PlotTransBiasExpToPdf}
\title{Plot Transcriptional Strand Bias on Expression level to PDF}
\usage{
PlotTransBiasExpToPdf(annotated.SBS.vcf, expression.level, ref.genome, n,
  plot.type, file)
}
\arguments{
\item{annotated.SBS.vcf}{An annotated SBS VCF as a data.table which contains
sequence context and transcript information. Please refer to
\code{\link{AnnotateSBSVCF}} for more details.}

\item{expression.level}{A two columns data.frame which contains the
transcription level of genes. The first column should be the gene symbols.
The second column should be the numeric expression level e.g.TPM
(Transcripts Per Kilobase Million).}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{n}{The number of bins that will be plotted in the graph.}

\item{plot.type}{A vector of character indicating types to be plotted. It
should be within "C>A", "C>G", "C>T", "T>A", "T>C", "T>G".}

\item{file}{The name of output file}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot Transcriptional Strand Bias on Expression level to PDF
}
\examples{
file <- c(system.file("extdata",
                      "Strelka.SBS.GRCh37.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadAndSplitStrelkaSBSVCFs(file)
SBS.vcf <- list.of.vcfs$SBS.vcfs[[1]]             
if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
  annotated.SBS.vcf <- AnnotateSBSVCF(SBS.vcf, ref.genome = "hg19",
                                      trans.ranges = trans.ranges.GRCh37)
  PlotTransBiasExpToPdf(annotated.SBS.vcf = annotated.SBS.vcf, 
                        expression.level = sample.gene.expression.value.GRCh37, 
                        ref.genome = "hg19", n = 4, 
                        plot.type = c("C>A","C>G","C>T","T>A","T>C"), 
                        file = file.path(tempdir(), "test.pdf"))
}
}
