% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandbias_functions.R
\name{PlotTransBiasGeneExp}
\alias{PlotTransBiasGeneExp}
\title{Plot transcription strand bias with respect to gene expression values.}
\usage{
PlotTransBiasGeneExp(annotated.SBS.vcf, expression.data,
  Ensembl.gene.ID.col, expression.value.col, num.of.bins, plot.type,
  ymax = NULL)
}
\arguments{
\item{annotated.SBS.vcf}{An SBS VCF annotated by \code{\link{AnnotateSBSVCF}}.}

\item{expression.data}{A \code{\link{data.table}} which contains the
expression values of genes. \cr See \code{\link{GeneExpressionData}} for more
details.}

\item{Ensembl.gene.ID.col}{Name of column which has the Ensembl gene ID
information in \code{expression.data}.}

\item{expression.value.col}{Name of column which has the gene expression
values in \code{expression.data}.}

\item{num.of.bins}{The number of bins that will be plotted on the graph.}

\item{plot.type}{A character string indicating one mutation type to be
plotted. It should be one of "C>A", "C>G", "C>T", "T>A", "T>C", "T>G".}

\item{ymax}{Limit for the y axis. If not specified, it defaults to NULL and
the y axis limit equals 1.5 times of the maximum mutation counts in a
specific mutation type.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot transcription strand bias with respect to gene expression values.
}
\examples{
file <- c(system.file("extdata",
                      "Strelka.SBS.GRCh37.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadAndSplitStrelkaSBSVCFs(file)
SBS.vcf <- list.of.vcfs$SBS.vcfs[[1]]             
if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
  annotated.SBS.vcf <- AnnotateSBSVCF(SBS.vcf, ref.genome = "hg19",
                                      trans.ranges = trans.ranges.GRCh37)
  PlotTransBiasGeneExp(annotated.SBS.vcf = annotated.SBS.vcf, 
                       expression.data = gene.expression.data.HepG2, 
                       Ensembl.gene.ID.col = "Ensembl.gene.ID", 
                       expression.value.col = "TPM", num.of.bins = 4, 
                       plot.type = "C>A")
}
}
