% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandbias_functions.R
\name{PlotTransBiasGeneExpToPdf}
\alias{PlotTransBiasGeneExpToPdf}
\title{Plot Transcription Strand Bias on Expression level to PDF.}
\usage{
PlotTransBiasGeneExpToPdf(annotated.SBS.vcf, file, expression.level,
  Ensembl.gene.ID.col, TPM.col, num.of.bins, plot.type = c("C>A", "C>G",
  "C>T", "T>A", "T>C", "T>G"))
}
\arguments{
\item{annotated.SBS.vcf}{An SBS VCF annotated by \code{\link{AnnotateSBSVCF}}.}

\item{file}{The name of output file.}

\item{expression.level}{A \code{data.frame} which contains the transcription
level of genes. \cr See \code{\link{gene.expression.level.example.GRCh37}}
for more details.}

\item{Ensembl.gene.ID.col}{Name of column which has the Ensembl gene ID
information in \code{experession.level}.}

\item{TPM.col}{Name of column which has the TPM (Transcripts Per Kilobase Million)
information in \code{experession.level}.}

\item{num.of.bins}{The number of bins that will be plotted in the graph.}

\item{plot.type}{A vector of character indicating types to be plotted. It
can be one or more types from "C>A", "C>G", "C>T", "T>A", "T>C", "T>G".
The default is to print all the six mutation types.}
}
\value{
\code{invisible(TRUE)}
}
\description{
Plot Transcription Strand Bias on Expression level to PDF.
}
\examples{
file <- c(system.file("extdata",
                      "Strelka.SBS.GRCh37.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadAndSplitStrelkaSBSVCFs(file)
SBS.vcf <- list.of.vcfs$SBS.vcfs[[1]]             
if (requireNamespace("BSgenome.Hsapiens.1000genomes.hs37d5", quietly = TRUE)) {
  annotated.SBS.vcf <- AnnotateSBSVCF(SBS.vcf, ref.genome = "hg19",
                                      trans.ranges = trans.ranges.GRCh37)
  PlotTransBiasGeneExpToPdf(annotated.SBS.vcf = annotated.SBS.vcf, 
                            expression.level = gene.expression.level.example.GRCh37, 
                            Ensembl.gene.ID.col = "Ensembl.gene.ID", 
                            TPM.col = "TPM", num.of.bins = 4, 
                            plot.type = c("C>A","C>G","C>T","T>A","T>C"), 
                            file = file.path(tempdir(), "test.pdf"))
}
}
