% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{ReadAndSplitMutectVCFs}
\alias{ReadAndSplitMutectVCFs}
\title{Read and split Mutect VCF files}
\usage{
ReadAndSplitMutectVCFs(
  files,
  names.of.VCFs = NULL,
  tumor.col.names = NA,
  suppress.discarded.variants.warnings = TRUE
)
}
\arguments{
\item{files}{Character vector of file paths to the Mutect VCF files.}

\item{names.of.VCFs}{Optional. Character vector of names of the VCF files.
The order of names in \code{names.of.VCFs} should match the order of VCF
file paths in \code{files}. If \code{NULL}(default), this function will
remove all of the path up to and including the last path separator (if any)
in \code{files} and file paths without extensions (and the leading dot)
will be used as the names of the VCF files.}

\item{tumor.col.names}{Optional. Character vector of column names in VCFs
which contain the tumor sample information. The order of names in
\code{tumor.col.names} should match the order of VCFs specified in
\code{files}. If \code{tumor.col.names} is equal to \code{NA}(default),
this function will use the 10th column in all the VCFs to calculate VAFs.
See \code{\link{GetMutectVAF}} for more details.}

\item{suppress.discarded.variants.warnings}{Logical. Whether to suppress
warning messages showing information about the discarded variants. Default
is TRUE.}
}
\description{
Read and split Mutect VCF files
}
\section{Value}{
 A list containing the following objects:
\itemize{
\item \code{SBS}: List of VCFs with only single base substitutions.
\item \code{DBS}: List of VCFs with only doublet base substitutions as called
by Mutect.
\item \code{ID}: List of VCFs with only small insertions and deletions.
\item \code{other.subs}: \strong{Non-NULL only if} there is list of VCF
like data.frames with rows for coordinate substitutions involving 3 or more
nucleotides (e.g. ACT > TGA or AACT > GGTA) and rows for complex indels.
\item \code{multiple.alt}: \strong{Non-NULL only if} there is list of VCF
like data.frames with rows for variants with multiple alternative alleles,
for example ACA mutated to both AGA and ATA at the same position.
\item \code{not.analyzed}: \strong{Non-NULL only if} there is list of VCF
like data.frames with rows for variants that were discarded immediately
after reading in the VCFs. The variants not analyzed can belong to the
following categories:
\itemize{
\item Duplicated "CHROM" and "POS" values.
\item Chromosome names that contain "#".
\item Chromosome names that contain "GL".
\item Chromosome names that contain "KI".
\item Chromosome names that contain "random".
}
\itemize{
\item Chromosome names that contain "Hs".
\item Chromosome names that contain "M".
}
}
}

\examples{
file <- c(system.file("extdata/Mutect-vcf",
                      "Mutect.GRCh37.s1.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadAndSplitMutectVCFs(file)
}
\seealso{
\code{\link{MutectVCFFilesToCatalog}}
}
