% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadAndSplitStrelkaSBSVCFs}
\alias{ReadAndSplitStrelkaSBSVCFs}
\title{Read and split Strelka SBS VCF files.}
\usage{
ReadAndSplitStrelkaSBSVCFs(vector.of.file.paths)
}
\arguments{
\item{vector.of.file.paths}{Character vector of
file paths to the Strelka SBS VCF files.}
}
\value{
A list of 3 in-memory objects as follows:
\enumerate{
   \item \code{SBS.vcfs} List of data.frames of pure SBS mutations -- no DBS or 3+BS mutations.

   \item \code{DBS.vcfs} List of data.frames of pure DBS mutations -- no SBS or 3+BS mutations.

   \item \code{ThreePlus} List of data.tables with the key CHROM, LOW.POS, HIGH.POS. containing
   rows that that in the input that did not represent SBSs or DBSs.

   }
}
\description{
Read and split Strelka SBS VCF files.
}
\seealso{
\code{\link{StrelkaSBSVCFFilesToCatalog}}
}
