% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadAndSplitStrelkaSNSVCFs}
\alias{ReadAndSplitStrelkaSNSVCFs}
\title{Read and split Strelka SNS VCF files.}
\usage{
ReadAndSplitStrelkaSNSVCFs(vector.of.file.paths)
}
\arguments{
\item{vector.of.file.paths}{Character vector of
file paths to Strelka VCF files.}
}
\value{
A list of 3 in-memory objects as follows:
\enumerate{
   \item \code{SNS.vcfs} List of data.frames of pure SNS mutations -- no DNS or 3+BS mutations.

   \item \code{DNS.vcfs} List of data.frames of pure DNS mutations -- no SNS or 3+BS mutations.

   \item \code{ThreePlus} List of data.tables with the key CHROM, LOW.POS, HIGH.POS. containing
   rows that that in the input that did not represent SNSs or DNSs.

   }
}
\description{
Read and split Strelka SNS VCF files.
}
\seealso{
\code{\link{StrelkaSNSVCFFilesToCatalog}}
}
