% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ReadBedRanges}
\alias{ReadBedRanges}
\title{Read chromosome and position information from a bed format file.}
\usage{
ReadBedRanges(file)
}
\arguments{
\item{file}{Path to the file in bed format.}
}
\value{
A data.table keyed by chrom, start, and end. It uses one-based
  coordinates.
}
\description{
Read chromosome and position information from a bed format file.
}
\note{
Only chromosomes 1-22 and X and Y will be kept.
}
\keyword{internal}
