% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ReadBedRanges}
\alias{ReadBedRanges}
\title{Read chromosome and position information from a bed format file.}
\usage{
ReadBedRanges(path)
}
\arguments{
\item{path}{Path to the file in bed format.}
}
\value{
A data.table keyed by chrom, chromStart, and chromEnd.
}
\description{
Read chromosome and position information from a bed format file.
}
\keyword{internal}
