% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ReadBedTranscriptRanges}
\alias{ReadBedTranscriptRanges}
\title{Read transcript ranges and strands from a bed format file.}
\usage{
ReadBedTranscriptRanges(path)
}
\arguments{
\item{path}{Path to the file with the transcript information (in bed format).}
}
\value{
A data.table keyed by chrom, chromStart, and chromEnd.
}
\description{
This function is mostly for testing purpose, may be removed in the future.
}
\keyword{internal}
