% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{ReadCatalog}
\alias{ReadCatalog}
\title{Read catalog.}
\usage{
ReadCatalog(path, ref.genome, region, type, strict = TRUE)
}
\arguments{
\item{path}{Path to a catalog on disk in the standardized format.}

\item{ref.genome}{A character string acting as a genome identifier, one of
"GRCh37", "hg19", "GRCh38", "hg38".}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}

\item{type}{A character string acting as a catalog type identifier, one of
"counts", "density", "signature".}

\item{strict}{If TRUE, then stop if additional checks on the input fail.}
}
\value{
An S3 object with class "catalog".
See \code{\link{CreateCatalogAttribute}} for more details.
}
\description{
Read a catalog in standardized format from path.
}
\details{
See also \code{\link{WriteCatalog}}
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
