% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{ReadCatalog}
\alias{ReadCatalog}
\alias{ReadCat96}
\alias{ReadCat192}
\alias{ReadCat1536}
\alias{ReadCatDNS78}
\alias{ReadCatDNS144}
\alias{ReadCatDNS136}
\alias{ReadCatID}
\title{Read Catalog Functions}
\usage{
ReadCat96(path, strict = TRUE)

ReadCat192(path, strict = TRUE)

ReadCat1536(path, strict = TRUE)

ReadCatDNS78(path, strict = TRUE)

ReadCatDNS144(path, strict = TRUE)

ReadCatDNS136(path, strict = TRUE)

ReadCatID(path, strict = TRUE)
}
\arguments{
\item{path}{Path to a catalog on disk in the "PCAWG7" format.}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}
}
\value{
A catalog in canonical in-memory format.
}
\description{
Read a catalog in PCAWG7 format from path
}
\details{
\code{ReadCat96} Read a 96 SNS catalog from path

\code{ReadCat192} Read a 192 SNS catalog from path

\code{ReadCat1536} Read a 1536 SNS catalog from path

\code{ReadCatDNS78} Read a 78 DNS catalog from path

\code{ReadCatDNS144} Read a 144 DNS catalog from path

\code{ReadCatDNS136} Read a 136 DNS catalog from path

\code{ReadCatID} Read a ID (insertion/deletion) catalog from path
Please take note that the deletions Repeat Size ranges from 0 to 5+
in the catalog, but for plotting and end user documentation
it ranges from 1 to 6+.

See also \code{\link{WriteCatalog}}
}
