% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{ReadCatalog}
\alias{ReadCatalog}
\title{Read catalog.}
\usage{
ReadCatalog(path, ref.genome, region, catalog.type, strict = TRUE)
}
\arguments{
\item{path}{Path to a catalog on disk in the standardized format.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{One of "genome", "exome".}

\item{catalog.type}{One of "counts", "density", "counts.signature",
"density.signature".}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}
}
\value{
A catalog in standard in-memory format with attributes added.
See \code{\link{as.catalog}} for more details.
}
\description{
Read a catalog in standardized format from path.
}
\details{
See also \code{\link{WriteCatalog}}
}
\note{
In ID (insertion and deletion) catalogs, deletion repeat sizes
  range from 0 to 5+, but for plotting and end-user documentation
  deletion repeat sizes range from 1 to 6+.
}
