% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{ReadCatalogInternal}
\alias{ReadCatalogInternal}
\title{Internal read catalog function to be wrapped in a tryCatch}
\usage{
ReadCatalogInternal(
  file,
  ref.genome = NULL,
  region = "unknown",
  catalog.type = "counts",
  strict = TRUE,
  stop.on.error = TRUE
)
}
\arguments{
\item{file}{Path to a catalog on disk in the standardized format.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{region A character string designating a genomic region;
see \code{\link{as.catalog}} and \code{\link{ICAMS}}.}

\item{catalog.type}{One of "counts", "density", "counts.signature",
"density.signature".}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}

\item{stop.on.error}{If TRUE, call \code{stop} on error; otherwise
return a 1-column matrix of NA's with the attribute "error"
containing error information. The number of rows may not
be the correct number for the expected catalog type.}
}
\description{
Internal read catalog function to be wrapped in a tryCatch
}
\keyword{internal}
