% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadMutectVCF}
\alias{ReadMutectVCF}
\title{Read in the data lines of a Variant Call Format (VCF) file created by
    MuTect}
\usage{
ReadMutectVCF(file)
}
\arguments{
\item{file}{The name/path of the VCF file, or a complete URL.}
}
\value{
A data frame storing mutation records of a VCF file with VAFs added.
}
\description{
Read in the data lines of a Variant Call Format (VCF) file created by
    MuTect
}
\keyword{internal}
