% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadMutectVCFs}
\alias{ReadMutectVCFs}
\title{Read Mutect VCF files.}
\usage{
ReadMutectVCFs(files)
}
\arguments{
\item{files}{Character vector of file paths to the VCF files.}
}
\value{
A list of vcfs from \code{files}.
}
\description{
Read Mutect VCF files.
}
\keyword{internal}
