% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadMutectVCFs}
\alias{ReadMutectVCFs}
\title{Read Mutect VCF files.}
\usage{
ReadMutectVCFs(vector.of.file.paths)
}
\arguments{
\item{vector.of.file.paths}{Character vector of
file paths to the VCF files.}
}
\value{
A list of vcfs from vector.of.file.paths.
}
\description{
Read Mutect VCF files.
}
\keyword{internal}
