% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadStrelkaIDVCF}
\alias{ReadStrelkaIDVCF}
\title{Read in the data lines of an ID VCF created by Strelka version 1}
\usage{
ReadStrelkaIDVCF(file)
}
\arguments{
\item{file}{The name/path of the VCF file, or a complete URL.}
}
\value{
A data frame storing mutation records of a VCF file.
}
\description{
Read in the data lines of an ID VCF created by Strelka version 1
}
\note{
In ID (small insertion and deletion) catalogs, deletion repeat sizes
  range from 0 to 5+, but for plotting and end-user documentation
  deletion repeat sizes range from 1 to 6+.
}
\keyword{internal}
