% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_related_functions.R
\name{ReadStrelkaIDVCFs}
\alias{ReadStrelkaIDVCFs}
\title{Read Strelka ID (small insertion and deletion) VCF files}
\usage{
ReadStrelkaIDVCFs(
  files,
  names.of.VCFs = NULL,
  suppress.discarded.variants.warnings = TRUE
)
}
\arguments{
\item{files}{Character vector of file paths to the VCF files.}

\item{names.of.VCFs}{Character vector of names of the VCF files. The order of
names in \code{names.of.VCFs} should match the order of VCF file paths in
\code{files}. If \code{NULL}(default), this function will remove all of the
path up to and including the last path separator (if any) in \code{files}
and file paths without extensions (and the leading dot) will be used as the
names of the VCF files.}

\item{suppress.discarded.variants.warnings}{Logical. Whether to suppress
warning messages showing information about the discarded variants. Default
is TRUE.}
}
\description{
Read Strelka ID (small insertion and deletion) VCF files
}
\section{Value}{

A list of elements:
\itemize{
\item \code{ID.vcfs}: List of data.frames containing ID mutations.
\item \code{discarded.variants}: \strong{Only appearing when} there is list of VCF
like data.frames with rows for variants that were discarded immediately
after reading in the VCFs. The variants not analyzed can belong to the
following categories:
\itemize{
\item Duplicated "CHROM" and "POS" values.
\item Chromosome names that contain "#".
\item Chromosome names that contain "GL".
\item Chromosome names that contain "KI".
\item Chromosome names that contain "random".
}
\itemize{
\item Chromosome names that contain "Hs".
\item Chromosome names that contain "M".
}
}
}

\section{Note}{

In ID (small insertion and deletion) catalogs, deletion repeat sizes range
from 0 to 5+, but for plotting and end-user documentation deletion repeat
sizes range from 1 to 6+.
}

\examples{
file <- c(system.file("extdata/Strelka-ID-vcf",
                      "Strelka.ID.GRCh37.s1.vcf",
                      package = "ICAMS"))
list.of.vcfs <- ReadStrelkaIDVCFs(file)
}
\seealso{
\code{\link{StrelkaIDVCFFilesToCatalog}}
}
