% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadStrelkaIDVCFs}
\alias{ReadStrelkaIDVCFs}
\title{Read Strelka ID (insertion and deletion) VCF files from paths}
\usage{
ReadStrelkaIDVCFs(vector.of.file.paths)
}
\arguments{
\item{vector.of.file.paths}{A vector containing the paths of the VCF files.}
}
\value{
A list of vcfs from vector.of.file.paths.
}
\description{
Read Strelka ID (insertion and deletion) VCF files from paths
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
