% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{ReadStrelkaSBSVCFs}
\alias{ReadStrelkaSBSVCFs}
\title{Read Strelka SBS (single base substitutions) VCF files.}
\usage{
ReadStrelkaSBSVCFs(files)
}
\arguments{
\item{files}{Character vector of file paths to the VCF files.}
}
\value{
A list of vcfs from \code{files}.
}
\description{
Read Strelka SBS (single base substitutions) VCF files.
}
\keyword{internal}
