% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{RemoveRangesOnBothStrand}
\alias{RemoveRangesOnBothStrand}
\title{Remove ranges that fall on both strands}
\usage{
RemoveRangesOnBothStrand(stranded.ranges)
}
\arguments{
\item{stranded.ranges}{A keyed data table which has stranded ranges information.
It has four columns: chrom, start, end and strand.}
}
\value{
A data table which has removed ranges that fall on both strands from
  the input \code{stranded.ranges}.
}
\description{
Remove ranges that fall on both strands
}
\keyword{internal}
