% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{RenameColumnsWithNameStrand}
\alias{RenameColumnsWithNameStrand}
\title{Is there any column in \code{df} with name "strand"?
If there is, change its name to "strand_old" so that it will
conflict with code in other parts of ICAMS package.}
\usage{
RenameColumnsWithNameStrand(df)
}
\description{
Is there any column in \code{df} with name "strand"?
If there is, change its name to "strand_old" so that it will
conflict with code in other parts of ICAMS package.
}
\keyword{internal}
