% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{RenameColumnsWithNameVAF}
\alias{RenameColumnsWithNameVAF}
\title{Is there any column in \code{df} with name "VAF"?
If there is, change its name to "VAF_old" so that it will
conflict with code in other parts of ICAMS package.}
\usage{
RenameColumnsWithNameVAF(df)
}
\description{
Is there any column in \code{df} with name "VAF"?
If there is, change its name to "VAF_old" so that it will
conflict with code in other parts of ICAMS package.
}
\keyword{internal}
