% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{revc}
\alias{revc}
\title{Reverse complement every string in \code{string.vec}.}
\usage{
revc(string.vec)
}
\arguments{
\item{string.vec}{a vector of type character.}
}
\value{
A vector of type characters with the reverse complement of every
  string in \code{string.vec}.
}
\description{
Reverse complement every string in \code{string.vec}.
}
\examples{
# Reverse complement a character of DNA sequence
string1 <- "ATGC"
string2 <- "atgc"
revc(string1)        # "GCAT"
revc(string2)        # "GCAT"

# Reverse complement a vector of character of DNA sequences
strings <- c("ATGC", "TCGT")
revc(strings)        # "GCAT" "ACGA"
}
