% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{revc}
\alias{revc}
\title{Reverse complement every string in string.vec}
\usage{
revc(string.vec)
}
\arguments{
\item{string.vec}{a vector of type character.}
}
\value{
A vector of type characters with the reverse complement of every
  string in string.vec.
}
\description{
Reverse complement every string in string.vec
}
