% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SigPro.to.ICAMS.ID}
\alias{SigPro.to.ICAMS.ID}
\title{Linker from SigProExtractor-formatted to PCAWG(ICAMS)-formatted indel names.}
\format{A 83*1 matrix. Its contents (first column) contain PCAWG(ICAMS)-
formatted indel names in PCAWG(ICAMS) order. Its rownames refer to the 
corresponding SigProExtractor indel names.}
\usage{
SigPro.to.ICAMS.ID
}
\description{
This data is designed for converting SigProExtractor-formatted indel names
to ICAMS-formatted indel names.
}
\examples{
SigPro.to.ICAMS.ID
#              ICAMS.ID.names      
# 1:Del:C:0    "DEL:C:1:0"    
# 1:Del:C:1    "DEL:C:1:1"
# 1:Del:C:2    "DEL:C:1:2"  
# 1:Del:C:3    "DEL:C:1:3" 
# 1:Del:C:4    "DEL:C:1:4"  
#       ...           ...  
}
\keyword{datasets}
