% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{SplitListOfMutectVCFs}
\alias{SplitListOfMutectVCFs}
\title{Split each Mutect VCF into SBS, DBS, and ID VCFs (plus two
VCF-like data frame with left-over rows).}
\usage{
SplitListOfMutectVCFs(list.of.vcfs)
}
\arguments{
\item{list.of.vcfs}{List of VCFs as in-memory data.frames.}
}
\description{
Split each Mutect VCF into SBS, DBS, and ID VCFs (plus two
VCF-like data frame with left-over rows).
}
\section{Value}{
 A list containing the following objects:
\enumerate{
\item \code{SBS}: List of VCFs with only single base substitutions.
\item \code{DBS}: List of VCFs with only doublet base substitutions as called
by Mutect.
\item \code{ID}: List of VCFs with only small insertions and deletions.
\item \code{other.subs}: List of VCF like data.frames with rows for coordinate
substitutions involving 3 or more nucleotides (e.g. ACT > TGA or AACT >
GGTA) and rows for complex indels.
\item \code{multiple.alt}: List of VCF like data.frames with rows for variants
with multiple alternative alleles, for example ACT mutated to both AGT and
ACT at the same position.
\item \code{discarded.variants}: \strong{Only appearing when there are
variants that were discarded.} List of VCF like data.frames with rows for
variants that are excluded in the analysis. The discarded variants belong
to the following categories:
\itemize{
\item Duplicated "CHROM" and "POS" values.
\item Chromosome names that contain "#".
\item Chromosome names that contain "GL".
\item Chromosome names that contain "KI".
\item Chromosome names that contain "random".
\item Chromosome names that contain "Hs".
\item Chromosome names that contain "M".
}
}
}

\keyword{internal}
