% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{SplitListOfStrelkaSBSVCFs}
\alias{SplitListOfStrelkaSBSVCFs}
\title{Split a list of in-memory Strelka SBS VCF into SBS, DBS, and variants involving
> 2 consecutive bases}
\usage{
SplitListOfStrelkaSBSVCFs(
  list.of.vcfs,
  suppress.discarded.variants.warnings = TRUE
)
}
\arguments{
\item{list.of.vcfs}{A list of in-memory data frames containing Strelka SBS
VCF file contents.}

\item{suppress.discarded.variants.warnings}{Logical. Whether to suppress
warning messages showing information about the discarded variants. Default
is TRUE.}
}
\description{
SBSs are single base substitutions,
e.g. C>T, A<G,....  DBSs are double base substitutions,
e.g. CC>TT, AT>GG, ...  Variants involving > 2 consecutive
bases are rare, so this function just records them. These
would be variants such ATG>CCT, AGAT>TCTA, ...
}
\section{Value}{

A list of elements as follows:
\itemize{
\item \code{SBS.vcfs}: List of data.frames of pure SBS mutations -- no DBS or
3+BS mutations.
\item \code{DBS.vcfs}: List of data.frames of pure DBS mutations -- no SBS or
3+BS mutations.
\item \code{ThreePlus}: \strong{Non-NULL only if} there is list of
data.tables with the key CHROM, LOW.POS, HIGH.POS which contain rows in
the input that did not represent SBSs or DBSs.
\item \code{multiple.alt}: \strong{Non-NULL only if} there is list of VCF
like data.frames with rows for variants with multiple alternative alleles,
for example ACA mutated to both AGA and ATA at the same position.
\item \code{not.analyzed}: \strong{Non-NULL only if} there is list of VCF
like data.frames with rows for variants that were discarded immediately
after reading in the VCFs. The variants not analyzed can belong to the
following categories:
\itemize{
\item Duplicated "CHROM" and "POS" values.
\item Chromosome names that contain "#".
\item Chromosome names that contain "GL".
\item Chromosome names that contain "KI".
\item Chromosome names that contain "random".
}
\itemize{
\item Chromosome names that contain "Hs".
\item Chromosome names that contain "M".
}
}
}

\keyword{internal}
