% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{SplitOneMutectVCF}
\alias{SplitOneMutectVCF}
\title{Split a mutect2 VCF into SBS, DBS, and ID VCFs, plus a list of other mutations}
\usage{
SplitOneMutectVCF(vcf.df)
}
\arguments{
\item{vcf.df}{An in-memory data.frame representing a Mutect VCF, including
VAFs, which are added by \code{\link{ReadMutectVCF}}.}
}
\value{
A list of in-memory objects with the elements:
\enumerate{
   \item \code{SBS.vcf}:   Data frame of pure SBS mutations -- no DBS or 3+BS mutations
   \item \code{DBS.vcf}:   Data frame of pure DBS mutations -- no SBS or 3+BS mutations
   \item{ThreePlus}: Data table with the key CHROM, LOW.POS, HIGH.POS and additional
   information (reference sequence, alternative sequence, context, etc.)
   Additional information not fully implemented at this point because of
   limited immediate biological interest.
   \item{multiple.alt}: Rows that were removed before processing because they had
   more than one alternative allele.
   }
}
\description{
Split a mutect2 VCF into SBS, DBS, and ID VCFs, plus a list of other mutations
}
\keyword{internal}
