% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{SplitOneMutectVCF}
\alias{SplitOneMutectVCF}
\title{Split a mutect2 VCF into SNS, DNS, and ID VCFs, plus a list of other mutations

#' @param vcf.df An in-memory data.frame representing a Mutect VCF, including
 VAFs, which are added by \code{\link{ReadMutectVCF}}.}
\usage{
SplitOneMutectVCF(vcf.df)
}
\value{
A list with the SNS, DNS, and ID portions of the VCF file, plus two
data.frames of other mutations
}
\description{
Split a mutect2 VCF into SNS, DNS, and ID VCFs, plus a list of other mutations

#' @param vcf.df An in-memory data.frame representing a Mutect VCF, including
 VAFs, which are added by \code{\link{ReadMutectVCF}}.
}
\keyword{internal}
