% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{SplitOneMutectVCF}
\alias{SplitOneMutectVCF}
\title{Split a mutect2 VCF into SBS, DBS, and ID VCFs, plus a list of other mutations}
\usage{
SplitOneMutectVCF(vcf.df)
}
\arguments{
\item{vcf.df}{An in-memory data.frame representing a Mutect VCF, including
VAFs, which are added by \code{\link{ReadMutectVCF}}.}
}
\value{
A list with the SBS, DBS, and ID portions of the VCF file, plus two
data.frames of other mutations
}
\description{
Split a mutect2 VCF into SBS, DBS, and ID VCFs, plus a list of other mutations
}
\keyword{internal}
