% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{SplitOneVCF}
\alias{SplitOneVCF}
\title{Split a VCF into SBS, DBS, and ID VCFs, plus a list of other mutations}
\usage{
SplitOneVCF(vcf.df, name.of.VCF = NULL)
}
\arguments{
\item{vcf.df}{An in-memory data.frame representing a VCF, including
VAFs, which are added by \code{\link{ReadVCF}}.}

\item{name.of.VCF}{Name of the VCF file.}
}
\value{
A list with 3 in-memory VCFs and discarded variants that were not
  incorporated into the first 3 VCFs:

 * \code{SBS}: VCF with only single base substitutions.

 * \code{DBS}: VCF with only doublet base substitutions
  as called by Mutect.

 * \code{ID}: VCF with only small insertions and deletions.

 * \code{discarded.variants}: \strong{Non-NULL only if} there are variants
 that were excluded from the analysis. See the added extra column
 \code{discarded.reason} for more details.
 @md
}
\description{
Split a VCF into SBS, DBS, and ID VCFs, plus a list of other mutations
}
\keyword{internal}
