% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{StrelkaIDVCFFilesToCatalog}
\alias{StrelkaIDVCFFilesToCatalog}
\title{Create ID (indel) catalog from Strelka ID VCF files}
\usage{
StrelkaIDVCFFilesToCatalog(vector.of.file.paths, genome)
}
\arguments{
\item{vector.of.file.paths}{A vector containing the paths of the Strelka ID VCF files.}

\item{genome}{A particular reference genome(without quotation marks). Use
\link[BSgenome]{available.genomes} to get the list of "BSgenome data
packages" currently available. There are 2 types of predefined reference
genome which are incorporated in this function. User can invoke a
predefined human GRCh38/hg38 BSgenome data package by typing \code{genome =
"GRCh38"} or \code{genome = "hg38"}. User can invoke a predefined human
GRCh37/hg19 BSgenome data package by typing \code{genome = "GRCh37"} or
\code{genome = "hg19"}.}
}
\value{
An ID (indel) catalog
}
\description{
Create ID (indel) catalog from the Strelka ID VCFs specified by vector.of.file.paths
}
\details{
This function calls \code{\link{VCFsToIDCatalogs}}
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
