% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{StrelkaIDVCFFilesToCatalog}
\alias{StrelkaIDVCFFilesToCatalog}
\title{Create ID (indel) catalog from Strelka ID VCF files}
\usage{
StrelkaIDVCFFilesToCatalog(vector.of.file.paths, ref.genome, region)
}
\arguments{
\item{vector.of.file.paths}{Character vector of
file paths to the Strelka ID VCF files.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}
}
\value{
An ID (indel) catalog with attributes added. See
  \code{\link{as.catalog}} for more details.
}
\description{
Create ID (indel) catalog from the Strelka ID VCFs specified by vector.of.file.paths
}
\details{
This function calls \code{\link{VCFsToIDCatalogs}}
}
\note{
In ID (insertion and deletion) catalogs, deletion repeat sizes
  range from 0 to 5+, but for plotting and end-user documentation
  deletion repeat sizes range from 1 to 6+.
}
