% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{StrelkaIDVCFFilesToCatalog}
\alias{StrelkaIDVCFFilesToCatalog}
\title{Create ID (indel) catalog from Strelka ID VCF files}
\usage{
StrelkaIDVCFFilesToCatalog(vector.of.file.paths, genome)
}
\arguments{
\item{vector.of.file.paths}{A vector containing the paths of the Strelka ID VCF files.}

\item{genome}{Name of a particular reference genome
(without quotations marks).}
}
\value{
An ID (indel) catalog
}
\description{
Create ID (indel) catalog from the Strelka ID VCFs specified by vector.of.file.paths
}
\details{
This function calls \code{\link{VCFsToIDCatalogs}}
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
