% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{StrelkaSNSVCFFilesToCatalog}
\alias{StrelkaSNSVCFFilesToCatalog}
\title{Create SNS and DNS catalogs from Strelka SNS VCF files.}
\usage{
StrelkaSNSVCFFilesToCatalog(vector.of.file.paths, ref.genome, trans.ranges,
  region)
}
\arguments{
\item{vector.of.file.paths}{Character vector of
file paths to the Strelka SNS VCF files.}

\item{ref.genome}{A genome argument as described in \code{\link{ICAMS}}.}

\item{trans.ranges}{A data.table which contains transcript range and
strand information.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}
}
\value{
A list of S3 objects with class "catalog". See
  \code{\link{CreateCatalogAttribute}} for more details. There are 3 SNS
  catalogs (one each for 96, 192, and 1536) and 3 DNS catalogs (one each for
  78, 136, and 144)
}
\description{
Create 3 SNS catalogs (96, 192, 1536) and 3 DNS catalogs (78, 136, 144)
from the Strelka SNS VCFs specified by vector.of.file.paths
}
\details{
This function calls \code{\link{VCFsToSNSCatalogs}} and
\code{\link{VCFsToDNSCatalogs}}.
}
\note{
SNS 192 and DNS 144 catalog only contains mutations in transcribed regions.
}
