% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TCFromDenSigDen}
\alias{TCFromDenSigDen}
\title{density -> <anything>
density.signature -> density.signature, counts.signature}
\usage{
TCFromDenSigDen(s, t)
}
\description{
density -> <anything>
density.signature -> density.signature, counts.signature
}
\keyword{internal}
