% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{TestMutectVCFToCatalog}
\alias{TestMutectVCFToCatalog}
\title{test \code{MutectVCFToCatalog}.}
\usage{
TestMutectVCFToCatalog()
}
\description{
test \code{MutectVCFToCatalog}.
}
\details{
Stop if the catalogs created do not match the expected values.
}
