% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TranscriptRanges}
\alias{TranscriptRanges}
\alias{trans.ranges.GRCh37}
\alias{trans.ranges.GRCh38}
\title{Transcript ranges data}
\format{A data.table which contains transcript range and strand information
  for a particular reference genome. It contains chromosome name, start, end
  position, strand information and gene name and is keyed by chrom,
  chromStart, and chromEnd. It uses one-based coordinate system.}
\usage{
trans.ranges.GRCh37

trans.ranges.GRCh38
}
\description{
Transcript ranges and strand information for a particular reference genome.
}
\details{
\code{trans.ranges.GRCh37} A data.table which contains transcript range and
strand information for \strong{Human} GRCh37. It is derived from a raw
\strong{GFF3} format file
(ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_30/GRCh37_mapping/gencode.v30lift37.annotation.gff3.gz),
only genes that are associated with a CCDS ID are kept for
transcriptional strand bias analysis. Needed for
\code{\link{StrelkaSBSVCFFilesToCatalog}},
\code{\link{MutectVCFFilesToCatalog}},  \code{\link{VCFsToSBSCatalogs}} and
\code{\link{VCFsToDBSCatalogs}}.

\code{trans.ranges.GRCh38} A data.table which contains transcript range and
strand information for \strong{Human} GRCh38. It is derived from a raw
\strong{GFF3} format file
(ftp://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_30/gencode.v30.annotation.gff3.gz),
only genes that are associated with a CCDS ID are kept for
transcriptional strand bias analysis. Needed for
\code{\link{StrelkaSBSVCFFilesToCatalog}},
\code{\link{MutectVCFFilesToCatalog}},  \code{\link{VCFsToSBSCatalogs}} and
\code{\link{VCFsToDBSCatalogs}}.
}
\keyword{datasets}
