% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TransformCatalog}
\alias{TransformCatalog}
\title{TODO}
\usage{
TransformCatalog(catalog, target.ref.genome, target.region, target.type)
}
\arguments{
\item{catalog}{TODO}

\item{target.ref.genome}{TODO}

\item{target.region}{TODO}

\item{target.type}{TODO}
}
\value{
TODO
}
\description{
TODO
}
