% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TransformCatalog}
\alias{TransformCatalog}
\title{Transform catalog function}
\usage{
TransformCatalog(catalog, source.abundance, target.abundance = NULL,
  which.n, source.type, target.type = source.type)
}
\arguments{
\item{catalog}{A matrix of mutation counts/signature. Rownames indicate the mutation
types. Each column contains the mutation counts/signature for one sample.}

\item{source.abundance}{Either an abundance variable or string specifying an abundance.}

\item{target.abundance}{Either an abundance variable or string specifying an abundance.}

\item{which.n}{The n for the n-mers, one of 2, 3, 4, 5 for 2-mers, 3-mers, etc.}

\item{source.type}{A character specifying the type of the input catalog
("counts", "signature" or "density")}

\item{target.type}{A character specifying the type of the output catalog
("counts", "signature" or "density")}
}
\value{
A matrix of mutation counts/signature. Rownames indicate the mutation
  types. Each column contains the mutation counts/signature for one sample.
}
\description{
Transform catalog function
}
