% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TransformCatalog}
\alias{TransformCatalog}
\title{Transform between counts and density spectrum catalogs
and counts and density signature catalogs.}
\usage{
TransformCatalog(catalog, target.ref.genome, target.region,
  target.catalog.type, target.abundance = NULL)
}
\arguments{
\item{catalog}{An SBS or DBS catalog as described in \code{\link{ICAMS}};
must \strong{not} be an ID (indel) catalog.}

\item{target.ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{target.region}{A \code{region} argument; see \code{\link{as.catalog}}
and \code{\link{ICAMS}}.}

\item{target.catalog.type}{A character string acting as a catalog type
identifier, one of "counts", "density", "counts.signature",
"density.signature"; see \code{\link{as.catalog}}.}

\item{target.abundance}{Optional, only needed when
\code{target.ref.genome} is not one of the reference genomes
known to ICAMS (see \code{\link{ICAMS}}. 
The abundance should contain the counts of different
source sequences for mutations. See
\code{\link{all.abundance}}.}
}
\value{
A catalog as defined in \code{\link{ICAMS}}.
}
\description{
Transform between counts and density spectrum catalogs
and counts and density signature catalogs.
}
\details{
Only the following transformations are legal:

\enumerate{

\item \code{counts -> counts} (used to transform
   between \code{target.ref.genome} and/or \code{target.region})

\item \code{counts -> density}

\item \code{counts -> (counts.signature, density.signature)}

\item \code{density -> counts} (the semantics are to
infer the genome-wide or exome-wide counts based on the
densities)

\item \code{density -> (counts.signature, density.signature)}

\item \preformatted{counts.signature -> (counts.signature, density.signature)}

\item \preformatted{density.signature -> counts.signature}

\item \code{density.signature -> density.signature} (a null operation)

\item \code{density -> density} (a null operation)
}
}
