% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TransformCatalog}
\alias{TransformCatalog}
\title{Transform between count and density catalogs
 and signatures and between different
source-sequence abundances.}
\usage{
TransformCatalog(catalog, which.n, source.type,
  target.type = source.type, source.abundance = NULL,
  target.abundance = NULL)
}
\arguments{
\item{catalog}{An SNS or DNS catalog as described in \code{\link{ICAMS}};
must \strong{not} be an ID (indel) catalog.}

\item{which.n}{The length of the source sequences, one of 2:5.}

\item{source.type}{A character specifying type of the input catalog,
one of \code{"counts"}, \code{"signature"} or \code{"density"}.}

\item{target.type}{A character specifying type of the output catalog,
with the same possible values as \code{source.type}.}

\item{source.abundance}{Either \code{NULL} or a numeric vector with one element
for each source sequence for the mutation types in \code{catalog}
or a string specifying such a vector, one of \code{"GRCh37.genome"},
\code{"GRCh37.exome"}, \code{"GRCh38.genome"}, or \code{"GRCh38.exome"}.
This is the
abundance upon which the counts, densities, or proportions
in \code{catalog} are based. For example, for SNS in
trinucleotide context, e.g. ACT > AGT or TAC > TTC, the source
sequences are ACT and TAC.}

\item{target.abundance}{Same possibilities as \code{source.abundance}.}
}
\value{
A catalog as defined in \code{\link{ICAMS}}
}
\description{
Transform between count and density catalogs
 and signatures and between different
source-sequence abundances.
}
\details{
Only certain parings of type and abundance are legal, as follows:
\enumerate{

\item The type \code{"density"} must always be associated with a \code{NULL}
abundance.

\item The type \code{"signature"} is allowed to be associated with a \code{NULL}
abundance. A \code{NULL} abundance indicates that the signature
is a "density-based" signature (see \code{\link{ICAMS}}).

\item The type \code{"counts"} must \strong{not} be associated with
 the \code{NULL} abundance.
}

Only the following transformations are legal:

\enumerate{
\item \code{counts -> counts}
\item \code{counts -> density}
\item \code{counts -> signature}
\item \code{density -> counts} (in which case the semantics are to
infer the genome-wide or exome wide counts based on the
densities.)
\item \code{density -> signature}
\item \code{signature -> signature}
}
}
