% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{TransRownames.ID.PCAWG.SigPro}
\alias{TransRownames.ID.PCAWG.SigPro}
\title{For indels, convert PCAWG-formatted rownames (used in ICAMS)
into sigproextractor-formatted rownames.}
\usage{
TransRownames.ID.PCAWG.SigPro(vector.of.rownames)
}
\description{
E.g. 
TransRownames.ID.PCAWG.SigPro("DEL:C:1:0") returns 
1:Del:C:0;
TransRownames.ID.PCAWG.SigPro("INS:repeat:2:5+") returns
2:Ins:R:5
}
\details{
That is, executing the conversion
DEL:C:1:0 -> 1:Del:C:0
INS:repeat:2:5+ -> 2:Ins:R:5
}
\keyword{internal}
