% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{VCFsToDBSCatalogs}
\alias{VCFsToDBSCatalogs}
\title{Create DBS catalogs from VCFs}
\usage{
VCFsToDBSCatalogs(list.of.DBS.vcfs, ref.genome, trans.ranges, region)
}
\arguments{
\item{list.of.DBS.vcfs}{List of in-memory data frames of pure DBS mutations
-- no SBS or 3+BS mutations. The list names will be the sample ids in the
output catalog.}

\item{ref.genome}{A \code{ref.genome} argument as described in
\code{\link{ICAMS}}.}

\item{trans.ranges}{a \code{\link[data.table]{data.table}} which contains
transcript range and strand information. Please refer to
\code{\link{TranscriptRanges}} for more details.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}
}
\value{
A list of 3 DBS catalogs, one each for 78, 144, 136: catDBS78
  catDBS144 catDBS136. Each catalog has attributes added. See
  \code{\link{as.catalog}} for more details.
}
\description{
Create a list of 3 catalogs (one each for DBS78, DBS144 and DBS136)
out of the contents in list.of.DBS.vcfs. The VCFs must not contain
any type of mutation other then DBSs.
}
\note{
DBS 144 catalog only contains mutations in transcribed regions.
}
