% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{VCFsToDNSCatalogs}
\alias{VCFsToDNSCatalogs}
\title{Create DNS catalogs from VCFs}
\usage{
VCFsToDNSCatalogs(list.of.DNS.vcfs, ref.genome, trans.ranges, region)
}
\arguments{
\item{list.of.DNS.vcfs}{List of in-memory data frames of pure DNS mutations
-- no SNS or 3+BS mutations. The list names will be the sample ids in the
output catalog.}

\item{ref.genome}{A genome argument as described in \code{\link{ICAMS}}.}

\item{trans.ranges}{A data frame containing transcript ranges.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}
}
\value{
A list of 3 DNS catalogs, one each for 78, 144, 136: catDNS78
  catDNS144 catDNS136. Each catalog has attributes added. See
  \code{\link{as.catalog}} for more details.
}
\description{
Create a list of 3 catalogs (one each for DNS78, DNS144 and DNS136)
out of the contents in list.of.DNS.vcfs. The VCFs must not contain
any type of mutation other then DNSs.
}
\note{
DNS 144 catalog only contains mutations in transcribed regions.
}
