% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{VCFsToDNSCatalogs}
\alias{VCFsToDNSCatalogs}
\title{Create DNS catalogs from VCFs}
\usage{
VCFsToDNSCatalogs(list.of.DNS.vcfs, genome, trans.ranges)
}
\arguments{
\item{list.of.DNS.vcfs}{List of in-memory data frames of pure DNS mutations
-- no SNS or 3+BS mutations. The list names will be the sample ids in the
output catalog.}

\item{genome}{Name of a particular reference genome
(without quotations marks).}

\item{trans.ranges}{A data frame containing transcript ranges.}
}
\value{
A list of 3 catalogs, one each for DNS78, DNS144, QUAD136:
  catDNS78
  catDNS144
  catQUAD136
}
\description{
Create a list of 3 catalogs (one each for DNS78, DNS144 and QUAD136)
out of the contents in list.of.DNS.vcfs. The VCFs must not contain
any type of mutation other then DNSs.
}
\keyword{internal}
