% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{VCFsToIDCatalogs}
\alias{VCFsToIDCatalogs}
\title{Create ID (insertion and deletion) catalog from ID VCFs}
\usage{
VCFsToIDCatalogs(list.of.vcfs, genome)
}
\arguments{
\item{list.of.vcfs}{List of in-memory VCFs. The list names will be
the sample ids in the output catalog.}

\item{genome}{A particular reference genome(without quotation marks). Use
\link[BSgenome]{available.genomes} to get the list of "BSgenome data
packages" currently available. There are 2 types of predefined reference
genome which are incorporated in this function. User can invoke a
predefined human GRCh38/hg38 BSgenome data package by typing \code{genome =
"GRCh38"} or \code{genome = "hg38"}. User can invoke a predefined human
GRCh37/hg19 BSgenome data package by typing \code{genome = "GRCh37"} or
\code{genome = "hg19"}.}
}
\value{
An ID (indel) catalog
}
\description{
Create ID (insertion and deletion) catalog from ID VCFs
}
