% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{VCFsToIDCatalogs}
\alias{VCFsToIDCatalogs}
\title{Create ID (indel) catalog from VCFs}
\usage{
VCFsToIDCatalogs(list.of.vcfs, genome)
}
\arguments{
\item{list.of.vcfs}{List of in-memory VCFs. The list names will be
the sample ids in the output catalog.}

\item{genome}{Name of a particular reference genome
(without quotations marks).}
}
\value{
An ID (indel) catalog
}
\description{
Create ID (indel) catalog from VCFs
}
\keyword{internal}
