% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{VCFsToIDCatalogs}
\alias{VCFsToIDCatalogs}
\title{Create ID (insertion and deletion) catalog from ID VCFs}
\usage{
VCFsToIDCatalogs(list.of.vcfs, genome)
}
\arguments{
\item{list.of.vcfs}{List of in-memory VCFs. The list names will be
the sample ids in the output catalog.}

\item{genome}{A genome argument as described in \code{\link{ICAMS}}.}
}
\value{
An ID (indel) catalog
}
\description{
Create ID (insertion and deletion) catalog from ID VCFs
}
