% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDELS_related_functions.R
\name{VCFsToIDCatalogs}
\alias{VCFsToIDCatalogs}
\title{Create ID (insertion and deletion) catalog from ID VCFs}
\usage{
VCFsToIDCatalogs(list.of.vcfs, ref.genome, region)
}
\arguments{
\item{list.of.vcfs}{List of in-memory VCFs. The list names will be
the sample ids in the output catalog.}

\item{ref.genome}{A genome argument as described in \code{\link{ICAMS}}.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}
}
\value{
An S3 object containing an ID (indel) catalog with class
  "catalog". See \code{\link{as.catalog}} for more details.
}
\description{
Create ID (insertion and deletion) catalog from ID VCFs
}
