% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_to_catalog_functions.R
\name{VCFsToSNSCatalogs}
\alias{VCFsToSNSCatalogs}
\title{Create SNS catalogs from SNS VCFs}
\usage{
VCFsToSNSCatalogs(list.of.SNS.vcfs, ref.genome, trans.ranges, region)
}
\arguments{
\item{list.of.SNS.vcfs}{List of in-memory data frames of pure SNS mutations
-- no DNS or 3+BS mutations. The list names will be the sample ids in the
output catalog.}

\item{ref.genome}{A genome argument as described in \code{\link{ICAMS}}.}

\item{trans.ranges}{A data frame containing transcript ranges.}

\item{region}{A character string acting as a region identifier, one of
"genome", "exome".}
}
\value{
A list of 3 SNS catalogs, one each for 96, 192, 1536: catSNS96
  catSNS192 catSNS1536. Each catalog has attributes added. See
  \code{\link{CreateCatalogAttribute}} for more details.
}
\description{
Create a list of 3 catalogs (one each for 96, 192, 1536)
out of the contents in list.of.SNS.vcfs. The SNS VCFs must not contain
DNSs, indels, or other types of mutations.
}
\note{
SNS 192 catalog only contains mutations in transcribed regions.
}
