% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCat}
\alias{WriteCat}
\title{Write a catalog to a file.}
\usage{
WriteCat(catalog, file, num.row, row.order, row.header, strict)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}} with attributes added.
See \code{\link{as.catalog}} for more details.}

\item{file}{The path of the file to be written.}

\item{num.row}{The number of rows in the file to be written.}

\item{row.order}{The row order to be used for writing the file.}

\item{row.header}{The row header to be used for writing the file.}

\item{strict}{If TRUE, then stop if additional checks on the input fail.}
}
\description{
This internal function is called by exported functions to do the
actual writing of the catalog.
}
\keyword{internal}
