% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCat}
\alias{WriteCat}
\title{Write a catalog to a file.}
\usage{
WriteCat(catalog, path, num.row, row.order, row.header, strict)
}
\arguments{
\item{catalog}{A catalog.}

\item{path}{The path of the file to be written.}

\item{num.row}{The number of rows in the file to be written.}

\item{row.order}{The row order to be used for writing the file.}

\item{row.header}{The row header to be used for writing the file.}

\item{strict}{If TRUE, then stop if additional checks on the input fail.}
}
\value{
A catalog in standard in-memory format.
}
\description{
This internal function is called by exported functions to do the
actual writing of the catalog.
}
\keyword{internal}
