% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCatalog}
\alias{WriteCatalog}
\title{Write catalog.}
\usage{
WriteCatalog(catalog, path, strict = TRUE)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}}.}

\item{path}{The path of the file to be written on disk.}

\item{strict}{If TRUE, then fail if additional checks on the input fail.}
}
\description{
Write a catalog to a file on disk.
}
\details{
See also \code{\link{ReadCatalog}}
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
