% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCatalog}
\alias{WriteCatalog}
\title{Write a catalog}
\usage{
WriteCatalog(catalog, path, strict = TRUE)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}};
see also \code{\link{as.catalog}}.}

\item{path}{The path to the file to be created .}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}
}
\description{
Write a catalog to a file.
}
\details{
See also \code{\link{ReadCatalog}}.
}
\note{
In the ID (insertion and deletion) catalog, deletion repeat size
  ranges from 0 to 5+, but for plotting and end user documentation it ranges
  from 1 to 6+.
}
