% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_catalog.R
\name{WriteCatalog}
\alias{WriteCatalog}
\alias{WriteCatSNS96}
\alias{WriteCatSNS192}
\alias{WriteCatSNS1536}
\alias{WriteCatDNS78}
\alias{WriteCatDNS144}
\alias{WriteCatDNS136}
\alias{WriteCatID}
\title{Write Catalog Functions}
\usage{
WriteCatSNS96(ct, path, strict = TRUE)

WriteCatSNS192(ct, path, strict = TRUE)

WriteCatSNS1536(ct, path, strict = TRUE)

WriteCatDNS78(ct, path, strict = TRUE)

WriteCatDNS144(ct, path, strict = TRUE)

WriteCatDNS136(ct, path, strict = TRUE)

WriteCatID(ct, path, strict = TRUE)
}
\arguments{
\item{ct}{A matrix of mutation catalog.}

\item{path}{The path of the file to be written on disk.}

\item{strict}{If TRUE, do additional checks on the input,
and stop if the checks fail.}
}
\description{
Write a mutation catalog to a file on disk
}
\details{
\code{WriteCatSNS96} Write a SNS 96 mutation catalog to a file on disk

\code{WriteCatSNS192} Write a SNS 192 mutation catalog to a file on disk

\code{WriteCatSNS1536} Write a SNS 1536 mutation catalog to a file on disk

\code{WriteCatDNS78} Write a DNS 78 mutation catalog to a file on disk

\code{WriteCatDNS144} Write a DNS 144 mutation catalog to a file on disk

\code{WriteCatDNS136} Write a 136 DNS catalog from path

\code{WriteCatID} Write a ID (insertion/deletion) catalog to a file on disk
Please take note that deletion repeat size ranges from 0 to 5+
in the catalog, but for plotting and end user documentation
it ranges from 1 to 6+.

See also \code{\link{ReadCatalog}}
}
