% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_catalog_formats.R
\name{WriteCatalogIndelSigPro}
\alias{WriteCatalogIndelSigPro}
\title{Write Indel Catalogs in SigProExtractor format.}
\usage{
WriteCatalogIndelSigPro(catalog, file, strict = TRUE)
}
\arguments{
\item{catalog}{A catalog as defined in \code{\link{ICAMS}};
see also \code{\link{as.catalog}}.}

\item{file}{The path to the file to be created.}

\item{strict}{If TRUE, do additional checks on the input, and stop if the
checks fail.}
}
\description{
Write Indel Catalogs in SigProExtractor format to a csv file.
}
\note{
In ID (insertion and deletion) catalogs in SigProExtractor format, 
deletion repeat sizes range from 0 to 5, rather than 0 to 5+.
}
\keyword{internal}
